### 2025-06-24 #8467 @editorialbot
[PRE REVIEW]: automesh: Automatic mesh generation in Rust
**Submitting author:** @hovey (<a href="http://orcid.org/0000-0003-0915-5765">Chad Hovey</a>)
**Repository:** https://github.com/autotwin/automesh
**Branch with paper.md** (empty if default branch): joss_2025
**Version:** v0.3.3
