+++ .github/workflows/draft-pdf.yml
@@ -6,7 +6,7 @@ on:
   pull_request:
     branches: main
   push:
-    branches: main
+    branches: [main, joss_2025]
 
 permissions: {}
 jobs:
