+++ src/tree/hex/vertex_template_7/mod.rs
@@ -194,7 +194,6 @@ fn template(
                                     {
                                         if let Some(cell_ab_subsubcells) = tree
                                             .cell_subcells_contain_leaves(
-                                            // .cell_subcells_contain_cells(
                                                 &tree[cell_ab_index],
                                                 face_index_b,
                                             )
@@ -207,21 +206,18 @@ fn template(
                                             {
                                                 if let Some(cell_c_a_subsubcells) = tree
                                                     .cell_subcells_contain_leaves(
-                                                    // .cell_subcells_contain_cells(
                                                         &tree[cell_c_a_index],
                                                         face_index_a,
                                                     )
                                                 {
                                                     if let Some(cell_c_b_subsubcells) = tree
                                                         .cell_subcells_contain_leaves(
-                                                        // .cell_subcells_contain_cells(
                                                             &tree[cell_c_b_index],
                                                             face_index_b,
                                                         )
                                                     {
                                                         if let Some(cell_c_ab_subsubcells) = tree
                                                                 .cell_subcells_contain_leaves(
-                                                                // .cell_subcells_contain_cells(
                                                                 &tree[cell_c_ab_index],
                                                                 face_index_b,
                                                             )
