+++ src/tree/hex/vertex_template_5/mod.rs
@@ -496,7 +496,7 @@ fn template(
                                                         )
                                                     {
                                                         if let Some(cell_c_ab_subsubcells) = tree
-                                                            .cell_subcells_contain_cells(
+                                                            .cell_subcells_contain_leaves(
                                                                 &tree[cell_c_ab_index],
                                                                 face_index_b,
                                                             )
