### 2025-06-25 6d79333
questions about 6 and 7
