+++ src/tree/hex/vertex_template_8/mod.rs
@@ -1,0 +1,539 @@
+use super::super::{Faces, HexConnectivity, Indices, Octree};
+
+pub fn apply(
+    cells_nodes: &[usize],
+    tree: &Octree,
+    element_node_connectivity: &mut HexConnectivity,
+) {
+    tree.iter()
+        .filter_map(|cell| tree.cell_contains_leaves(cell))
+        .for_each(|(cell_subcells, cell_faces)| {
+            template(
+                0,
+                1,
+                5,
+                5,
+                4,
+                0,
+                5,
+                15,
+                10,
+                0,
+                5,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                0,
+                4,
+                1,
+                1,
+                5,
+                4,
+                0,
+                5,
+                15,
+                10,
+                0,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                0,
+                3,
+                4,
+                0,
+                1,
+                5,
+                0,
+                0,
+                5,
+                15,
+                10,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                0,
+                5,
+                3,
+                4,
+                0,
+                1,
+                10,
+                10,
+                0,
+                5,
+                15,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                1,
+                2,
+                5,
+                7,
+                5,
+                1,
+                15,
+                15,
+                10,
+                0,
+                5,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                1,
+                4,
+                2,
+                3,
+                7,
+                5,
+                5,
+                5,
+                15,
+                10,
+                0,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                1,
+                0,
+                4,
+                1,
+                3,
+                7,
+                5,
+                0,
+                5,
+                15,
+                10,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                1,
+                5,
+                0,
+                5,
+                1,
+                3,
+                15,
+                10,
+                0,
+                5,
+                15,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                2,
+                3,
+                5,
+                6,
+                7,
+                3,
+                10,
+                10,
+                15,
+                5,
+                0,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                2,
+                4,
+                3,
+                2,
+                6,
+                7,
+                5,
+                0,
+                10,
+                15,
+                5,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                2,
+                1,
+                4,
+                3,
+                2,
+                6,
+                15,
+                5,
+                0,
+                10,
+                15,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                2,
+                5,
+                1,
+                7,
+                3,
+                2,
+                15,
+                15,
+                5,
+                0,
+                10,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
[…]
