### 2025-06-25 f611190
strong balancing over vertices and querying only corner subsubcells in vertex templates
