+++ src/tree/hex/vertex_template_8/mod.rs
@@ -468,40 +468,46 @@ fn template(
                 tree.cell_contains_leaves(&tree[cell_a_index])
             {
                 if let Some(cell_ab_index) = cell_a_faces[face_index_b] {
-                    if let Some(cell_b_subsubcells) =
-                        tree.cell_subcells_contain_leaves(&tree[cell_b_index], face_index_b)
-                    {
+                    if let Some(cell_b_subsubcells) = tree.cell_subcell_contains_leaves(
+                        &tree[cell_b_index],
+                        face_index_b,
+                        cell_subsubcell_b_index,
+                    ) {
                         if let Some((cell_ab_subcells, cell_ab_faces)) =
                             tree.cell_contains_leaves(&tree[cell_ab_index])
                         {
                             if let Some(cell_c_index) = cell_faces[face_index] {
-                                if let Some(cell_c_subsubcells) = tree
-                                    .cell_subcells_contain_leaves(&tree[cell_c_index], face_index)
-                                {
+                                if let Some(cell_c_subsubcells) = tree.cell_subcell_contains_leaves(
+                                    &tree[cell_c_index],
+                                    face_index,
+                                    cell_subsubcell_c_index,
+                                ) {
                                     if let Some(cell_c_a_index) = cell_a_faces[face_index] {
                                         if let Some(cell_c_a_subsubcells) = tree
-                                            .cell_subcells_contain_leaves(
+                                            .cell_subcell_contains_leaves(
                                                 &tree[cell_c_a_index],
                                                 face_index,
+                                                cell_subsubcell_c_a_index,
                                             )
                                         {
                                             if let Some(cell_c_b_index) =
                                                 tree[cell_b_index].get_faces()[face_index]
                                             {
                                                 if let Some(cell_c_b_subsubcells) = tree
-                                                    // .cell_subcells_contain_leaves(
-                                                    .cell_subcells_contain_cells(
+                                                    .cell_subcell_contains_leaves(
                                                         &tree[cell_c_b_index],
                                                         face_index,
+                                                        cell_subsubcell_c_b_index,
                                                     )
                                                 {
                                                     if let Some(cell_c_ab_index) =
                                                         cell_ab_faces[face_index]
                                                     {
                                                         if let Some(cell_c_ab_subsubcells) = tree
-                                                            .cell_subcells_contain_leaves(
+                                                            .cell_subcell_contains_leaves(
                                                                 &tree[cell_c_ab_index],
                                                                 face_index,
+                                                                cell_subsubcell_c_ab_index,
                                                             )
                                                         {
                                                             element_node_connectivity.push([
