+++ src/tree/hex/vertex_template_3/mod.rs
@@ -277,18 +277,20 @@ fn template(
                                             tree.cell_contains_leaves(&tree[cell_ab_index])
                                         {
                                             if let Some(cell_c_subsubcells) = tree
-                                                .cell_subcells_contain_leaves(
+                                                .cell_subcell_contains_leaves(
                                                     &tree[cell_c_index],
                                                     face_index_c,
+                                                    cell_subsubcell_c_index,
                                                 )
                                             {
                                                 if let Some((cell_c_a_subcells, _)) =
                                                     tree.cell_contains_leaves(&tree[cell_c_a_index])
                                                 {
                                                     if let Some(cell_c_b_subsubcells) = tree
-                                                        .cell_subcells_contain_leaves(
+                                                        .cell_subcell_contains_leaves(
                                                             &tree[cell_c_b_index],
                                                             face_index_c,
+                                                            cell_subsubcell_c_b_index,
                                                         )
                                                     {
                                                         if let Some((cell_c_ab_subcells, _)) = tree
