### 2025-06-26 507022d
fix DOI issues
