+++ pubs/joss_2025/paper.bib
@@ -31,7 +31,7 @@
   doi={10.1007/s00707-009-0274-0}
 }
 @article{bracci2019hexalab,
-  title={HexaLab. net: An online viewer for hexahedral meshes},
+  title={HexaLab.net: An online viewer for hexahedral meshes},
   author={Bracci, Matteo and Tarini, Marco and Pietroni, Nico and Livesu, Marco and Cignoni, Paolo},
   journal={Computer-Aided Design},
   volume={110},
@@ -135,7 +135,8 @@
   booktitle={Proceedings of the 18th international meshing roundtable},
   pages={65--84},
   year={2009},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-04319-2_5}
 }
 @article{marechal2016all,
   title={All hexahedral boundary layers generation},
@@ -149,8 +150,10 @@
 }
 @techreport{meagher1980,
   author       = {D. J. Meagher},
+  title={Octree Encoding: A New Technique for the Representation, Manipulation and Display of Arbitrary 3-D Objects by Computer},
   institution  = {Rensselaer Polytechnic Institute},
   number       = {IPL-TR-80-111},
+  month        = {10},
   year         = {1980},
 }
 @misc{meshlab,
@@ -175,7 +178,7 @@
 }
 @misc{numpy_format,
   author       = {{NumPy Developers}},
-  title        = {numpy.lib.format — NumPy v1.24 Manual},
+  title        = {numpy.lib.format, NumPy v1.24 Manual},
   year         = {2025},
   url          = {https://numpy.org/doc/stable/reference/generated/numpy.lib.format.html},
   note         = {Accessed: 2025-03-06}
@@ -203,7 +206,8 @@
   booktitle={Proceedings of the 22nd International Meshing Roundtable},
   pages={39--56},
   year={2014},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-02335-9_3}
 }
 @article{polonsky2023toward,
   title={Toward accurate prediction of partial-penetration laser weld performance informed by three-dimensional characterization--Part I: high fidelity interrogation},
@@ -213,7 +217,7 @@
   pages={100006},
   year={2023},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.tmater.2023.100006}
+  doi={10.1016/j.tmater.2023.100006}
 }
 @article{rust_programming_language,
   author       = {Steve Klabnik and Carol Nichols},
@@ -227,7 +231,8 @@
   title={EXODUS II: a finite element data model},
   author={Schoof, Larry A and Yarberry, Victor R},
   year={1994},
-  institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)}
+  institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)},
+  doi={10.2172/10102115}
 }
 @misc{siemens_simcenter,
   author       = {{Siemens AG}},
@@ -257,14 +262,16 @@
   author={Taubin, Gabriel},
   booktitle={Proceedings of the 22nd annual conference on Computer graphics and interactive techniques},
   pages={351--358},
-  year={1995}
+  year={1995},
+  doi={10.1145/3596711.3596724}
 }
 @misc{tetgen,
   author       = {H. Si},
   title        = {TetGen: A Delaunay-Based Quality Tetrahedral Mesh Generator},
   year         = {2025},
   url          = {https://wias-berlin.de/software/index.jsp?id=TetGen&lang=1},
-  note         = {Accessed: 2025-04-15}
+  note         = {Accessed: 2025-04-15},
+  doi          = {10.1145/2629697}
 }
 @misc{truegrid,
   author       = {{TrueGrid}},
