### 2025-06-26 20f57da
additional citations for Livesu and Gurobi
