+++ pubs/joss_2025/paper.bib
@@ -96,6 +96,13 @@
   url          = {https://www.pointwise.com/},
   note         = {Accessed: 2025-03-10}
 }
+@misc{gurobi,
+  author       = {{Gurobi Optimization, LLC}},
+  title        = {Gurobi Optimizer},
+  year         = {2025},
+  url          = {https://www.gurobi.com/},
+  note         = {Accessed: 2025-03-10}
+}
 @misc{hexotic,
   author       = {{INRIA Gamma Team}},
   title        = {Hexotic: A Software for Hexahedral Mesh Generation},
@@ -117,9 +124,11 @@
   title={cinolib: a generic programming header only C++ library for processing polygonal and polyhedral meshes},
   author={Livesu, Marco},
   journal={Transactions on Computational Science XXXIV},
+  series={Lecture Notes in Computer Science},
   pages={64--76},
   year={2019},
   publisher={Springer},
+  url={https://github.com/mlivesu/cinolib/},
   doi={10.1007/978-3-662-59958-7_4}
 }
 @article{livesu2021optimal,
@@ -208,6 +217,17 @@
   year={2014},
   organization={Springer},
   doi={10.1007/978-3-319-02335-9_3}
+}
+@article{pitzalis2021generalized,
+  title={Generalized adaptive refinement for grid-based hexahedral meshing},
+  author={Pitzalis, Luca and Livesu, Marco and Cherchi, Gianmarco and Gobbetti, Enrico and Scateni, Riccardo},
+  journal={ACM Transactions on Graphics (TOG)},
+  volume={40},
+  number={6},
+  pages={1--13},
+  year={2021},
+  doi={10.1145/3478513.3480508},
+  publisher={ACM New York, NY, USA}
 }
 @article{polonsky2023toward,
   title={Toward accurate prediction of partial-penetration laser weld performance informed by three-dimensional characterization--Part I: high fidelity interrogation},
