### 2025-06-26 9df8200
shuffle citation location
