+++ src/tree/hex/vertex_template_9/mod.rs
@@ -1,0 +1,190 @@
+use super::super::{Faces, HexConnectivity, Indices, Octree};
+
+pub fn apply(
+    cells_nodes: &[usize],
+    tree: &Octree,
+    element_node_connectivity: &mut HexConnectivity,
+) {
+    tree.iter()
+        .filter_map(|cell| tree.cell_contains_leaves(cell))
+        .for_each(|(cell_subcells, cell_faces)| {
+            template(
+                0,
+                1,
+                5,
+                5,
+                10,
+                0,
+                5,
+                15,
+                10,
+                2,
+                5,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                1,
+                2,
+                5,
+                7,
+                15,
+                5,
+                15,
+                15,
+                10,
+                0,
+                5,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                2,
+                3,
+                5,
+                6,
+                15,
+                5,
+                10,
+                10,
+                15,
+                1,
+                0,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+            template(
+                3,
+                0,
+                5,
+                4,
+                10,
+                0,
+                0,
+                10,
+                15,
+                3,
+                0,
+                cell_faces,
+                cell_subcells,
+                cells_nodes,
+                tree,
+                element_node_connectivity,
+            );
+        })
+}
+
+#[allow(clippy::too_many_arguments)]
+fn template(
+    face_index: usize,
+    face_index_a: usize,
+    face_index_b: usize,
+    cell_subcell_index: usize,
+    cell_subsubcell_a_index: usize,
+    cell_subsubcell_ab_index: usize,
+    cell_subsubcell_b_index: usize,
+    cell_subsubcell_c_index: usize,
+    cell_subsubcell_c_a_index: usize,
+    cell_subcell_c_ab_index: usize,
+    cell_subsubcell_c_b_index: usize,
+    cell_faces: &Faces,
+    cell_subcells: &Indices,
+    cells_nodes: &[usize],
+    tree: &Octree,
+    element_node_connectivity: &mut HexConnectivity,
+) {
+    if let Some(cell_a_index) = cell_faces[face_index_a] {
+        if let Some(cell_b_index) = cell_faces[face_index_b] {
+            if let Some(cell_a_subsubcells) = tree.cell_subcell_contains_leaves(
+                &tree[cell_a_index],
+                face_index_a,
+                cell_subsubcell_a_index,
+            ) {
+                if let Some(cell_ab_index) = tree[cell_a_index].get_faces()[face_index_b] {
+                    if let Some(cell_b_subsubcells) = tree.cell_subcell_contains_leaves(
+                        &tree[cell_b_index],
+                        face_index_b,
+                        cell_subsubcell_b_index,
+                    ) {
+                        if let Some(cell_ab_subsubcells) = tree.cell_subcell_contains_leaves(
+                            &tree[cell_ab_index],
+                            face_index_a,
+                            cell_subsubcell_ab_index,
+                        ) {
+                            if let Some(cell_c_index) = cell_faces[face_index] {
+                                if let Some(cell_c_subsubcells) = tree.cell_subcell_contains_leaves(
+                                    &tree[cell_c_index],
+                                    face_index,
+                                    cell_subsubcell_c_index,
+                                ) {
+                                    if let Some(cell_c_a_index) =
+                                        tree[cell_c_index].get_faces()[face_index_a]
+                                    {
[…]
