+++ src/tree/hex/mod.rs
@@ -12,3 +12,4 @@ pub mod vertex_template_5; // (O, A, AB, B) => (o, A, ab, b)
 pub mod vertex_template_6; // (O, A, AB, b) => (o, A, ab, b)
 pub mod vertex_template_7; // (O, a, ab, b) => (o, a, ab, b)
 pub mod vertex_template_8; // (O, A, AB, b) => (o, a, ab, b)
+pub mod vertex_template_9; // (O, a, ab, b) => (o, a, AB, b)
