### 2025-06-26 535fc88
9 done and all vertex templates now parallel
