+++ src/tree/hex/vertex_template_2/mod.rs
@@ -1,234 +1,22 @@
-use super::super::{Faces, HexConnectivity, Indices, Octree};
+use super::super::{Faces, HEX, Indices, Octree};
 
-pub fn apply(
-    cells_nodes: &[usize],
-    tree: &Octree,
-    element_node_connectivity: &mut HexConnectivity,
-) {
-    tree.iter()
-        .filter_map(|cell| tree.cell_contains_leaves(cell))
-        .for_each(|(cell_subcells, cell_faces)| {
-            template(
-                0,
-                3,
-                5,
-                4,
-                0,
-                10,
-                10,
-                15,
-                0,
-                0,
-                5,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                1,
-                0,
-                5,
-                5,
-                1,
-                10,
-                15,
-                10,
-                0,
-                5,
-                0,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                2,
-                1,
-                5,
-                7,
-                3,
-                15,
-                15,
-                10,
-                5,
-                5,
-                0,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                3,
-                2,
-                5,
-                6,
-                2,
-                15,
-                10,
-                15,
-                5,
-                0,
-                5,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                4,
-                1,
-                2,
-                3,
-                1,
-                15,
-                5,
-                15,
-                5,
-                0,
-                10,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                2,
-                4,
-                1,
-                3,
-                2,
-                5,
-                15,
-                5,
-                0,
-                10,
-                0,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                3,
-                4,
-                2,
-                2,
-                0,
-                5,
-                10,
-                15,
-                0,
-                0,
-                5,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                0,
-                4,
-                3,
-                0,
-                1,
-                0,
-                0,
-                10,
-                5,
-                5,
-                15,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                5,
-                2,
-                1,
-                7,
-                6,
-                15,
-                15,
-                5,
-                10,
-                10,
-                0,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                0,
-                5,
-                1,
-                5,
-                4,
-                15,
-                5,
-                15,
-                10,
-                0,
-                10,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                5,
-                3,
-                2,
-                6,
-                4,
-                10,
-                15,
-                5,
-                0,
-                10,
-                0,
-                cell_faces,
-                cell_subcells,
-                cells_nodes,
-                tree,
-                element_node_connectivity,
-            );
-            template(
-                1,
[…]
