### 2025-06-26 54501a1
9 done and all vertex templates now parallel
