### 2025-06-26 1d0a5f3
vertex template 10
