+++ src/tree/hex/mod.rs
@@ -8,8 +8,8 @@ pub mod face_template_0;
 pub mod face_template_1;
 mod vertex_template_1; // (O, A, AB, B) => (o, a, ab, b)
 mod vertex_template_10; // (O, a, AB, b) => (O, a, ab, b)
-// mod vertex_template_11; // (O, a, AB, b) => (o, a, ab, b)
-// mod vertex_template_12; // (O, a, AB, b) => (O, a, AB, b)
+mod vertex_template_11; // (O, a, AB, b) => (O, a, AB, b)
+// mod vertex_template_12; // (O, a, AB, b) => (o, a, ab, b)
 mod vertex_template_2; // (O, a, ab, b) => (O, a, ab, b)
 mod vertex_template_3; // (O, A, AB, B) => (o, A, AB, b)
 mod vertex_template_4; // (O, A, AB, B) => (o, A, AB, B)
@@ -80,6 +80,12 @@ pub fn vertex_template(index: usize, cells_nodes: &[usize], tree: &Octree) -> He
             tree,
             vertex_template_10::DATA,
             vertex_template_10::template,
+        ),
+        11 => apply_vertex_template(
+            cells_nodes,
+            tree,
+            vertex_template_11::DATA,
+            vertex_template_11::template,
         ),
         _ => panic!(),
     }
