### 2025-06-27 13be8dc
vertex template 12
