+++ src/tree/hex/edge_template_3/mod.rs
@@ -1,200 +1,388 @@
 use super::super::{
-    Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NodeMap, Octree, face_direction,
+    Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NUM_OCTANTS, NodeMap, Octree,
 };
+use conspire::math::{TensorVec, tensor_rank_1};
 
 pub fn apply(
     cells_nodes: &[usize],
     nodes_map: &mut NodeMap,
+    node_index: &mut usize,
     tree: &Octree,
     element_node_connectivity: &mut HexConnectivity,
-    nodal_coordinates: &Coordinates,
+    nodal_coordinates: &mut Coordinates,
 ) {
     tree.iter()
         .filter_map(|cell| tree.cell_contains_leaves(cell))
-        .for_each(|(cell_subcells, cell_faces)| {
-            cell_faces
-                .iter()
-                .enumerate()
-                .for_each(|(face_index, face_cell)| {
-                    if let Some(face_cell_index) = face_cell {
-                        if let Some(face_subsubcells) =
-                            tree.cell_subcells_contain_leaves(&tree[*face_cell_index], face_index)
-                        {
-                            template_inner(face_index).into_iter().for_each(
-                                |[
-                                    adjacent_face,
-                                    cell_subcell_a,
-                                    cell_subcell_b,
-                                    adjacent_cell_subcell_a,
-                                    adjacent_cell_subcell_b,
-                                    face_subsubcell_a,
-                                    face_subsubcell_b,
-                                    face_subsubcell_c,
-                                    face_subsubcell_d,
-                                    adjacent_face_subsubcell_a,
-                                    adjacent_face_subsubcell_b,
-                                    adjacent_face_subsubcell_c,
-                                    adjacent_face_subsubcell_d,
-                                ]| {
-                                    if let Some(adjacent_cell) = cell_faces[adjacent_face] {
-                                        if let Some((adjacent_cell_subcells, adjacent_cell_faces)) =
-                                            tree.cell_contains_leaves(&tree[adjacent_cell])
-                                        {
-                                            if let Some(adjacent_cell_face_cell) =
-                                                adjacent_cell_faces[face_index]
+        .for_each(|(cell_subcells, _)| {
+            template(
+                0,
+                1,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                0,
+                2,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                0,
+                4,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                1,
+                3,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                1,
+                5,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                2,
+                3,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                2,
+                6,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                node_index,
+                tree,
[…]
