+++ src/tree/hex/edge_template_1/mod.rs
@@ -1,7 +1,7 @@
 use super::super::{
-    Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NUM_OCTANTS, NodeMap, Octree,
+    Cell, Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NodeMap, Octree, mirror_face,
 };
-use conspire::math::{TensorVec, tensor_rank_1};
+use conspire::math::{TensorRank1, TensorVec, tensor_rank_1};
 
 pub fn apply(
     cells_nodes: &[usize],
@@ -11,367 +11,316 @@ pub fn apply(
     element_node_connectivity: &mut HexConnectivity,
     nodal_coordinates: &mut Coordinates,
 ) {
-    tree.iter()
-        .filter_map(|cell| tree.cell_contains_leaves(cell))
-        .for_each(|(cell_subcells, _)| {
-            template(
-                0,
-                1,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                0,
-                2,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                0,
-                4,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                1,
-                3,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                1,
-                5,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                2,
-                3,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                2,
-                6,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                3,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                4,
-                5,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                4,
-                6,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                5,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                6,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
[…]

[…]
