+++ src/tree/hex/edge_template_4/mod.rs
@@ -18,6 +18,90 @@ pub fn apply(
                 6,
                 3,
                 5,
+                7,
+                5,
+                2,
+                0,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                5,
+                7,
+                5,
+                1,
+                3,
+                1,
+                6,
+                4,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                5,
+                4,
+                0,
+                5,
+                6,
+                7,
+                0,
+                1,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                7,
+                6,
+                5,
+                2,
+                2,
+                3,
+                4,
+                5,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                6,
+                2,
+                3,
+                2,
+                3,
+                7,
+                0,
+                4,
+                cells_nodes,
+                cell_subcells,
+                nodes_map,
+                tree,
+                element_node_connectivity,
+                nodal_coordinates,
+            );
+            template(
+                7,
+                3,
+                2,
+                1,
+                1,
+                5,
+                2,
+                6,
                 cells_nodes,
                 cell_subcells,
                 nodes_map,
@@ -34,6 +118,10 @@ fn template(
     subcell_b: usize,
     face_m: usize,
     face_n: usize,
+    subcell_face_m_p: usize,
+    subcell_face_m_q: usize,
+    subcell_face_n_p: usize,
+    subcell_face_n_q: usize,
     cells_nodes: &[usize],
     cell_subcells: &[usize; NUM_OCTANTS],
     nodes_map: &mut NodeMap,
@@ -63,11 +151,13 @@ fn template(
                                                 if let Some((subcell_b_face_n_subcells, _)) = tree
                                                     .cell_contains_leaves(&tree[subcell_b_face_n])
                                                 {
-                                                    let lngth = *tree[subcell_a_face_m_subcells[7]]
+                                                    let lngth = *tree[subcell_a_face_m_subcells
+                                                        [subcell_face_m_p]]
                                                         .get_lngth()
                                                         as f64;
                                                     let coordinates_1 = &nodal_coordinates
-                                                        [cells_nodes[subcell_a_face_m_subcells[7]]
+                                                        [cells_nodes[subcell_a_face_m_subcells
+                                                            [subcell_face_m_p]]
                                                             - NODE_NUMBERING_OFFSET]
                                                         - face_direction(face_m) * lngth;
                                                     let node_1 = *nodes_map
@@ -78,7 +168,8 @@ fn template(
                                                         ))
                                                         .expect("nonexistent entry");
                                                     let coordinates_2 = &nodal_coordinates
-                                                        [cells_nodes[subcell_b_face_m_subcells[5]]
+                                                        [cells_nodes[subcell_b_face_m_subcells
+                                                            [subcell_face_m_q]]
                                                             - NODE_NUMBERING_OFFSET]
                                                         - face_direction(face_m) * lngth;
                                                     let node_2 = *nodes_map
@@ -89,7 +180,8 @@ fn template(
                                                         ))
                                                         .expect("nonexistent entry");
                                                     let coordinates_3 = &nodal_coordinates
-                                                        [cells_nodes[subcell_a_face_m_subcells[7]]
[…]
