### 2025-06-27 2b82c67
halfway done with edge template 13
