### 2025-06-27 19b295f
halfway done with edge template 13
