### 2025-06-27 3d16f8b
halfway done with edge template 13
