### 2025-06-30 09ac083
restart of dualizaiton discussion
