+++ book/dualization/code/quadtree_plot.py
@@ -1,0 +1,185 @@
+"""This module, quadtree.py, creates a simple quadtree and plots it."""
+
+from pathlib import Path
+from typing import Final
+from typing import NamedTuple
+
+
+import matplotlib.pyplot as plt
+from matplotlib import patches
+
+
+# class Seed(NamedTuple):
+#     """The (x, y) point used to trigger refinement."""
+#     x: float
+#     y: float
+
+class ColorScheme(NamedTuple):
+    """A color scheme for the quadtree.
+    The plasma colorscheme goes from dark purple to bright yellow.  The plasma_r
+    reverses the plasma colorscheme, thus it goes from bright yellow to dark purple.
+    """
+    edgecolor: str
+    facecolor: str
+    zorder: int
+    alpha: float = 0.3
+
+
+def level_to_colorscheme(level: int) -> ColorScheme:
+    """Returns a color scheme based on the level of the quadtree."""
+    colors = [
+        "dimgray", "lightgray", "silver", "gainsboro", "whitesmoke", "white"
+    ]
+    edgecolor = "black"
+    facecolor = colors[level % len(colors)]
+    zorder = 2 + level
+    return ColorScheme(edgecolor, facecolor, zorder)
+
+
+class QuadTree:
+    """Defines a quadtree composed of a single parent quad and recursive
+    children quads.
+    """
+
+    def __init__(self, x, y, width, height, level=0, max_level=2):
+        # (x, y, width, height)
+        self.boundary = (x, y, width, height)
+        self.level = level
+        self.max_level = max_level
+        self.children = []
+        self.subdivide()
+
+    def subdivide(self):
+        """Divides the parent quad into four quad children."""
+        if self.level < self.max_level:
+            x, y, width, height = self.boundary
+            half_width = width / 2.0
+            half_height = height / 2.0
+
+            # Create four children
+            self.children.append(
+                QuadTree(
+                    x,
+                    y,
+                    half_width,
+                    half_height,
+                    self.level + 1,
+                    self.max_level,
+                )
+            )  # Top-left
+            self.children.append(
+                QuadTree(
+                    x + half_width,
+                    y,
+                    half_width,
+                    half_height,
+                    self.level + 1,
+                    self.max_level,
+                )
+            )  # Top-right
+            self.children.append(
+                QuadTree(
+                    x,
+                    y + half_height,
+                    half_width,
+                    half_height,
+                    self.level + 1,
+                    self.max_level,
+                )
+            )  # Bottom-left
+            self.children.append(
+                QuadTree(
+                    x + half_width,
+                    y + half_height,
+                    half_width,
+                    half_height,
+                    self.level + 1,
+                    self.max_level,
+                )
+            )  # Bottom-right
+
+    def draw(self, ax, edgecolor="black", color="dimgray"):
+        """Draw the quadtree."""
+        x, y, width, height = self.boundary
+        # Draw the boundary rectangle
+        rect = patches.Rectangle(
+            (x, y),
+            width,
+            height,
+            # linewidth=1,
+            linestyle="solid",
+            edgecolor="black",
+            # facecolor="dimgray",
+            facecolor=color
[…]
