### 2025-06-30 fcce0c1
vertex template 13
