+++ src/tree/hex/vertex_template_13/mod.rs
@@ -5,7 +5,7 @@ pub const DATA: [[usize; 11]; 8] = [
     [0, 4, 1, 1, 5, 4, 0, 5, 7, 15, 0],   // (0, 4, 1) = (1, 2, 5) = (5, 3, 0)
     [1, 0, 4, 1, 5, 7, 5, 0, 2, 10, 10],  // (1, 0, 4) = (2, 5, 1) = (5, 3, 2)
     [1, 4, 2, 3, 15, 5, 5, 5, 6, 10, 0],  // (1, 4, 2) = (2, 3, 5) = (5, 0, 1)
-    [2, 1, 4, 3, 5, 6, 15, 5, 0, 0, 15], // (2, 1, 4) = (3, 5, 2) = (4, 0, 3)
+    [2, 1, 4, 3, 5, 6, 15, 5, 0, 0, 15],  // (2, 1, 4) = (3, 5, 2) = (4, 0, 3)
     [2, 4, 3, 2, 10, 7, 5, 0, 4, 10, 5],  // (2, 4, 3) = (3, 0, 5) = (5, 1, 2)
     [4, 1, 0, 1, 0, 2, 5, 5, 4, 10, 15],  // (4, 1, 0) = (3, 2, 4) = (0, 5, 3)
     [4, 2, 1, 3, 5, 0, 5, 15, 5, 10, 10], // (4, 2, 1) = (0, 3, 4) = (1, 5, 0)
