### 2025-06-30 d948071
vertex template 14
