### 2025-06-30 91763fe
vertex templates 14-21
