+++ src/tree/hex/vertex_template_21/mod.rs
@@ -1,14 +1,8 @@
 use super::super::{Faces, HEX, Indices, Octree};
 
-pub const DATA: [[usize; 11]; 8] = [
-    [0, 1, 5, 5, 10, 0, 5, 15, 6, 10, 5], // (0, 1, 5) = (3, 4, 0) = (5, 2, 3)
-    [0, 4, 1, 1, 5, 4, 0, 5, 7, 15, 0],   // (0, 4, 1) = (1, 2, 5) = (5, 3, 0)
-    [1, 0, 4, 1, 5, 7, 5, 0, 2, 10, 10],  // (1, 0, 4) = (2, 5, 1) = (5, 3, 2)
-    [1, 4, 2, 3, 15, 5, 5, 5, 6, 10, 0],  // (1, 4, 2) = (2, 3, 5) = (5, 0, 1)
-    [2, 1, 4, 3, 5, 6, 15, 5, 0, 0, 15],  // (2, 1, 4) = (3, 5, 2) = (4, 0, 3)
-    [2, 4, 3, 2, 10, 7, 5, 0, 4, 10, 5],  // (2, 4, 3) = (3, 0, 5) = (5, 1, 2)
-    [4, 1, 0, 1, 0, 2, 5, 5, 4, 10, 15],  // (4, 1, 0) = (3, 2, 4) = (0, 5, 3)
-    [4, 2, 1, 3, 5, 0, 5, 15, 5, 10, 10], // (4, 2, 1) = (0, 3, 4) = (1, 5, 0)
+pub const DATA: [[usize; 11]; 2] = [
+    [0, 1, 5, 5, 10, 0, 5, 15, 6, 10, 3],
+    [0, 4, 1, 1, 5, 4, 0, 5, 7, 15, 2],
 ];
 
 #[allow(clippy::too_many_arguments)]
@@ -23,7 +17,7 @@ pub fn template(
     cell_subsubcell_c_index: usize,
     cell_subcell_c_a_index: usize,
     cell_subsubcell_c_ab_index: usize,
-    cell_subsubcell_c_b_index: usize,
+    cell_subcell_c_b_index: usize,
     cell_faces: &Faces,
     cell_subcells: &Indices,
     cells_nodes: &[usize],
@@ -60,12 +54,8 @@ pub fn template(
                                             if let Some(cell_c_b_index) =
                                                 tree[cell_b_index].get_faces()[face_index]
                                             {
-                                                if let Some(cell_c_b_subsubcells) = tree
-                                                    .cell_subcell_contains_leaves(
-                                                        &tree[cell_c_b_index],
-                                                        face_index,
-                                                        cell_subsubcell_c_b_index,
-                                                    )
+                                                if let Some((cell_c_b_subcells, _)) =
+                                                    tree.cell_contains_leaves(&tree[cell_c_b_index])
                                                 {
                                                     if let Some(cell_c_ab_index) = tree
                                                         [cell_c_a_index]
@@ -92,8 +82,8 @@ pub fn template(
                                                                     [cell_subcell_c_a_index]],
                                                                 cells_nodes[cell_c_ab_subsubcells
                                                                     [cell_subsubcell_c_ab_index]],
-                                                                cells_nodes[cell_c_b_subsubcells
-                                                                    [cell_subsubcell_c_b_index]],
+                                                                cells_nodes[cell_c_b_subcells
+                                                                    [cell_subcell_c_b_index]],
                                                             ]
                                                         })
                                                     } else {
