+++ src/tree/hex/vertex_template_17/mod.rs
@@ -1,14 +1,30 @@
 use super::super::{Faces, HEX, Indices, Octree};
 
-pub const DATA: [[usize; 11]; 8] = [
-    [0, 1, 5, 5, 10, 0, 5, 15, 6, 10, 5], // (0, 1, 5) = (3, 4, 0) = (5, 2, 3)
-    [0, 4, 1, 1, 5, 4, 0, 5, 7, 15, 0],   // (0, 4, 1) = (1, 2, 5) = (5, 3, 0)
-    [1, 0, 4, 1, 5, 7, 5, 0, 2, 10, 10],  // (1, 0, 4) = (2, 5, 1) = (5, 3, 2)
-    [1, 4, 2, 3, 15, 5, 5, 5, 6, 10, 0],  // (1, 4, 2) = (2, 3, 5) = (5, 0, 1)
-    [2, 1, 4, 3, 5, 6, 15, 5, 0, 0, 15],  // (2, 1, 4) = (3, 5, 2) = (4, 0, 3)
-    [2, 4, 3, 2, 10, 7, 5, 0, 4, 10, 5],  // (2, 4, 3) = (3, 0, 5) = (5, 1, 2)
-    [4, 1, 0, 1, 0, 2, 5, 5, 4, 10, 15],  // (4, 1, 0) = (3, 2, 4) = (0, 5, 3)
-    [4, 2, 1, 3, 5, 0, 5, 15, 5, 10, 10], // (4, 2, 1) = (0, 3, 4) = (1, 5, 0)
+pub const DATA: [[usize; 11]; 24] = [
+    [0, 1, 5, 5, 4, 0, 5, 15, 10, 2, 3],
+    [0, 4, 1, 1, 5, 4, 0, 5, 15, 6, 2],
+    [0, 3, 4, 0, 1, 5, 0, 0, 5, 7, 6],
+    [0, 5, 3, 4, 0, 1, 10, 10, 0, 3, 7],
+    [1, 2, 5, 7, 5, 1, 15, 15, 10, 0, 2],
+    [1, 4, 2, 3, 7, 5, 5, 5, 15, 4, 0],
+    [1, 0, 4, 1, 3, 7, 5, 0, 5, 6, 4],
+    [1, 5, 0, 5, 1, 3, 15, 10, 0, 2, 6],
+    [2, 3, 5, 6, 7, 3, 10, 10, 15, 1, 0],
+    [2, 4, 3, 2, 6, 7, 5, 0, 10, 5, 1],
+    [2, 1, 4, 3, 2, 6, 15, 5, 0, 4, 5],
+    [2, 5, 1, 7, 3, 2, 15, 15, 5, 0, 4],
+    [3, 0, 5, 4, 6, 2, 0, 10, 15, 3, 1],
+    [3, 4, 0, 0, 4, 6, 0, 0, 10, 7, 3],
+    [3, 2, 4, 2, 0, 4, 10, 5, 0, 5, 7],
+    [3, 5, 2, 6, 2, 0, 10, 15, 5, 1, 5],
+    [4, 0, 3, 0, 2, 3, 0, 0, 10, 7, 5],
+    [4, 1, 0, 1, 0, 2, 5, 5, 0, 6, 7],
+    [4, 2, 1, 3, 1, 0, 5, 15, 5, 4, 6],
+    [4, 3, 2, 2, 3, 1, 0, 10, 15, 5, 4],
+    [5, 0, 1, 5, 7, 6, 10, 5, 15, 2, 0],
+    [5, 3, 0, 4, 5, 7, 10, 0, 5, 3, 2],
+    [5, 2, 3, 6, 4, 5, 15, 10, 0, 1, 3],
[…]
