+++ book/dualization/code/plasma_gray_colorschemes.py
@@ -2,17 +2,48 @@ import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
 from typing import NamedTuple
 
-import numpy as np
+
+N_COLORS = 8  # Number of discrete colors to extract
 
 
-N_COLORS = 6  # Number of discrete colors to extract
+# Pre-compute the colors for better performance
+def plasma_color_palette(n_colors: int, reversed: bool):
+    """Create a palette of discrete plasma colors."""
+    colormap = plt.cm.plasma_r if reversed else plt.cm.plasma
+    color_indices = [i / (n_colors - 1) for i in range(n_colors)]
+    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
-class ColorScheme(NamedTuple):
-    """A color scheme for the quadtree.
-    The plasma colorscheme goes from dark purple to bright yellow.  The plasma_r
-    reverses the plasma colorscheme, thus it goes from bright yellow to dark purple.
-    """
+def grayscale_color_palette(n_colors: int, reversed: bool):
+    """Create a palette of discrete grayscale colors between 0.05 and 0.95."""
+    # Define the range for the grayscale values
+    min_gray = 0.05
+    max_gray = 0.95
+
+    # Calculate the step size based on the number of colors
+    step = (max_gray - min_gray) / (n_colors - 1)
+
+    # Generate color indices within the specified range
+    color_indices = [min_gray + i * step for i in range(n_colors)]
+
+    # Select the colormap and reverse if needed
+    colormap = plt.cm.gray_r if reversed else plt.cm.gray
+
+    # Convert the grayscale values to hex colors
+    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
+
+
+# Pre-computed color palettes (more efficient if called many times)
+# The plasma colorscheme goes from dark purple to bright yellow.
+# The plasma_r (reversed) goes from bright yellow to dark purple.
+PLASMA_COLORS = plasma_color_palette(N_COLORS, reversed=False)
+PLASMA_R_COLORS = plasma_color_palette(N_COLORS, reversed=True)
+GRAYSCALE_COLORS = grayscale_color_palette(N_COLORS, reversed=False)
+GRAYSCALE_R_COLORS = grayscale_color_palette(N_COLORS, reversed=True)
+
+
+class Color(NamedTuple):
+    """A color for a particular level of a quadtree."""
 
     edgecolor: str
     facecolor: str
@@ -20,81 +51,45 @@ class ColorScheme(NamedTuple):
     alpha: float = 0.8
 
 
-def grayscale_colors(n_colors: int):
-    """Creates n_colors grayscale colors from dark to light.
-
-    Args:
-        n_colors: Number of grayscale colors to generate
-
-    Returns:
-        List of grayscale color names/hex codes
-    """
-    if n_colors <= 0:
-        return []
-    elif n_colors == 1:
-        return ["gray"]
-
-    # Generate grayscale values from dark to light
-    # Using a range that goes from darker grays to lighter grays/white
-    gray_values = np.linspace(0.2, 0.95, n_colors)
-
-    # Convert to hex colors
-    grayscale_colors = [mcolors.to_hex((val, val, val)) for val in gray_values]
-
-    return grayscale_colors
-
-
 def level_to_colorscheme(
-    level: int, n_levels, use_plasma: bool = True, reverse: bool = False
-) -> ColorScheme:
-    """Returns a color scheme based on the level of the quadtree.
+    level: int, n_levels: int, plasma: bool, reversed: bool
+) -> Color:
[…]
