### 2025-07-01 2740217
slow is smooth, smooth is fast
