+++ book/dualization/code/colorschemes.py
@@ -4,7 +4,33 @@ import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
 from matplotlib import patches
 
-N_COLORS = 8  # Number of discrete colors to extract
+
+class QuadColors:
+    """A collection of color schemes for quadtree levels."""
+
+    def __init__(
+        self,
+        n_levels: int,
+        edgecolor: str,
+        alpha: float,
+        plasma: bool = True,
+        reversed: bool = False,
+    ):
+        self.n_levels = n_levels
+        self.edgecolor = edgecolor
+        self.alpha = alpha
+        self.plasma = plasma
+        self.reversed = reversed
+
+        match (plasma, reversed):
+            case (True, True):
+                self.facecolors = plasma_color_palette(n_levels, reversed=True)
+            case (True, False):
+                self.facecolors = plasma_color_palette(n_levels, reversed=False)
+            case (False, True):
+                self.facecolors = grayscale_color_palette(n_levels, reversed=True)
+            case (False, False):
+                self.facecolors = grayscale_color_palette(n_levels, reversed=False)
 
 
 # Pre-compute the colors for better performance
@@ -15,6 +41,7 @@ def plasma_color_palette(n_colors: int, reversed: bool):
     return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
+# Pre-compute the grays for better performance
 def grayscale_color_palette(n_colors: int, reversed: bool):
     """Create a palette of discrete grayscale colors between 0.05 and 0.95."""
     # Define the range for the grayscale values
@@ -34,140 +61,124 @@ def grayscale_color_palette(n_colors: int, reversed: bool):
     return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
-# Pre-computed color palettes (more efficient if called many times)
-# The plasma colorscheme goes from dark purple to bright yellow.
-# The plasma_r (reversed) goes from bright yellow to dark purple.
-PLASMA_COLORS = plasma_color_palette(N_COLORS, reversed=False)
-PLASMA_R_COLORS = plasma_color_palette(N_COLORS, reversed=True)
-GRAYSCALE_COLORS = grayscale_color_palette(N_COLORS, reversed=False)
-GRAYSCALE_R_COLORS = grayscale_color_palette(N_COLORS, reversed=True)
-
-
-class Color(NamedTuple):
-    """A color for a particular level of a quadtree."""
-
-    edgecolor: str
-    facecolor: str
-    zorder: int
-    alpha: float = 0.8
-
-
-def level_to_colorscheme(
-    level: int, n_levels: int, plasma: bool, reversed: bool
-) -> Color:
-    """Returns a color based on the level of the quadtree.
-
-    Args:
-        level: The level of the quadtree node
-        n_levels: Total number of levels in the quadtree
-        plasma: If True, use plasma colors; if False, use original grayscale
-        reversed: If True, use reversed plasma or reversed grayscale
-    """
-    match (plasma, reversed):
-        case (True, True):
-            facecolor = PLASMA_R_COLORS[level]
-        case (True, False):
-            facecolor = PLASMA_COLORS[level]
-        case (False, True):
-            facecolor = GRAYSCALE_R_COLORS[level]
-        case (False, False):
-            facecolor = GRAYSCALE_COLORS[level]
-
-    edgecolor = "black"
-    zorder = 2 + level
-    return Color(edgecolor, facecolor, zorder)
-
-
[…]
