### 2025-07-01 0667690
pause for the night
