+++ book/dualization/code/quadtree_plot.py
@@ -2,12 +2,15 @@
 
 from pathlib import Path
 from typing import Final
-from typing import NamedTuple
+# from typing import NamedTuple
 
 
 import matplotlib.pyplot as plt
 from matplotlib import patches
 
+from colorschemes import QuadColors
+
+# from colorschemes import Color, level_to_color
 
 # class Seed(NamedTuple):
 #     """The (x, y) point used to trigger refinement."""
@@ -15,25 +18,25 @@ from matplotlib import patches
 #     y: float
 
 
-class ColorScheme(NamedTuple):
-    """A color scheme for the quadtree.
-    The plasma colorscheme goes from dark purple to bright yellow.  The plasma_r
-    reverses the plasma colorscheme, thus it goes from bright yellow to dark purple.
-    """
-
-    edgecolor: str
-    facecolor: str
-    zorder: int
-    alpha: float = 0.3
-
-
-def level_to_colorscheme(level: int) -> ColorScheme:
-    """Returns a color scheme based on the level of the quadtree."""
-    colors = ["dimgray", "lightgray", "silver", "gainsboro", "whitesmoke", "white"]
-    edgecolor = "black"
-    facecolor = colors[level % len(colors)]
-    zorder = 2 + level
-    return ColorScheme(edgecolor, facecolor, zorder)
+# class ColorScheme(NamedTuple):
+#     """A color scheme for the quadtree.
+#     The plasma colorscheme goes from dark purple to bright yellow.  The plasma_r
+#     reverses the plasma colorscheme, thus it goes from bright yellow to dark purple.
+#     """
+#
+#     edgecolor: str
+#     facecolor: str
+#     zorder: int
+#     alpha: float = 0.3
+#
+#
+# def level_to_colorscheme(level: int) -> ColorScheme:
+#     """Returns a color scheme based on the level of the quadtree."""
+#     colors = ["dimgray", "lightgray", "silver", "gainsboro", "whitesmoke", "white"]
+#     edgecolor = "black"
+#     facecolor = colors[level % len(colors)]
+#     zorder = 2 + level
+#     return ColorScheme(edgecolor, facecolor, zorder)
 
 
 class QuadTree:
@@ -46,17 +49,22 @@ class QuadTree:
         self.boundary = (x, y, width, height)
         self.level = level
         self.max_level = max_level
+        self.has_children = False
         self.children = []
         self.subdivide()
 
     def subdivide(self):
         """Divides the parent quad into four quad children."""
-        if self.level < self.max_level:
+        if self.level <= self.max_level:
+            print(
+                f"Subdividing quad at level {self.level} with boundary {self.boundary}"
+            )
             x, y, width, height = self.boundary
             half_width = width / 2.0
             half_height = height / 2.0
 
             # Create four children
+            self.has_children = True  # overwrite
             self.children.append(
                 QuadTree(
                     x,
@@ -98,74 +106,79 @@ class QuadTree:
                 )
             )  # Bottom-right
 
-    def draw(self, ax, edgecolor="black", color="dimgray"):
+    def draw(self, ax, quadcolors: QuadColors):
         """Draw the quadtree."""
         x, y, width, height = self.boundary
         # Draw the boundary rectangle
+        print(
[…]
