### 2025-07-01 6ae65d1
add error checking
