+++ book/dualization/code/quadtree_colors.py
@@ -32,29 +32,40 @@ class QuadColors:
                 self.facecolors = grayscale_color_palette(n_levels, reversed=False)
 
 
+def n_colors_valid(n_colors: int) -> bool:
+    """Check if the number of colors is valid.  Ensure n_colors is
+    at least 2 for a valid palette.
+    """
+    if n_colors < 2:
+        raise ValueError(f"n_colors {n_colors} must be at least 2 for a color palette.")
+    return True
+
+
 # Pre-compute the colors for better performance
 def plasma_color_palette(n_colors: int, reversed: bool):
     """Create a palette of discrete plasma colors."""
-    colormap = plt.cm.plasma_r if reversed else plt.cm.plasma
-    color_indices = [i / (n_colors - 1) for i in range(n_colors)]
-    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
+    if n_colors_valid(n_colors):
+        colormap = plt.cm.plasma_r if reversed else plt.cm.plasma
+        color_indices = [i / (n_colors - 1) for i in range(n_colors)]
+        return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
 # Pre-compute the grays for better performance
 def grayscale_color_palette(n_colors: int, reversed: bool):
     """Create a palette of discrete grayscale colors between 0.05 and 0.95."""
     # Define the range for the grayscale values
-    min_gray = 0.05
-    max_gray = 0.95
+    if n_colors_valid(n_colors):
+        min_gray = 0.05
+        max_gray = 0.95
 
-    # Calculate the step size based on the number of colors
-    step = (max_gray - min_gray) / (n_colors - 1)
+        # Calculate the step size based on the number of colors
+        step = (max_gray - min_gray) / (n_colors - 1)
 
-    # Generate color indices within the specified range
-    color_indices = [min_gray + i * step for i in range(n_colors)]
+        # Generate color indices within the specified range
+        color_indices = [min_gray + i * step for i in range(n_colors)]
 
-    # Select the colormap and reverse if needed
-    colormap = plt.cm.gray_r if reversed else plt.cm.gray
+        # Select the colormap and reverse if needed
+        colormap = plt.cm.gray_r if reversed else plt.cm.gray
 
-    # Convert the grayscale values to hex colors
-    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
+        # Convert the grayscale values to hex colors
+        return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
