### 2025-07-01 a3acdf7
more error checking
