+++ book/dualization/code/quadtree_colors.py
@@ -30,6 +30,9 @@ class QuadColors:
                 self.facecolors = grayscale_color_palette(n_levels, reversed=True)
             case (False, False):
                 self.facecolors = grayscale_color_palette(n_levels, reversed=False)
+            case _:
+                # Catch-all fallback, shouldn't happen with current logic
+                self.facecolors = ["#FF00FF"] * n_levels  # magenta for debugging
 
 
 def n_colors_valid(n_colors: int) -> bool:
@@ -42,30 +45,31 @@ def n_colors_valid(n_colors: int) -> bool:
 
 
 # Pre-compute the colors for better performance
-def plasma_color_palette(n_colors: int, reversed: bool):
+def plasma_color_palette(n_colors: int, reversed: bool) -> list[str]:
     """Create a palette of discrete plasma colors."""
-    if n_colors_valid(n_colors):
-        colormap = plt.cm.plasma_r if reversed else plt.cm.plasma
-        color_indices = [i / (n_colors - 1) for i in range(n_colors)]
-        return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
+    assert n_colors_valid(n_colors)
+    colormap = plt.cm.plasma_r if reversed else plt.cm.plasma
+    color_indices = [i / (n_colors - 1) for i in range(n_colors)]
+    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
 # Pre-compute the grays for better performance
-def grayscale_color_palette(n_colors: int, reversed: bool):
+def grayscale_color_palette(n_colors: int, reversed: bool) -> list[str]:
     """Create a palette of discrete grayscale colors between 0.05 and 0.95."""
     # Define the range for the grayscale values
-    if n_colors_valid(n_colors):
-        min_gray = 0.05
-        max_gray = 0.95
+    assert n_colors_valid(n_colors)
 
-        # Calculate the step size based on the number of colors
-        step = (max_gray - min_gray) / (n_colors - 1)
+    min_gray = 0.05
+    max_gray = 0.95
 
-        # Generate color indices within the specified range
-        color_indices = [min_gray + i * step for i in range(n_colors)]
+    # Calculate the step size based on the number of colors
+    step = (max_gray - min_gray) / (n_colors - 1)
 
-        # Select the colormap and reverse if needed
-        colormap = plt.cm.gray_r if reversed else plt.cm.gray
+    # Generate color indices within the specified range
+    color_indices = [min_gray + i * step for i in range(n_colors)]
 
-        # Convert the grayscale values to hex colors
-        return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
+    # Select the colormap and reverse if needed
+    colormap = plt.cm.gray_r if reversed else plt.cm.gray
+
+    # Convert the grayscale values to hex colors
+    return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
