### 2025-07-01 22b92b2
base quadtree with colors working again
