+++ book/dualization/code/quadtree_plot.py
@@ -16,18 +16,21 @@ class QuadTree:
     children quads.
     """
 
-    def __init__(self, x, y, width, height, level=0, max_level=2):
+    def __init__(self, *, x, y, width, height, level=0, max_level=2):
         # (x, y, width, height)
         self.boundary = (x, y, width, height)
         self.level = level
         self.max_level = max_level
         self.has_children = False
         self.children = []
+        assert level <= max_level, (
+            f"QuadTree level {level} exceeds max_level {max_level}."
+        )
         self.subdivide()
 
     def subdivide(self):
         """Divides the parent quad into four quad children."""
-        if self.level <= self.max_level:
+        if self.level < self.max_level:
             print(
                 f"Subdividing quad at level {self.level} with boundary {self.boundary}"
             )
@@ -39,42 +42,42 @@ class QuadTree:
             self.has_children = True  # overwrite
             self.children.append(
                 QuadTree(
-                    x,
-                    y,
-                    half_width,
-                    half_height,
-                    self.level + 1,
-                    self.max_level,
+                    x=x,
+                    y=y,
+                    width=half_width,
+                    height=half_height,
+                    level=self.level + 1,
+                    max_level=self.max_level,
                 )
             )  # Top-left
             self.children.append(
                 QuadTree(
-                    x + half_width,
-                    y,
-                    half_width,
-                    half_height,
-                    self.level + 1,
-                    self.max_level,
+                    x=x + half_width,
+                    y=y,
+                    width=half_width,
+                    height=half_height,
+                    level=self.level + 1,
+                    max_level=self.max_level,
                 )
             )  # Top-right
             self.children.append(
                 QuadTree(
-                    x,
-                    y + half_height,
-                    half_width,
-                    half_height,
-                    self.level + 1,
-                    self.max_level,
+                    x=x,
+                    y=y + half_height,
+                    width=half_width,
+                    height=half_height,
+                    level=self.level + 1,
+                    max_level=self.max_level,
                 )
             )  # Bottom-left
             self.children.append(
                 QuadTree(
-                    x + half_width,
-                    y + half_height,
-                    half_width,
-                    half_height,
-                    self.level + 1,
-                    self.max_level,
+                    x=x + half_width,
+                    y=y + half_height,
+                    width=half_width,
+                    height=half_height,
+                    level=self.level + 1,
+                    max_level=self.max_level,
                 )
             )  # Bottom-right
 
@@ -107,49 +110,51 @@ class QuadTree:
 
 def main():
     # User input begin
-    MAX_LEVEL: Final[int] = 2
+    level_min: Final[int] = 0
+    level_max: Final[int] = 1
     SAVE: Final[bool] = True
     SHOW: Final[bool] = True
     EXT: Final[str] = ".png"  # ".pdf" | ".png" | ".svg"
     DPI: Final[int] = 300
 
-    XMIN = -2  # 1
-    XMAX = 2  # 3
-    YMIN = -2  # -1
[…]
