### 2025-07-01 b27e3b0
a working colorized quadtree
