+++ book/dualization/code/quadtree_plot.py
@@ -1,14 +1,39 @@
-"""This module, quadtree.py, creates a simple quadtree and plots it."""
+"""This module creates a quadtree and plots it."""
 
 from pathlib import Path
-from typing import Final
-# from typing import NamedTuple
+from typing import Final, NamedTuple
 
 
 import matplotlib.pyplot as plt
 from matplotlib import patches
 
-from colorschemes import QuadColors
+# from book.dualization.code.color_schemes import QuadColors
+from color_complement import ColorComplement
+from color_schemes import QuadColors
+
+
+class Point(NamedTuple):
+    """A point in 2D space."""
+
+    x: float  # x-coordinate
+    y: float  # y-coordinate
+
+
+class Seeds(NamedTuple):
+    """A collection of points (seeds) in 2D space."""
+
+    points: list[Point]  # List of Point objects
+
+
+class Boundary(NamedTuple):
+    """A boundary defined by its minimum and maximum
+    x and y coordinates."""
+
+    xmin: float  # Minimum x-coordinate
+    xmax: float  # Maximum x-coordinate
+
+    ymin: float  # Minimum y-coordinate
+    ymax: float  # Maximum y-coordinate
 
 
 class QuadTree:
@@ -16,9 +41,20 @@ class QuadTree:
     children quads.
     """
 
-    def __init__(self, *, x, y, width, height, level=0, max_level=2, verbose=False):
+    def __init__(
+        self,
+        *,
+        x: float,
+        y: float,
+        width: float,
+        height: float,
+        level: int,
+        max_level: int,
+        seeds: list[Point],
+        verbose: bool,
+    ):
         # (x, y, width, height)
-        self.boundary = (x, y, width, height)
+        self.boundary = Boundary(xmin=x, xmax=x + width, ymin=y, ymax=y + height)
         self.level = level
         self.max_level = max_level
         self.has_children = False
@@ -27,21 +63,28 @@ class QuadTree:
             f"QuadTree level {level} exceeds max_level {max_level}."
         )
         self.verbose = verbose
-        self.subdivide()
 
-    def subdivide(self):
+        if self.contains_any_point(seeds):
+            # If the quad contains any of the seed points, subdivide it
+            self.subdivide(seeds=seeds)
+
+    def subdivide(self, seeds: list[Point]):
         """Divides the parent quad into four quad children."""
         if self.level < self.max_level:
             if self.verbose:
                 print(
                     f"Subdividing quad at level {self.level} with boundary {self.boundary}"
                 )
-            x, y, width, height = self.boundary
+            x = self.boundary.xmin
+            y = self.boundary.ymin
+            width = self.boundary.xmax - self.boundary.xmin
+            height = self.boundary.ymax - self.boundary.ymin
             half_width = width / 2.0
             half_height = height / 2.0
 
-            # Create four children
             self.has_children = True  # overwrite
+
+            # Create four children
             self.children.append(
                 QuadTree(
                     x=x,
@@ -50,6 +93,8 @@ class QuadTree:
                     height=half_height,
                     level=self.level + 1,
                     max_level=self.max_level,
+                    seeds=seeds,
+                    verbose=self.verbose,
                 )
             )  # Top-left
             self.children.append(
@@ -60,6 +105,8 @@ class QuadTree:
                     height=half_height,
                     level=self.level + 1,
                     max_level=self.max_level,
+                    seeds=seeds,
+                    verbose=self.verbose,
                 )
             )  # Top-right
             self.children.append(
@@ -70,6 +117,8 @@ class QuadTree:
                     height=half_height,
                     level=self.level + 1,
[…]
