### 2025-07-01 e3ed20c
add Tab color scheme
