+++ book/dualization/code/color_schemes.py
@@ -1,30 +1,31 @@
-"""This module plots discrete grayscale and plasma color schemes."""
+"""This module plots discrete grayscale, plasma and viridis color schemes."""
 
 import matplotlib.pyplot as plt
 from matplotlib import patches
 
-from quadtree_colors import QuadColors
+from quadtree_colors import QuadColors, ColorScheme
 
 
 def plot_color_schemes(n_colors: int, alpha: float):
     """Shows the different color schemes."""
 
-    fig, axes = plt.subplots(1, 4, figsize=(15, 5))
+    fig, axes = plt.subplots(2, 4, figsize=(10, 5))
 
     # Create sample rectangles to show the colors
     levels = range(n_colors)
     # n_levels = n_colors
 
+    # Row 1: Original color schemes
     # Grayscale
     grayscale = QuadColors(
         n_levels=n_colors,
         edgecolor="black",
         alpha=alpha,
-        plasma=False,
+        color_scheme=ColorScheme.GRAYSCALE,
         reversed=False,
     )
     print("Grayscale colors:", grayscale.facecolors)
-    ax1 = axes[0]
+    ax0 = axes[0, 0]
     for i, level in enumerate(levels):
         rect = patches.Rectangle(
             (0, i),
@@ -34,53 +35,25 @@ def plot_color_schemes(n_colors: int, alpha: float):
             edgecolor=grayscale.edgecolor,
             alpha=grayscale.alpha,
         )
-        ax1.add_patch(rect)
-        ax1.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
+        ax0.add_patch(rect)
+        ax0.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
 
-    ax1.set_xlim(-0.1, 1.1)
-    ax1.set_ylim(-0.1, n_colors)
-    ax1.set_title("Grayscale")
-    ax1.set_xticks([])
-    ax1.set_yticks([])
+    ax0.set_xlim(-0.1, 1.1)
+    ax0.set_ylim(-0.1, n_colors)
+    ax0.set_title("Grayscale")
+    ax0.set_xticks([])
+    ax0.set_yticks([])
 
-    # Grayscale reversed
-    grayscale_reversed = QuadColors(
-        n_levels=n_colors,
-        edgecolor="black",
-        alpha=0.8,
-        plasma=False,
-        reversed=True,
-    )
-    print("Grayscale colors reversed:", grayscale_reversed.facecolors)
-    ax2 = axes[1]
-    for i, level in enumerate(levels):
-        rect = patches.Rectangle(
-            (0, i),
-            1,
-            0.8,
-            facecolor=grayscale_reversed.facecolors[level],
-            edgecolor=grayscale_reversed.edgecolor,
-            alpha=grayscale_reversed.alpha,
-        )
-        ax2.add_patch(rect)
-        ax2.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
-
-    ax2.set_xlim(-0.1, 1.1)
-    ax2.set_ylim(-0.1, n_colors)
-    ax2.set_title("Grayscale Reversed")
-    ax2.set_xticks([])
-    ax2.set_yticks([])
-
-    # Plasma colors
+    # Plasma
     plasma = QuadColors(
         n_levels=n_colors,
         edgecolor="black",
-        alpha=0.8,
-        plasma=True,
+        alpha=alpha,
+        color_scheme=ColorScheme.PLASMA,
         reversed=False,
     )
     print("Plasma colors:", plasma.facecolors)
-    ax3 = axes[2]
+    ax1 = axes[0, 1]
[…]
