+++ book/dualization/code/quadtree_colors.py
@@ -1,7 +1,18 @@
 """This module provides color schemes for visualizing quadtree levels."""
 
+from typing import NamedTuple
+
 import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
+
+
+class ColorScheme(NamedTuple):
+    """Named tuple for color scheme options."""
+
+    GRAYSCALE: str = "grayscale"
+    PLASMA: str = "plasma"
+    TAB10: str = "tab10"
+    VIRIDIS: str = "viridis"
 
 
 class QuadColors:
@@ -12,26 +23,39 @@ class QuadColors:
         n_levels: int,
         edgecolor: str,
         alpha: float,
-        plasma: bool = True,
+        color_scheme: str,
         reversed: bool = False,
     ):
         self.n_levels = n_levels
         self.edgecolor = edgecolor
         self.alpha = alpha
-        self.plasma = plasma
+        self.color_scheme = color_scheme
         self.reversed = reversed
 
-        match (plasma, reversed):
-            case (True, True):
-                self.facecolors = plasma_color_palette(n_levels, reversed=True)
-            case (True, False):
-                self.facecolors = plasma_color_palette(n_levels, reversed=False)
-            case (False, True):
-                self.facecolors = grayscale_color_palette(n_levels, reversed=True)
-            case (False, False):
-                self.facecolors = grayscale_color_palette(n_levels, reversed=False)
+        # Validate color scheme
+        valid_schemes = [
+            ColorScheme.GRAYSCALE,
+            ColorScheme.PLASMA,
+            ColorScheme.TAB10,
+            ColorScheme.VIRIDIS,
+        ]
+        if color_scheme not in valid_schemes:
+            raise ValueError(
+                f"color_scheme must be one of {valid_schemes}, got '{color_scheme}'"
+            )
+
+        # Generate colors based on scheme
+        match color_scheme:
+            case ColorScheme.GRAYSCALE:
+                self.facecolors = grayscale_color_palette(n_levels, reversed=reversed)
+            case ColorScheme.PLASMA:
+                self.facecolors = plasma_color_palette(n_levels, reversed=reversed)
+            case ColorScheme.TAB10:
+                self.facecolors = tab10_color_palette(n_levels, reversed=reversed)
+            case ColorScheme.VIRIDIS:
+                self.facecolors = viridis_color_palette(n_levels, reversed=reversed)
             case _:
-                # Catch-all fallback, shouldn't happen with current logic
+                # Catch-all fallback, shouldn't happen with validation above
                 self.facecolors = ["#FF00FF"] * n_levels  # magenta for debugging
 
 
@@ -44,7 +68,6 @@ def n_colors_valid(n_colors: int) -> bool:
     return True
 
 
-# Pre-compute the colors for better performance
 def plasma_color_palette(n_colors: int, reversed: bool) -> list[str]:
     """Create a palette of discrete plasma colors."""
     assert n_colors_valid(n_colors)
@@ -53,7 +76,6 @@ def plasma_color_palette(n_colors: int, reversed: bool) -> list[str]:
     return [mcolors.to_hex(colormap(idx)) for idx in color_indices]
 
 
-# Pre-compute the grays for better performance
 def grayscale_color_palette(n_colors: int, reversed: bool) -> list[str]:
     """Create a palette of discrete grayscale colors between 0.05 and 0.95."""
     # Define the range for the grayscale values
@@ -72,4 +94,45 @@ def grayscale_color_palette(n_colors: int, reversed: bool) -> list[str]:
     colormap = plt.cm.gray_r if reversed else plt.cm.gray
 
[…]
