### 2025-07-01 4a99f2e
quadtree docs
