+++ book/dualization/code/color_schemes.py
@@ -1,249 +1,139 @@
-"""This module plots discrete grayscale, plasma and viridis color schemes."""
+"""This module provides color schemes for visualizing quadtree levels."""
 
+from typing import NamedTuple
+
+import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
-from matplotlib import patches
-
-from quadtree_colors import QuadColors, ColorScheme
 
 
-def plot_color_schemes(n_colors: int, alpha: float):
-    """Shows the different color schemes."""
+class ColorSchemes(NamedTuple):
+    """Named tuple for color scheme options."""
 
-    fig, axes = plt.subplots(2, 4, figsize=(10, 5))
-
-    # Create sample rectangles to show the colors
-    levels = range(n_colors)
-    # n_levels = n_colors
-
-    # Row 1: Original color schemes
-    # Grayscale
-    grayscale = QuadColors(
-        n_levels=n_colors,
-        edgecolor="black",
-        alpha=alpha,
-        color_scheme=ColorScheme.GRAYSCALE,
-        reversed=False,
-    )
-    print("Grayscale colors:", grayscale.facecolors)
-    ax0 = axes[0, 0]
-    for i, level in enumerate(levels):
-        rect = patches.Rectangle(
-            (0, i),
-            1,
-            0.8,
-            facecolor=grayscale.facecolors[level],
-            edgecolor=grayscale.edgecolor,
-            alpha=grayscale.alpha,
-        )
-        ax0.add_patch(rect)
-        ax0.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
-
-    ax0.set_xlim(-0.1, 1.1)
-    ax0.set_ylim(-0.1, n_colors)
-    ax0.set_title("Grayscale")
-    ax0.set_xticks([])
-    ax0.set_yticks([])
-
-    # Plasma
-    plasma = QuadColors(
-        n_levels=n_colors,
-        edgecolor="black",
-        alpha=alpha,
-        color_scheme=ColorScheme.PLASMA,
-        reversed=False,
-    )
-    print("Plasma colors:", plasma.facecolors)
-    ax1 = axes[0, 1]
-    for i, level in enumerate(levels):
-        rect = patches.Rectangle(
-            (0, i),
-            1,
-            0.8,
-            facecolor=plasma.facecolors[level],
-            edgecolor=plasma.edgecolor,
-            alpha=plasma.alpha,
-        )
-        ax1.add_patch(rect)
-        ax1.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
-    ax1.set_xlim(-0.1, 1.1)
-    ax1.set_ylim(-0.1, n_colors)
-    ax1.set_title("Plasma")
-    ax1.set_xticks([])
-    ax1.set_yticks([])
-
-    # Tab10 (default matplotlib color scheme)
-    tab10 = QuadColors(
-        n_levels=n_colors,
-        edgecolor="black",
-        alpha=alpha,
-        color_scheme=ColorScheme.TAB10,
-        reversed=False,
-    )
-    print("Tab10 colors:", tab10.facecolors)
-    ax2 = axes[0, 2]
-    for i, level in enumerate(levels):
-        rect = patches.Rectangle(
-            (0, i),
-            1,
-            0.8,
-            facecolor=tab10.facecolors[level],
-            edgecolor=tab10.edgecolor,
-            alpha=tab10.alpha,
-        )
-        ax2.add_patch(rect)
-        ax2.text(0.5, i + 0.4, f"Level {level}", ha="center", va="center")
-    ax2.set_xlim(-0.1, 1.1)
-    ax2.set_ylim(-0.1, n_colors)
-    ax2.set_title("Tab10")
-    ax2.set_xticks([])
-    ax2.set_yticks([])
-
[…]

[…]
