### 2025-07-02 38d7e78
refactor with a configuration
