+++ book/dualization/code/quadtree_plot.py
@@ -20,12 +20,6 @@ class Point(NamedTuple):
     y: float  # y-coordinate
 
 
-class Seeds(NamedTuple):
-    """A collection of points (seeds) in 2D space."""
-
-    points: list[Point]  # List of Point objects
-
-
 class Boundary(NamedTuple):
     """A boundary defined by its minimum and maximum
     x and y coordinates."""
@@ -210,24 +204,33 @@ class QuadTree:
             )
 
 
-def main():
-    # User input begin
-    level_min: Final[int] = 0
-    level_max: Final[int] = 5
-    alpha = 1.0
-    SAVE: Final[bool] = True
-    SHOW: Final[bool] = True
-    EXT: Final[str] = ".svg"  # ".pdf" | ".png" | ".svg"
-    DPI: Final[int] = 300
-    fig_width, fig_height = 6, 6  # inches, inches
+class Configuration(NamedTuple):
+    """User input configuration for the quadtree plot."""
 
-    xmin = -2
-    xmax = 6
-    ymin = -2
-    ymax = 6
-    width = xmax - xmin
-    height = ymax - ymin
-    verbose = False  # Set to True to see debug output
+    xmin: float  # Minimum x-coordinate for the quadtree
+    xmax: float  # Maximum x-coordinate for the quadtree
+    ymin: float  # Minimum y-coordinate for the quadtree
+    ymax: float  # Maximum y-coordinate for the quadtree
+
+    level_min: int  # Minimum level of the quadtree
+    level_max: int  # Maximum level of the quadtree
+
+    seeds: list[Point]  # List of seed points for the quadtree
+
+    alpha: float  # Transparency of the quadtree colors
+    save: bool  # Whether to save the plot
+    show: bool  # Whether to show the plot
+    dpi: int = 300  # Dots per inch for saving the plot
+    fig_width: float = 6.0  # Width of the figure in inches
+    fig_height: float = 6.0  # Height of the figure in inches
+    ext: str = ".svg"  # File extension for saving the plot
+
+    verbose: bool = False  # Whether to print debug information
+
+
+def quarter_plate_seeds() -> list[Point]:
+    """Helper function to create seeds for the Hughes quarter plate example."""
+
     # Similar to the round in the Hughes quarter plate problem
     # https://github.com/sandialabs/sibl/blob/master/geo/doc/dual/lesson_11.md
     # see also Cottrell 2009 IGA book, page 117.
@@ -237,10 +240,6 @@ def main():
     n_points = 9
     theta_values = np.linspace(theta_start, theta_stop, n_points)
     offset_x, offset_y = 4.0, 0.0
-    # seeds = [
-    #     Point(x=2.6, y=0.6),
-    #     Point(x=2.9, y=0.2),
-    # ]
     seeds = [
         Point(x=radius * np.cos(theta) + offset_x, y=radius * np.sin(theta) + offset_y)
         for theta in theta_values
@@ -251,21 +250,68 @@ def main():
         Point(x=0, y=0),
     ]
     seeds += corner_seeds
+    return seeds
+
+
+def main():
+    # Circle example
+
+    # Hughes quarter plate example
+    cc = Configuration(
+        xmin=-2,
+        xmax=6,
+        ymin=-2,
+        ymax=6,
+        #
+        level_min=0,
+        level_max=5,
+        #
+        seeds=quarter_plate_seeds(),  # type: ignore
+        #
+        alpha=1.0,
+        save=True,
+        show=True,
+        dpi=300,
[…]
