### 2025-07-02 3f2dee3
lots of refactors
