+++ book/dualization/README.md
@@ -28,6 +28,26 @@ Level 0 | 1 | 2
 --- | --- | ---
 ![](mwe/plot_quadtree_L3.png) | ![](mwe/plot_quadtree_L4.png) | ![](mwe/plot_quadtree_L5.png)
 
+### Circle from Segmentation
+
+3 | 4 | 5 | 6
+--- | --- | --- | ---
+![](code/circle_segmentation_diam_3.svg) | ![](code/circle_segmentation_diam_4.svg) | ![](code/circle_segmentation_diam_5.svg) | ![](code/circle_segmentation_diam_6.svg)
+
+13 | 14 | 15 | 16
+--- | --- | --- | ---
+![](code/circle_segmentation_diam_13.svg) | ![](code/circle_segmentation_diam_14.svg) | ![](code/circle_segmentation_diam_15.svg) | ![](code/circle_segmentation_diam_16.svg)
+
+![](code/circle_segmentation_diam_100.svg)
+
+### Circle from Boundary
+
+Consider a boundary of a circle defined by discrete `(x, y)` points.
+
+![](code/circle_loop_r_50_npts_36.svg)
+
+### Circle from Tesellation
+
 ### Quarter Plate
 
 With [Python](#source), we produce a Quadtree with zero to five levels of refinement.  Refinement is triggered based on whether or not a cell contains one or more seed points, shown as points along the quarter circle centered at `(4, 0)`.
@@ -42,9 +62,6 @@ Level 0 | 1 | 2
 
 ## Octree
 
-## Circle
-
-Consider a boundary of a circle defined by discrete `(x, y)` points.
 
 ## Sphere
 
