### 2025-07-02 df72e9c
trim trailing white space
