### 2025-07-02 6b4a2e7
note about ruff formatter
