### 2025-07-02 83b8163
nodes at centers of cells do not need to be in node_map
