### 2025-07-03 effa2bb
subdivision criterion
