### 2025-07-03 be3962d
a start on quadtree
