### 2025-07-21 #8467 @Andlon
Hi @finsberg, thanks for the shout-out. This sounds interesting but unfortunately because of personal reasons I am unable to take on commitments of any form at the moment.
