+++ Cargo.toml
@@ -15,7 +15,7 @@ version = "0.3.4"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-clap = { version = "=4.5.41", features = ["derive"] }
+clap = { version = "=4.5.42", features = ["derive"] }
 chrono = "=0.4.41"
 conspire = { version = "=0.5.7", features = ["fem"] }
 ndarray = { version = "=0.16.1", features = ["rayon"] }
