+++ sandbox/logs.md
@@ -1,711 +1,0 @@
-# logs
-
-**Goal:** A command line Rust application to produce ABAQUS and Exodus outputs.
-
-*In order of most recent to least recent.*
-
-## 2025-05-16
-
-* Rayon
-  * ndarray also has parallel
-  * multi-threading API
-* Tets
-  * `dual` change to `adaptive` in API
-  * Structured
-  * Unstructured
-  * Refactor
-* JOSS 2025
-  * Papers
-    * All papers
-    * Active papers
-* Preparation for ONR annual review
-* `autosim` workflow
-  * `.inp` update
-
-## 2025-03-30
-
-* Extraction
-  * mdbook update?
-* Concurrency
-  * looked at rayon, https://docs.rs/rayon/latest/rayon/
-  * current not needed
-  * defeaturing is currently the slowest
-  * surface triangularization
-* isosurface
-  * still working toward full-manifold
-  * lurking challenge (?) multi-material with non-manifold
-* JOSS submission
-  * Defeaturing
-    * canonical defeature A and B figure
-    * mdbook in need of update
-* Funding
-  * tbd
-* ONR annual review
-  * slides
-  * dates of attendance
-* Matrix multiplication
-
-## 2025-01-29
-
-* Goals
-  * "skin" of a segmentation -> quads -> (2 tris) -> stl (isosurface)
-    * similar to "dual marching cubes"
-    * smoothing of any generic edge connectivity (adjacency matrix in graph theory?)
-    * tris tests (CBH) - test and documentation
-    * stl -> tri (MRB) - implementation and architecture
-  * octree
-    * segementation - balanced quad tree origami "skin" surface -> quad surface -> tri surface
-    * 6 templates conference paper, turn in to connected trianges, https://www.cs.umd.edu/users/hjs//pubs/reddyasae01.pdf, page 12
-  * defeaturing - currently paused
-  * dualization - paused
-
-## 2025-01-22
-
-### Toward isosurfaces
-
-* **Option 1** (NW example)
-  * origami quadtree -> trimesh (.stl) -> Taubin smoothing
-* **Option 2** (NE example)
-  * marching cubes from *dual* segmentation
-
-### Questions
-
-* volume preservation (avoid shrinking volumes)
-* interface conformity (surface-to-surface matching)
-
-### TODO
-
-* MRB - refactor smoothing on generalized types, such as tris, quads, and hexes
-* CBH - smoothing unit tests
-  * The actual tessellation example from the Taubin paper
-  * Single 3D cube into 12 triangles, then 36 quads, then continued refinement
-* CBH - Sculpt to generate an isosurface without smoothing on a sphere or cube
-* CBH - continue smoothed simulations, compare with conforming and voxel cases
-
-## 2025-01-02
-
-* schedule meeting
-  * value proposition: open, auto, fast
-* private client repo
-* pause dualization, concerns about resulting mesh quality
-* defeaturing
-* surface reconstruction
-  * `npy -> stl`
-  * marching cubes
-  * dual marching cubes
-  * dual contouring (DC)
-  * cubical marching squares (CMS)
-  * https://www.reddit.com/r/VoxelGameDev/comments/qxi0u3/cubical_marching_squares_vs_dual_contouring_which/?rdt=44360
-* mesh from isosurface
-  * `stl -> exo`
-* perfect tets to 4-hexs, what are the element qualities
-
-## 2024-12-18
-
-* .svg figures instead of .png
-  * inherit colors from webpage
-  * more lightweight
-  * `custom.css` into mdbook, and point to it in the mdbook
-  * CBH to create a cross-correlation .svg images
-* order to get a dualized mesh - MRB to do next
-  * dualization with one material block
-  * then dualization with two material blocks, how to handle the interfaces?
[…]
