### 2025-07-31 #8467 @finsberg
Hi @YohannDudouit, @jeremiah-corrado, @cpmech, @iulian787 👋

Considering your expertise, would any of you be willing to review this submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://github.com/openjournals/joss-papers/blob/joss.08467/joss.08467/10.21105.joss.08467.pdf
Software: https://github.com/autotwin/automesh
