### 2025-08-06 6eeab6a
more idomatic type conversions
