### 2025-08-06 29e6fda
more idomatic type conversions
