+++ src/tessellation/mod.rs
@@ -1,9 +1,12 @@
 use super::{
-    Coordinate,
-    fem::{FiniteElementMethods, NODE_NUMBERING_OFFSET, TriangularFiniteElements},
+    Vector,
+    fem::{
+        FiniteElementMethods, HexahedralFiniteElements, NODE_NUMBERING_OFFSET, TRI,
+        TetrahedralFiniteElements, TriangularFiniteElements,
+    },
 };
-use conspire::math::TensorArray;
-use std::fmt;
+use conspire::math::{Tensor, TensorArray};
+use std::fmt::{self, Display, Formatter};
 use std::fs::File;
 use std::io::{BufWriter, Error};
 use stl_io::{IndexedMesh, Triangle, Vertex, read_stl, write_stl};
@@ -12,6 +15,63 @@ use stl_io::{IndexedMesh, Triangle, Vertex, read_stl, write_stl};
 #[derive(Debug, PartialEq)]
 pub struct Tessellation {
     data: IndexedMesh,
+}
+
+impl From<HexahedralFiniteElements> for Tessellation {
+    fn from(_: HexahedralFiniteElements) -> Self {
+        unimplemented!()
+    }
+}
+
+impl From<TetrahedralFiniteElements> for Tessellation {
+    fn from(_: TetrahedralFiniteElements) -> Self {
+        unimplemented!()
+    }
+}
+
+impl From<TriangularFiniteElements> for Tessellation {
+    fn from(finite_elements: TriangularFiniteElements) -> Self {
+        let mut normal = Vector::zero();
+        let mut vertices_tri = [0; TRI];
+        let nodal_coordinates = finite_elements.get_nodal_coordinates();
+        let vertices = nodal_coordinates
+            .iter()
+            .map(|coordinate| {
+                stl_io::Vertex::new([
+                    coordinate[0] as f32,
+                    coordinate[1] as f32,
+                    coordinate[2] as f32,
+                ])
+            })
+            .collect();
+        let faces = finite_elements
+            .get_element_node_connectivity()
+            .iter()
+            .map(|&connectivity| {
+                vertices_tri = [
+                    connectivity[0] - NODE_NUMBERING_OFFSET,
+                    connectivity[1] - NODE_NUMBERING_OFFSET,
+                    connectivity[2] - NODE_NUMBERING_OFFSET,
+                ];
+                normal = (&nodal_coordinates[vertices_tri[1]]
+                    - &nodal_coordinates[vertices_tri[0]])
+                    .cross(
+                        &(&nodal_coordinates[vertices_tri[2]]
+                            - &nodal_coordinates[vertices_tri[0]]),
+                    )
+                    .normalized();
+                stl_io::IndexedTriangle {
+                    normal: stl_io::Normal::new([
+                        normal[0] as f32,
+                        normal[1] as f32,
+                        normal[2] as f32,
+                    ]),
+                    vertices: vertices_tri,
+                }
+            })
+            .collect();
+        Tessellation::new(stl_io::IndexedMesh { vertices, faces })
+    }
 }
 
 impl Tessellation {
@@ -24,32 +84,6 @@ impl Tessellation {
         let mut file = File::open(file_path)?;
         let data = read_stl(&mut file)?;
         Ok(Self { data })
-    }
-    /// Converts the tessellation into finite elements, consuming the tessellation.
-    pub fn into_finite_elements(self) -> TriangularFiniteElements {
-        let data = self.get_data();
-        let element_blocks = vec![1; data.faces.len()];
-        let nodal_coordinates = data
-            .vertices
-            .iter()
-            .map(|&vertex| Coordinate::new([vertex[0].into(), vertex[1].into(), vertex[2].into()]))
-            .collect();
[…]
