### 2025-08-06 5a6ac3b
more idomatic type conversions
