+++ src/fem/tri/mod.rs
@@ -5,10 +5,14 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-    Coordinate, FiniteElementMethods, FiniteElementSpecifics, FiniteElements, Metrics,
-    NODE_NUMBERING_OFFSET, Tessellation, Vector,
+    super::tree::Edges, Coordinate, Coordinates, ELEMENT_NUMBERING_OFFSET, FiniteElementMethods,
+    FiniteElementSpecifics, FiniteElements, Metrics, NODE_NUMBERING_OFFSET, Tessellation,
+    VecConnectivity, Vector,
 };
-use conspire::math::{Tensor, TensorArray};
+use conspire::{
+    math::{Tensor, TensorArray, TensorVec},
+    mechanics::Scalar,
+};
 use ndarray::{Array2, s};
 use ndarray_npy::WriteNpyExt;
 use std::{
@@ -17,10 +21,16 @@ use std::{
     path::Path,
 };
 
-const J_EQUILATERAL: f64 = 0.8660254037844387;
+const FOUR_THIRDS: Scalar = 4.0 / 3.0;
+const FOUR_FIFTHS: Scalar = 4.0 / 5.0;
+const J_EQUILATERAL: Scalar = 0.8660254037844387;
+const REGULAR_DEGREE: i8 = 6;
 
 /// The number of nodes in a triangular finite element.
 pub const TRI: usize = 3;
+
+type Connectivity = super::Connectivity<TRI>;
+type Lengths = conspire::math::Vector;
 
 /// The triangular finite elements type.
 pub type TriangularFiniteElements = FiniteElements<TRI>;
@@ -224,4 +234,345 @@ impl TriangularFiniteElements {
             })
             .collect()
     }
+    /// Isotropic remeshing of the finite elements.
+    pub fn remesh(&mut self, iterations: usize, smoothing_iterations: usize) {
+        remesh(self, iterations, smoothing_iterations)
+    }
+}
+
+fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_iterations: usize) {
+    (0..iterations).for_each(|_| {
+        let mut edges: Edges = fem
+            .get_element_node_connectivity()
+            .iter()
+            .flat_map(|&[node_0, node_1, node_2]| {
+                [[node_0, node_1], [node_1, node_2], [node_2, node_0]].into_iter()
+            })
+            .collect();
+        edges.iter_mut().for_each(|edge| edge.sort());
+        edges.sort();
+        edges.dedup();
+        let mut lengths: Lengths = edges
+            .iter()
+            .map(|&[node_a, node_b]| {
+                (&fem.get_nodal_coordinates()[node_a - NODE_NUMBERING_OFFSET]
+                    - &fem.get_nodal_coordinates()[node_b - NODE_NUMBERING_OFFSET])
+                    .norm()
+            })
+            .collect();
+        let average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
+        // edges.iter_mut().zip(lengths.iter_mut()).for_each(|([node_a, node_b], length)|
+        //     if *length > FOUR_THIRDS * average_length {
+        //         println!("large {:?}", length)
+        //     } else if *length < FOUR_FIFTHS * average_length {
+        //         println!("small {:?}", length)
+        //     }
+        // );
+        // split_edges(fem, &edges, &lengths, average_length);
+        collapse_edges(fem, &mut edges, lengths, average_length);
+        // flip_edges(fem, edges);
+        // Vertex averaging is Laplace smoothing but while removing the normal component.
+    });
+    fem.nodal_influencers = vec![];
+}
+
+fn split_edges(
[…]
