### 2025-08-07 a9f5ef4
slightly better/cleaner
