+++ src/fem/tri/mod.rs
@@ -252,7 +252,7 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_itera
         edges.iter_mut().for_each(|edge| edge.sort());
         edges.sort();
         edges.dedup();
-        let mut lengths: Lengths = edges
+        let lengths: Lengths = edges
             .iter()
             .map(|&[node_a, node_b]| {
                 (&fem.get_nodal_coordinates()[node_a - NODE_NUMBERING_OFFSET]
@@ -388,94 +388,217 @@ fn collapse_edges(
 ) {
     let mut element_index_1 = 0;
     let mut element_index_2 = 0;
-    let mut node_c = 0;
-    let mut node_d = 0;
     let element_node_connectivity = &mut fem.element_node_connectivity;
     let node_element_connectivity = &mut fem.node_element_connectivity;
     let node_node_connectivity = &mut fem.node_node_connectivity;
-    let mut nodal_coordinates = &mut fem.nodal_coordinates;
+    let nodal_coordinates = &mut fem.nodal_coordinates;
     let mut coincident_nodes = vec![None; nodal_coordinates.len()];
     let mut degenerate_elements = vec![false; element_node_connectivity.len()];
     edges
         .iter()
-        .zip(lengths.iter())
-        .filter(|&(_, &length)| length < FOUR_FIFTHS * average_length)
-        .for_each(|(&[node_a, node_b], _)| {
-            [element_index_1, element_index_2, _, _] = edge_info(
-                node_a,
-                node_b,
-                element_node_connectivity,
-                node_element_connectivity,
-            );
-            nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] =
-                (nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone()
-                    + &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET])
-                    / 2.0;
-            nodal_coordinates[node_b - NODE_NUMBERING_OFFSET] =
-                nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone();
-            coincident_nodes[node_b - NODE_NUMBERING_OFFSET] = Some(node_a);
-            degenerate_elements[element_index_1] = true;
-            degenerate_elements[element_index_2] = true;
+        // .zip(lengths.iter())
+        // .filter(|&(_, &length)| length < FOUR_FIFTHS * average_length)
+        // .filter(|&[node_a, node_b]|
+        //     (&nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] - &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET]).norm() < FOUR_FIFTHS * average_length
+        // )
+        // .for_each(|(&[node_a, node_b], _)| {
+        .for_each(|&[node_a, node_b]| {
+            if (&nodal_coordinates[node_a - NODE_NUMBERING_OFFSET]
+                - &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET])
+                .norm()
+                < FOUR_FIFTHS * average_length
+            {
+                [element_index_1, element_index_2, _, _] = edge_info(
+                    node_a,
+                    node_b,
+                    element_node_connectivity,
+                    node_element_connectivity,
+                );
+                // nodal_coordinates[node_b - NODE_NUMBERING_OFFSET] =
+                //     nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone();
+                // coincident_nodes[node_b - NODE_NUMBERING_OFFSET] =
+                //     coincident_nodes[node_a - NODE_NUMBERING_OFFSET].or(Some(node_a));
 
-            // element_node_connectivity
-            //     .iter_mut()
[…]
