### 2025-08-08 #8768 @editorialbot
[REVIEW]: automesh: Automatic mesh generation in Rust
**Submitting author:** @hovey (<a href="http://orcid.org/0000-0003-0915-5765">Chad Hovey</a>)
**Repository:** https://github.com/autotwin/automesh
**Branch with paper.md** (empty if default branch): joss_2025
**Version:** v0.3.5
