### 2025-08-08 #8768 @hovey
wonderful!   thank you!🙏🏼 On Aug 7, 2025, at 22:47, Henrik Finsberg ***@***.***> wrote:﻿finsberg left a comment (openjournals/joss-reviews#8768)
@hovey, this is where the review takes place. Please keep an eye out for comments here from the reviewers, as well as any issues opened by them on your software repository. I recommend you aim to respond to these as soon as possible, and you can address them straight away as they come in if you like, to ensure we do not loose track of the reviewers.
