+++ src/tree/mod.rs
@@ -556,260 +556,260 @@ impl Octree {
             while index < self.len() {
                 if !self[index].is_voxel() && self[index].is_leaf() {
                     'faces: for (face, face_cell) in self[index].get_faces().iter().enumerate() {
-                        if let Some(neighbor) = face_cell
-                            && let Some(kids) = self[*neighbor].get_cells()
-                        {
-                            if strong {
-                                edges = from_fn(|_| false);
-                                vertices = from_fn(|_| false);
-                            }
-                            if match face {
-                                0 => {
-                                    if strong {
-                                        if let Some(edge_cell) = self[kids[3]].get_faces()[1] {
-                                            edges[0] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[7]].get_faces()[1] {
-                                            edges[1] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[6]].get_faces()[5] {
-                                            if let Some(vertex_cell) =
-                                                self[edge_cell].get_faces()[3]
-                                            {
-                                                vertices[0] = self[vertex_cell].is_not_leaf()
-                                            }
-                                            edges[2] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[7]].get_faces()[5] {
-                                            edges[3] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[2]].get_faces()[3] {
-                                            edges[4] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[6]].get_faces()[3] {
-                                            edges[5] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[2]].get_faces()[4] {
-                                            if let Some(vertex_cell) =
-                                                self[edge_cell].get_faces()[3]
-                                            {
-                                                vertices[1] = self[vertex_cell].is_not_leaf()
-                                            }
-                                            edges[6] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[3]].get_faces()[4] {
-                                            edges[7] = self[edge_cell].is_not_leaf()
-                                        }
-                                    }
-                                    self[kids[2]].is_not_leaf()
-                                        || self[kids[3]].is_not_leaf()
-                                        || self[kids[6]].is_not_leaf()
-                                        || self[kids[7]].is_not_leaf()
-                                        || edges.into_iter().any(|edge: bool| edge)
-                                        || vertices.into_iter().any(|vertex| vertex)
+                        if let Some(neighbor) = face_cell {
+                            if let Some(kids) = self[*neighbor].get_cells() {
+                                if strong {
+                                    edges = from_fn(|_| false);
+                                    vertices = from_fn(|_| false);
                                 }
-                                1 => {
-                                    if strong {
-                                        if let Some(edge_cell) = self[kids[2]].get_faces()[2] {
-                                            edges[0] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[6]].get_faces()[2] {
-                                            edges[1] = self[edge_cell].is_not_leaf()
-                                        }
-                                        if let Some(edge_cell) = self[kids[4]].get_faces()[5] {
-                                            if let Some(vertex_cell) =
-                                                self[edge_cell].get_faces()[0]
-                                            {
-                                                vertices[0] = self[vertex_cell].is_not_leaf()
+                                if match face {
+                                    0 => {
+                                        if strong {
[…]
