+++ src/tree/hex/edge_template_2/mod.rs
@@ -19,146 +19,135 @@ pub fn apply(
                     if let Some(face_cell_index) = face_cell
                         && let Some(face_subsubcells) =
                             tree.cell_subcells_contain_leaves(&tree[*face_cell_index], face_index)
-                        {
-                            template_inner(face_index).into_iter().for_each(
-                                |[
-                                    adjacent_face,
-                                    cell_subcell_a,
-                                    cell_subcell_b,
-                                    adjacent_cell_subcell_a,
-                                    adjacent_cell_subcell_b,
-                                    face_subsubcell_a,
-                                    face_subsubcell_b,
-                                    face_subsubcell_c,
-                                    face_subsubcell_d,
-                                    adjacent_face_subsubcell_a,
-                                    adjacent_face_subsubcell_b,
-                                    adjacent_face_subsubcell_c,
-                                    adjacent_face_subsubcell_d,
-                                ]| {
-                                    if let Some(adjacent_cell) = cell_faces[adjacent_face]
-                                        && let Some((adjacent_cell_subcells, adjacent_cell_faces)) =
-                                            tree.cell_contains_leaves(&tree[adjacent_cell])
-                                            && let Some(adjacent_cell_face_cell) =
-                                                adjacent_cell_faces[face_index]
-                                                && let Some(adjacent_face_subsubcells) = tree
-                                                    .cell_subcells_contain_leaves(
-                                                        &tree[adjacent_cell_face_cell],
-                                                        face_index,
-                                                    )
-                                                {
-                                                    let lngth = *tree
-                                                        [face_subsubcells[face_subsubcell_a]]
-                                                        .get_lngth()
-                                                        as f64;
-                                                    let coordinates_1 = &nodal_coordinates
-                                                        [cells_nodes
-                                                            [face_subsubcells[face_subsubcell_a]]
-                                                            - NODE_NUMBERING_OFFSET]
-                                                        - face_direction(face_index) * lngth;
-                                                    let node_1 = *nodes_map
-                                                        .get(&(
-                                                            (2.0 * coordinates_1[0]) as usize,
-                                                            (2.0 * coordinates_1[1]) as usize,
-                                                            (2.0 * coordinates_1[2]) as usize,
-                                                        ))
-                                                        .expect("nonexistent entry");
-                                                    let coordinates_2 = &nodal_coordinates
-                                                        [cells_nodes
-                                                            [face_subsubcells[face_subsubcell_b]]
-                                                            - NODE_NUMBERING_OFFSET]
-                                                        - face_direction(face_index) * lngth;
-                                                    let node_2 = *nodes_map
-                                                        .get(&(
-                                                            (2.0 * coordinates_2[0]) as usize,
-                                                            (2.0 * coordinates_2[1]) as usize,
-                                                            (2.0 * coordinates_2[2]) as usize,
-                                                        ))
-                                                        .expect("nonexistent entry");
-                                                    let coordinates_3 = &nodal_coordinates
-                                                        [cells_nodes[adjacent_face_subsubcells
-                                                            [adjacent_face_subsubcell_a]]
-                                                            - NODE_NUMBERING_OFFSET]
-                                                        - face_direction(face_index) * lngth;
-                                                    let node_3 = *nodes_map
-                                                        .get(&(
-                                                            (2.0 * coordinates_3[0]) as usize,
-                                                            (2.0 * coordinates_3[1]) as usize,
-                                                            (2.0 * coordinates_3[2]) as usize,
-                                                        ))
-                                                        .expect("nonexistent entry");
-                                                    let coordinates_4 = &nodal_coordinates
-                                                        [cells_nodes[adjacent_face_subsubcells
[…]
