+++ src/tree/hex/edge_template_3/mod.rs
@@ -180,187 +180,121 @@ fn template(
     let subcell_a_faces = tree[cell_subcells[subcell_a]].get_faces();
     if let Some(subcell_a_face_m) = subcell_a_faces[face_m]
         && let Some(subcell_a_face_n) = subcell_a_faces[face_n]
-            && let Some((subcell_a_face_m_subcells, _)) =
-                tree.cell_contains_leaves(&tree[subcell_a_face_m])
-                && let Some((subcell_a_face_n_subcells, _)) =
-                    tree.cell_contains_leaves(&tree[subcell_a_face_n])
-                    && let Some(diagonal_a) =
-                        tree[subcell_a_face_m_subcells[subcell_c]].get_faces()[face_n]
-                        && tree[diagonal_a].is_leaf()
-                            && let Some(subdiagonal_a) =
-                                tree[subcell_a_face_m_subcells[subcell_e]].get_faces()[face_n]
-                                && tree[subdiagonal_a].is_leaf() {
-                                    let subcell_b_faces =
-                                        tree[cell_subcells[subcell_b]].get_faces();
-                                    if let Some(subcell_b_face_m) = subcell_b_faces[face_m]
-                                        && let Some(subcell_b_face_n) = subcell_b_faces[face_n]
-                                            && let Some((subcell_b_face_m_subcells, _)) =
-                                                tree.cell_contains_leaves(&tree[subcell_b_face_m])
-                                                && let Some((subcell_b_face_n_subcells, _)) = tree
-                                                    .cell_contains_leaves(&tree[subcell_b_face_n])
-                                                    && let Some(diagonal_b) = tree
-                                                        [subcell_b_face_m_subcells[subcell_e]]
-                                                        .get_faces()[face_n]
-                                                        && tree[diagonal_b].is_leaf()
-                                                            && let Some(subdiagonal_b) = tree
-                                                                [subcell_b_face_m_subcells
-                                                                    [subcell_c]]
-                                                                .get_faces()[face_n]
-                                                                && tree[subdiagonal_b].is_leaf() {
-                                                                    let lngth = *tree
-                                                                        [subcell_a_face_m_subcells
-                                                                            [subcell_e]]
-                                                                        .get_lngth()
-                                                                        as f64;
-                                                                    nodal_coordinates.push(
-                                                                &nodal_coordinates[cells_nodes
-                                                                    [subcell_a_face_m_subcells
-                                                                        [subcell_e]]
-                                                                    - NODE_NUMBERING_OFFSET]
-                                                                    + &direction * lngth,
-                                                            );
-                                                                    nodal_coordinates.push(
-                                                                &nodal_coordinates[cells_nodes
-                                                                    [subcell_b_face_m_subcells
-                                                                        [subcell_c]]
-                                                                    - NODE_NUMBERING_OFFSET]
-                                                                    + direction * lngth,
-                                                            );
-                                                                    (0..2).for_each(|k| {
-                                                                        assert!(nodes_map.insert(
-                                                            (
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][0])
-                                                                    as usize,
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][1])
-                                                                    as usize,
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][2])
-                                                                    as usize,
-                                                            ),
-                                                            *node_index + k,
-                                                        ).is_none(), "duplicate entry")
-                                                                    });
-                                                                    if flip {
-                                                                        element_node_connectivity.push([
[…]
