+++ src/tree/hex/face_template_1/mod.rs
@@ -25,19 +25,19 @@ pub fn apply(
                     if let Some(face_cell_index) = face_cell
                         && let Some(face_subsubcells) =
                             tree.cell_subcells_contain_leaves(&tree[*face_cell_index], face_index)
-                        {
-                            template(
-                                cell_subcells,
-                                cells_nodes,
-                                nodes_map,
-                                face_index,
-                                face_subsubcells,
-                                tree,
-                                element_node_connectivity,
-                                nodal_coordinates,
-                                node_index,
-                            )
-                        }
+                    {
+                        template(
+                            cell_subcells,
+                            cells_nodes,
+                            nodes_map,
+                            face_index,
+                            face_subsubcells,
+                            tree,
+                            element_node_connectivity,
+                            nodal_coordinates,
+                            node_index,
+                        )
+                    }
                 })
         });
 }

[…]
