### 2025-08-13 eebee98
more idiomatic flatten() is noticeably slower
