### 2025-08-13 be5e93e
already working pretty well without collapse
