+++ src/fem/tri/mod.rs
@@ -6,8 +6,8 @@ use std::time::Instant;
 
 use super::{
     super::tree::Edges, Coordinate, Coordinates, ELEMENT_NUMBERING_OFFSET, FiniteElementMethods,
-    FiniteElementSpecifics, FiniteElements, Metrics, NODE_NUMBERING_OFFSET, Tessellation,
-    VecConnectivity, Vector,
+    FiniteElementSpecifics, FiniteElements, Metrics, NODE_NUMBERING_OFFSET, Smoothing,
+    Tessellation, VecConnectivity, Vector,
 };
 use conspire::{
     math::{Tensor, TensorArray, TensorVec},
@@ -235,12 +235,20 @@ impl TriangularFiniteElements {
             .collect()
     }
     /// Isotropic remeshing of the finite elements.
-    pub fn remesh(&mut self, iterations: usize, smoothing_iterations: usize) {
-        remesh(self, iterations, smoothing_iterations)
+    pub fn remesh(&mut self, iterations: usize, smoothing_method: &Smoothing) {
+        remesh(self, iterations, smoothing_method)
     }
 }
 
-fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_iterations: usize) {
+fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_method: &Smoothing) {
+    let foo = 0.1;
+    let bar = 0.6307;
+    let mut average_length = 0.0;
+    let mut lengths = Lengths::zero(0);
+    fem.boundary_nodes = vec![];
+    fem.exterior_nodes = vec![];
+    fem.interface_nodes = vec![];
+    fem.interior_nodes = vec![];
     (0..iterations).for_each(|_| {
         let mut edges: Edges = fem
             .get_element_node_connectivity()
@@ -252,7 +260,7 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_itera
         edges.iter_mut().for_each(|edge| edge.sort());
         edges.sort();
         edges.dedup();
-        let lengths: Lengths = edges
+        lengths = edges
             .iter()
             .map(|&[node_a, node_b]| {
                 (&fem.get_nodal_coordinates()[node_a - NODE_NUMBERING_OFFSET]
@@ -260,28 +268,23 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_itera
                     .norm()
             })
             .collect();
-        let average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
-        // edges.iter_mut().zip(lengths.iter_mut()).for_each(|([node_a, node_b], length)|
-        //     if *length > FOUR_THIRDS * average_length {
-        //         println!("large {:?}", length)
-        //     } else if *length < FOUR_FIFTHS * average_length {
-        //         println!("small {:?}", length)
-        //     }
-        // );
-        // fem.node_element_connectivity().unwrap();
-        // fem.node_node_connectivity().unwrap();
-        split_edges(fem, &edges, &lengths, average_length);
+        average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
+        split_edges(fem, &mut edges, &mut lengths, average_length);
         // collapse_edges(fem, &mut edges, lengths, average_length);
-        // flip_edges(fem, edges);
-        // Vertex averaging is Laplace smoothing but while removing the normal component.
+        flip_edges(fem, edges);
+        //
+        // if can update edges in flip_edges, won't have to recalculate above
+        // just leave lengths though, have to update every one in smooth, similar to recalculate
+        //
+        fem.nodal_influencers();
[…]
