### 2025-08-13 e2db733
already working pretty well without collapse
