+++ src/main.rs
@@ -10,6 +10,10 @@ use netcdf::Error as ErrorNetCDF;
 use std::{io::Error as ErrorIO, path::Path, time::Instant};
 use vtkio::Error as ErrorVtk;
 
+const TAUBIN_DEFAULT_ITERS: usize = 20;
+const TAUBIN_DEFAULT_BAND: f64 = 0.1;
+const TAUBIN_DEFAULT_SCALE: f64 = 0.6307;
+
 macro_rules! about {
     () => {
         format!(
@@ -265,6 +269,10 @@ enum Commands {
         /// Mesh output file (exo | mesh | stl | vtk)
         #[arg(long, short, value_name = "FILE")]
         output: String,
+
+        /// Number of remeshing iterations
+        #[arg(default_value_t = 5, long, short = 'n', value_name = "NUM")]
+        iterations: usize,
 
         /// Pass to quiet the terminal output
         #[arg(action, long, short)]
@@ -602,7 +610,7 @@ enum MeshSmoothCommands {
         hierarchical: bool,
 
         /// Number of smoothing iterations
-        #[arg(default_value_t = 20, long, short = 'n', value_name = "NUM")]
+        #[arg(default_value_t = TAUBIN_DEFAULT_ITERS, long, short = 'n', value_name = "NUM")]
         iterations: usize,
 
         /// Smoothing method (Laplace | Taubin) [default: Taubin]
@@ -610,11 +618,11 @@ enum MeshSmoothCommands {
         method: Option<String>,
 
         /// Pass-band frequency (for Taubin only)
-        #[arg(default_value_t = 0.1, long, short = 'k', value_name = "FREQ")]
+        #[arg(default_value_t = TAUBIN_DEFAULT_BAND, long, short = 'k', value_name = "FREQ")]
         pass_band: f64,
 
         /// Scaling parameter for all smoothing methods
-        #[arg(default_value_t = 0.6307, long, short, value_name = "SCALE")]
+        #[arg(default_value_t = TAUBIN_DEFAULT_SCALE, long, short, value_name = "SCALE")]
         scale: f64,
     },
 }
@@ -1016,10 +1024,11 @@ fn main() -> Result<(), ErrorWrapper> {
         Some(Commands::Remesh {
             input,
             output,
+            iterations,
             quiet,
         }) => {
             is_quiet = quiet;
-            remesh(input, output, quiet)
+            remesh(input, output, iterations, quiet)
         }
         Some(Commands::Smooth { subcommand }) => match subcommand {
             SmoothSubcommand::Hex(args) => {
@@ -1759,7 +1768,12 @@ fn octree(
     Ok(())
 }
 
-fn remesh(input: String, output: String, quiet: bool) -> Result<(), ErrorWrapper> {
+fn remesh(
+    input: String,
+    output: String,
+    iterations: usize,
+    quiet: bool,
+) -> Result<(), ErrorWrapper> {
     let input_extension = Path::new(&input).extension().and_then(|ext| ext.to_str());
     let output_extension = Path::new(&output).extension().and_then(|ext| ext.to_str());
     let mut finite_elements = match read_input::<TRI, TriangularFiniteElements>(
@@ -1778,10 +1792,23 @@ fn remesh(input: String, output: String, quiet: bool) -> Result<(), ErrorWrapper
         _ => Err(invalid_input(&input, input_extension)),
     }
     .unwrap();
-//
-// fix unwwrap() and use smoothing as a subcommand to recycle as much as possible
-//
-    finite_elements.remesh(5, &Smoothing::Taubin(20, 0.1, 0.607));
+    let time = Instant::now();
+    if !quiet {
[…]
