+++ src/fem/tri/mod.rs
@@ -267,11 +267,11 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
                     .norm()
             });
         average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
-        split_edges(fem, &mut edges, &mut lengths, average_length);
-        // collapse_edges(fem, &mut edges, lengths, average_length);
-        flip_edges(fem, &mut edges);
-        fem.nodal_influencers();
-        fem.smooth(smoothing_method).unwrap();
+        // split_edges(fem, &mut edges, &mut lengths, average_length);
+        collapse_edges(fem, &mut edges, &lengths, average_length);
+        // flip_edges(fem, &mut edges);
+        // fem.nodal_influencers();
+        // fem.smooth(smoothing_method).unwrap();
     });
 }
 
@@ -390,7 +390,7 @@ fn split_edges(
 fn collapse_edges(
     fem: &mut TriangularFiniteElements,
     edges: &mut Edges,
-    lengths: Lengths,
+    lengths: &Lengths,
     average_length: Scalar,
 ) {
     let mut element_index_1 = 0;
@@ -402,212 +402,54 @@ fn collapse_edges(
     let mut coincident_nodes = vec![None; nodal_coordinates.len()];
     let mut degenerate_elements = vec![false; element_node_connectivity.len()];
     edges
-        .iter()
-        // .zip(lengths.iter())
-        // .filter(|&(_, &length)| length < FOUR_FIFTHS * average_length)
-        // .filter(|&[node_a, node_b]|
-        //     (&nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] - &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET]).norm() < FOUR_FIFTHS * average_length
-        // )
-        // .for_each(|(&[node_a, node_b], _)| {
-        .for_each(|&[node_a, node_b]| {
-            if (&nodal_coordinates[node_a - NODE_NUMBERING_OFFSET]
-                - &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET])
-                .norm()
-                < FOUR_FIFTHS * average_length
-            {
-                [element_index_1, element_index_2, _, _] = edge_info(
-                    node_a,
-                    node_b,
-                    element_node_connectivity,
-                    node_element_connectivity,
-                );
-                // nodal_coordinates[node_b - NODE_NUMBERING_OFFSET] =
-                //     nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone();
-                // coincident_nodes[node_b - NODE_NUMBERING_OFFSET] =
-                //     coincident_nodes[node_a - NODE_NUMBERING_OFFSET].or(Some(node_a));
-
-                nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] =
-                    (nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone()
-                        + &nodal_coordinates[node_b - NODE_NUMBERING_OFFSET])
-                        / 2.0;
-                nodal_coordinates[node_b - NODE_NUMBERING_OFFSET] =
-                    nodal_coordinates[node_a - NODE_NUMBERING_OFFSET].clone();
-                coincident_nodes[node_b - NODE_NUMBERING_OFFSET] =
-                    coincident_nodes[node_a - NODE_NUMBERING_OFFSET].or(Some(node_a));
-
-                let foo = coincident_nodes[node_a - NODE_NUMBERING_OFFSET].or(Some(node_a));
-
-                // coincident_nodes
-                //     .iter()
-                //     .filter(|&&coincident_node| {
[…]
