+++ src/main.rs
@@ -1787,24 +1787,28 @@ fn remesh(
         quiet,
         true,
     )? {
-        InputTypes::Abaqus(finite_elements) => Ok(finite_elements),
-        InputTypes::Stl(tessellation) => Ok(tessellation.into()),
-        _ => Err(invalid_input(&input, input_extension)),
-    }
-    .unwrap();
+        InputTypes::Abaqus(finite_elements) => finite_elements,
+        InputTypes::Stl(tessellation) => tessellation.into(),
+        _ => return invalid_input(&input, input_extension),
+    };
     let time = Instant::now();
     if !quiet {
         println!("     \x1b[1;96mRemeshing\x1b[0m {output}");
     }
     finite_elements.node_element_connectivity()?;
     finite_elements.node_node_connectivity()?;
+    // finite_elements.remesh(
+    //     iterations,
+    //     &Smoothing::Taubin(
+    //         TAUBIN_DEFAULT_ITERS,
+    //         TAUBIN_DEFAULT_BAND,
+    //         TAUBIN_DEFAULT_SCALE,
+    //     ),
+    // );
+    let foo = 1;
     finite_elements.remesh(
         iterations,
-        &Smoothing::Taubin(
-            TAUBIN_DEFAULT_ITERS,
-            TAUBIN_DEFAULT_BAND,
-            TAUBIN_DEFAULT_SCALE,
-        ),
+        &Smoothing::Taubin(1, TAUBIN_DEFAULT_BAND, TAUBIN_DEFAULT_SCALE),
     );
     if !quiet {
         println!("        \x1b[1;92mDone\x1b[0m {:?}", time.elapsed());
