### 2025-08-14 2f1451c
Update conspire requirement from =0.5.7 to =0.5.8
Updates the requirements on [conspire](https://github.com/mrbuche/conspire.rs) to permit the latest version.
- [Release notes](https://github.com/mrbuche/conspire.rs/releases)
- [Commits](https://github.com/mrbuche/conspire.rs/compare/v0.5.7...v0.5.8)

---
updated-dependencies:
- dependency-name: conspire
  dependency-version: 0.5.8
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
