+++ src/fem/tri/mod.rs
@@ -257,6 +257,16 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
     fem.exterior_nodes = vec![];
     fem.interface_nodes = vec![];
     fem.interior_nodes = vec![];
+
+    let mut signatures = std::collections::HashSet::new();
+    for (i, triangle) in fem.get_element_node_connectivity().iter().enumerate() {
+        let mut sorted = triangle.clone();
+        sorted.sort();
+        if !signatures.insert(sorted.clone()) {
+            println!("Input mesh has duplicate triangle: {:?} at index {}", sorted, i);
+        }
+    }
+
     (0..iterations).for_each(|_| {
         edges
             .iter()
@@ -267,11 +277,11 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
                     .norm()
             });
         average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
-        split_edges(fem, &mut edges, &mut lengths, average_length);
-        // collapse_edges(fem, &mut edges, &lengths, average_length);
-        flip_edges(fem, &mut edges);
-        fem.nodal_influencers();
-        fem.smooth(smoothing_method).unwrap();
+        // split_edges(fem, &mut edges, &mut lengths, average_length);
+        collapse_edges(fem, &mut edges, &lengths, average_length);
+    //     flip_edges(fem, &mut edges);
+    //     fem.nodal_influencers();
+    //     fem.smooth(smoothing_method).unwrap();
     });
 }
 
@@ -393,63 +403,119 @@ fn collapse_edges(
     lengths: &Lengths,
     average_length: Scalar,
 ) {
-    let mut element_index_1 = 0;
-    let mut element_index_2 = 0;
+    let mut element_index_1;
+    let mut element_index_2;
     let element_node_connectivity = &mut fem.element_node_connectivity;
     let node_element_connectivity = &mut fem.node_element_connectivity;
     let node_node_connectivity = &mut fem.node_node_connectivity;
     let nodal_coordinates = &mut fem.nodal_coordinates;
-    let mut coincident_nodes = vec![None; nodal_coordinates.len()];
-    let mut degenerate_elements = vec![false; element_node_connectivity.len()];
-    edges
-        .iter_mut()
-        .zip(lengths.iter())
-        .filter(|&(_, &length)| length < FOUR_FIFTHS * average_length)
-        .for_each(|([node_a, node_b], _)| {
-            [element_index_1, element_index_2, _, _] = edge_info(
-                *node_a,
-                *node_b,
-                element_node_connectivity,
-                node_element_connectivity,
-            );
-            degenerate_elements[element_index_1] = true;
-            degenerate_elements[element_index_2] = true;
-            nodal_coordinates[*node_a - NODE_NUMBERING_OFFSET] =
-                (nodal_coordinates[*node_a - NODE_NUMBERING_OFFSET].clone()
-                    + &nodal_coordinates[*node_b - NODE_NUMBERING_OFFSET])
-                    / 2.0;
-            nodal_coordinates[*node_b - NODE_NUMBERING_OFFSET] =
-                nodal_coordinates[*node_a - NODE_NUMBERING_OFFSET].clone();
-            let mut asdf = vec![*node_a, *node_b];
-            if let Some(foo) = coincident_nodes[*node_a - NODE_NUMBERING_OFFSET] {
-                asdf.push(foo);
-                coincident_nodes[foo - NODE_NUMBERING_OFFSET] = Some(*node_a);
-                nodal_coordinates[foo - NODE_NUMBERING_OFFSET] =
[…]
