+++ src/fem/tri/mod.rs
@@ -5,7 +5,7 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-    super::tree::Edges, Coordinate, Coordinates, ELEMENT_NUMBERING_OFFSET, FiniteElementMethods,
+    super::tree::Edges, Coordinate, ELEMENT_NUMBERING_OFFSET, FiniteElementMethods,
     FiniteElementSpecifics, FiniteElements, Metrics, NODE_NUMBERING_OFFSET, Smoothing,
     Tessellation, VecConnectivity, Vector,
 };
@@ -22,7 +22,7 @@ use std::{
 };
 
 const FOUR_THIRDS: Scalar = 4.0 / 3.0;
-const FOUR_FIFTHS: Scalar = 4.0 / 5.0;
+// const FOUR_FIFTHS: Scalar = 4.0 / 5.0;
 const J_EQUILATERAL: Scalar = 0.8660254037844387;
 const REGULAR_DEGREE: i8 = 6;
 
@@ -257,16 +257,6 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
     fem.exterior_nodes = vec![];
     fem.interface_nodes = vec![];
     fem.interior_nodes = vec![];
-
-    let mut signatures = std::collections::HashSet::new();
-    for (i, triangle) in fem.get_element_node_connectivity().iter().enumerate() {
-        let mut sorted = triangle.clone();
-        sorted.sort();
-        if !signatures.insert(sorted.clone()) {
-            println!("Input mesh has duplicate triangle: {:?} at index {}", sorted, i);
-        }
-    }
-
     (0..iterations).for_each(|_| {
         edges
             .iter()
@@ -277,11 +267,11 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
                     .norm()
             });
         average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
-        // split_edges(fem, &mut edges, &mut lengths, average_length);
-        collapse_edges(fem, &mut edges, &lengths, average_length);
-    //     flip_edges(fem, &mut edges);
-    //     fem.nodal_influencers();
-    //     fem.smooth(smoothing_method).unwrap();
+        split_edges(fem, &mut edges, &mut lengths, average_length);
+        // collapse_edges(fem, &mut edges, &lengths, average_length);
+        flip_edges(fem, &mut edges);
+        fem.nodal_influencers();
+        fem.smooth(smoothing_method).unwrap();
     });
 }
 
@@ -397,126 +387,126 @@ fn split_edges(
     lengths.append(&mut new_lengths);
 }
 
-fn collapse_edges(
-    fem: &mut TriangularFiniteElements,
-    edges: &mut Edges,
-    lengths: &Lengths,
-    average_length: Scalar,
-) {
-    let mut element_index_1;
-    let mut element_index_2;
-    let element_node_connectivity = &mut fem.element_node_connectivity;
-    let node_element_connectivity = &mut fem.node_element_connectivity;
-    let node_node_connectivity = &mut fem.node_node_connectivity;
-    let nodal_coordinates = &mut fem.nodal_coordinates;
+// fn collapse_edges(
+//     fem: &mut TriangularFiniteElements,
+//     edges: &mut Edges,
+//     lengths: &Lengths,
+//     average_length: Scalar,
+// ) {
+//     let mut element_index_1;
+//     let mut element_index_2;
+//     let element_node_connectivity = &mut fem.element_node_connectivity;
+//     let node_element_connectivity = &mut fem.node_element_connectivity;
+//     let node_node_connectivity = &mut fem.node_node_connectivity;
[…]
