+++ book/cli/smooth.md
@@ -18,3 +18,10 @@ automesh smooth hex --help
 automesh smooth tri --help
 <!-- cmdrun automesh smooth tri --help -->
 ```
+
+## Mesh Smooth Tri Remesh
+
+```sh
+automesh smooth tri remesh --help
+<!-- cmdrun automesh smooth tri remesh --help -->
+```
