### 2025-08-21 ed30a65
Update smooth.md
