### 2025-08-22 1515a45
major progress on CLI refactor
