+++ src/cli/input/mod.rs
@@ -1,0 +1,157 @@
+use super::ErrorWrapper;
+use automesh::{FiniteElementMethods, Nel, Remove, Scale, Tessellation, Translate, Voxels};
+use std::{path::Path, time::Instant};
+
+pub fn invalid_input(file: &str, extension: Option<&str>) -> ErrorWrapper {
+    ErrorWrapper {
+        message: format!(
+            "Invalid extension .{} from input file {}",
+            extension.unwrap_or("UNDEFINED"),
+            file
+        ),
+    }
+}
+
+pub struct Input<'a> {
+    file: &'a str,
+    nel: Nel,
+    remove: Remove,
+    scale: Scale,
+    translate: Translate,
+}
+
+pub enum FiniteElementInputTypes<const N: usize, T>
+where
+    T: FiniteElementMethods<N>,
+{
+    Abaqus(T),
+    Stl(T),
+}
+
+impl<const N: usize, T> TryFrom<&str> for FiniteElementInputTypes<N, T>
+where
+    T: FiniteElementMethods<N> + From<Tessellation>,
+{
+    type Error = ErrorWrapper;
+    fn try_from(file: &str) -> Result<Self, Self::Error> {
+        let extension = Path::new(&file).extension().and_then(|ext| ext.to_str());
+        match extension {
+            Some("inp") => Ok(FiniteElementInputTypes::Abaqus(T::from_inp(file)?)),
+            Some("stl") => Ok(FiniteElementInputTypes::Stl(T::from(
+                Tessellation::try_from(file)?,
+            ))),
+            _ => Err(invalid_input(file, extension)),
+        }
+    }
+}
+
+impl<const N: usize, T> TryFrom<String> for FiniteElementInputTypes<N, T>
+where
+    T: FiniteElementMethods<N> + From<Tessellation>,
+{
+    type Error = ErrorWrapper;
+    fn try_from(file: String) -> Result<Self, Self::Error> {
+        Self::try_from(file.as_str())
+    }
+}
+
+impl<'a> TryFrom<Input<'a>> for Voxels {
+    type Error = ErrorWrapper;
+    fn try_from(
+        Input {
+            file,
+            nel,
+            remove,
+            scale,
+            translate,
+        }: Input,
+    ) -> Result<Self, Self::Error> {
+        let extension = Path::new(&file).extension().and_then(|ext| ext.to_str());
+        match extension {
+            Some("npy") => Ok(Voxels::from_npy(file, remove, scale, translate)?),
+            Some("spn") => Ok(Voxels::from_spn(file, nel, remove, scale, translate)?),
+            _ => Err(invalid_input(file, extension)),
+        }
+    }
+}
+
+pub fn read_finite_elements<const N: usize, T>(
+    file: &str,
+    quiet: bool,
+    title: bool,
+) -> Result<T, ErrorWrapper>
+where
+    T: FiniteElementMethods<N> + From<Tessellation>,
+{
+    let time = Instant::now();
+    if !quiet {
+        if title {
+            println!(
+                "\x1b[1m    {} {}\x1b[0m",
+                env!("CARGO_PKG_NAME"),
+                env!("CARGO_PKG_VERSION")
+            );
+        }
+        print!("     \x1b[1;96mReading\x1b[0m {file}");
+    }
+    let finite_elements = match FiniteElementInputTypes::<N, T>::try_from(file)? {
+        FiniteElementInputTypes::Abaqus(finite_elements) => finite_elements,
+        FiniteElementInputTypes::Stl(finite_elements) => finite_elements,
+    };
+    println!(
+        "\x1b[0m\n        \x1b[1;92mDone\x1b[0m {:?}",
+        time.elapsed()
+    );
+    Ok(finite_elements)
[…]
