### 2025-08-22 9f35f77
major progress on CLI refactor
