### 2025-08-26 #8768 @jeremylt
Update: The paper and documentation look pretty good to me. I need to carve out some time to play around with the package now and try to recreate the images in the paper.
