+++ src/cli/remesh/mod.rs
@@ -1,4 +1,12 @@
+use super::{
+    super::{TAUBIN_DEFAULT_BAND, TAUBIN_DEFAULT_ITERS, TAUBIN_DEFAULT_SCALE},
+    ErrorWrapper,
+    input::read_finite_elements,
+    output::write_finite_elements,
+};
+use automesh::{FiniteElementMethods, Smoothing, TriangularFiniteElements};
 use clap::Subcommand;
+use std::time::Instant;
 
 pub const REMESH_DEFAULT_ITERS: usize = 5;
 
@@ -14,4 +22,32 @@ pub enum MeshRemeshCommands {
         #[arg(action, long, short)]
         quiet: bool,
     },
+}
+
+pub fn remesh(
+    input: String,
+    output: String,
+    iterations: usize,
+    quiet: bool,
+) -> Result<(), ErrorWrapper> {
+    let mut finite_elements =
+        read_finite_elements::<_, TriangularFiniteElements>(&input, quiet, true)?;
+    let time = Instant::now();
+    if !quiet {
+        println!("   \x1b[1;96mRemeshing\x1b[0m isotropically with {iterations} iterations")
+    }
+    finite_elements.node_element_connectivity()?;
+    finite_elements.node_node_connectivity()?;
+    finite_elements.remesh(
+        iterations,
+        &Smoothing::Taubin(
+            TAUBIN_DEFAULT_ITERS,
+            TAUBIN_DEFAULT_BAND,
+            TAUBIN_DEFAULT_SCALE,
+        ),
+    );
+    if !quiet {
+        println!("        \x1b[1;92mDone\x1b[0m {:?}", time.elapsed());
+    }
+    write_finite_elements(output, finite_elements, quiet)
 }
