+++ src/main.rs
@@ -1,7 +1,6 @@
 use automesh::{
-    Extraction, FiniteElementMethods, HEX, HexahedralFiniteElements, Octree, Remove, Scale,
-    Smoothing, TET, TRI, Tessellation, TetrahedralFiniteElements, Translate,
-    TriangularFiniteElements,
+    FiniteElementMethods, HEX, HexahedralFiniteElements, Octree, Remove, Scale, Smoothing, TET,
+    TRI, TetrahedralFiniteElements, Translate, TriangularFiniteElements,
 };
 use clap::{Parser, Subcommand};
 use conspire::math::TensorVec;
@@ -11,14 +10,17 @@ mod cli;
 use cli::{
     ErrorWrapper,
     convert::{ConvertSubcommand, convert_mesh, convert_segmentation},
+    defeature::defeature,
+    diff::diff,
+    extract::extract,
     input::{read_finite_elements, read_segmentation},
-    output::{write_finite_elements, write_segmentation},
+    output::{write_finite_elements, write_metrics},
     remesh::{MeshRemeshCommands, REMESH_DEFAULT_ITERS, remesh},
+    smooth::{
+        MeshSmoothCommands, SmoothSubcommand, TAUBIN_DEFAULT_BAND, TAUBIN_DEFAULT_ITERS,
+        TAUBIN_DEFAULT_SCALE, apply_smoothing_method, smooth,
+    },
 };
-
-const TAUBIN_DEFAULT_ITERS: usize = 20;
-const TAUBIN_DEFAULT_BAND: f64 = 0.1;
-const TAUBIN_DEFAULT_SCALE: f64 = 0.6307;
 
 macro_rules! about {
     () => {
@@ -557,165 +559,6 @@ struct MeshTriArgs {
     adapt: bool,
 }
 
-#[derive(Subcommand, Debug)]
-enum MeshSmoothCommands {
-    /// Applies smoothing to the mesh before output
-    Smooth {
-        #[command(subcommand)]
-        remeshing: Option<MeshRemeshCommands>,
-
-        /// Pass to enable hierarchical control
-        #[arg(action, long, short = 'c')]
-        hierarchical: bool,
-
-        /// Number of smoothing iterations
-        #[arg(default_value_t = TAUBIN_DEFAULT_ITERS, long, short = 'n', value_name = "NUM")]
-        iterations: usize,
-
-        /// Smoothing method (Laplace | Taubin) [default: Taubin]
-        #[arg(long, short, value_name = "NAME")]
-        method: Option<String>,
-
-        /// Pass-band frequency (for Taubin only)
-        #[arg(default_value_t = TAUBIN_DEFAULT_BAND, long, short = 'k', value_name = "FREQ")]
-        pass_band: f64,
-
-        /// Scaling parameter for all smoothing methods
-        #[arg(default_value_t = TAUBIN_DEFAULT_SCALE, long, short, value_name = "SCALE")]
-        scale: f64,
-    },
-}
-
-#[derive(Subcommand)]
-enum SmoothSubcommand {
-    /// Smooths an all-hexahedral mesh
-    Hex(SmoothHexArgs),
-    /// Smooths an all-tetrahedral mesh
-    Tet(SmoothTetArgs),
-    /// Smooths an all-triangular mesh
-    Tri(SmoothTriArgs),
-}
-
-#[derive(clap::Args)]
-struct SmoothHexArgs {
-    /// Pass to enable hierarchical control
-    #[arg(action, long, short = 'c')]
-    hierarchical: bool,
-
-    /// Mesh input file (inp)
-    #[arg(long, short, value_name = "FILE")]
-    input: String,
-
-    /// Smoothed mesh output file (exo | inp | mesh | vtk)
[…]
