+++ src/cli/mod.rs
@@ -1,7 +1,11 @@
 pub mod convert;
+pub mod defeature;
+pub mod diff;
+pub mod extract;
 pub mod input;
 pub mod output;
 pub mod remesh;
+pub mod smooth;
 
 use ndarray_npy::{ReadNpyError, WriteNpyError};
 use netcdf::Error as ErrorNetCDF;
