+++ src/cli/smooth/mod.rs
@@ -1,0 +1,261 @@
+use super::{
+    ErrorWrapper,
+    input::read_finite_elements,
+    output::{write_finite_elements, write_metrics},
+    remesh::MeshRemeshCommands,
+};
+use automesh::{FiniteElementMethods, Smoothing, Tessellation};
+use clap::Subcommand;
+use std::time::Instant;
+
+pub const TAUBIN_DEFAULT_ITERS: usize = 20;
+pub const TAUBIN_DEFAULT_BAND: f64 = 0.1;
+pub const TAUBIN_DEFAULT_SCALE: f64 = 0.6307;
+
+#[derive(Subcommand, Debug)]
+pub enum MeshSmoothCommands {
+    /// Applies smoothing to the mesh before output
+    Smooth {
+        #[command(subcommand)]
+        remeshing: Option<MeshRemeshCommands>,
+
+        /// Pass to enable hierarchical control
+        #[arg(action, long, short = 'c')]
+        hierarchical: bool,
+
+        /// Number of smoothing iterations
+        #[arg(default_value_t = TAUBIN_DEFAULT_ITERS, long, short = 'n', value_name = "NUM")]
+        iterations: usize,
+
+        /// Smoothing method (Laplace | Taubin) [default: Taubin]
+        #[arg(long, short, value_name = "NAME")]
+        method: Option<String>,
+
+        /// Pass-band frequency (for Taubin only)
+        #[arg(default_value_t = TAUBIN_DEFAULT_BAND, long, short = 'k', value_name = "FREQ")]
+        pass_band: f64,
+
+        /// Scaling parameter for all smoothing methods
+        #[arg(default_value_t = TAUBIN_DEFAULT_SCALE, long, short, value_name = "SCALE")]
+        scale: f64,
+    },
+}
+
+#[derive(Subcommand)]
+pub enum SmoothSubcommand {
+    /// Smooths an all-hexahedral mesh
+    Hex(SmoothHexArgs),
+    /// Smooths an all-tetrahedral mesh
+    Tet(SmoothTetArgs),
+    /// Smooths an all-triangular mesh
+    Tri(SmoothTriArgs),
+}
+
+#[derive(clap::Args)]
+pub struct SmoothHexArgs {
+    /// Pass to enable hierarchical control
+    #[arg(action, long, short = 'c')]
+    pub hierarchical: bool,
+
+    /// Mesh input file (inp)
+    #[arg(long, short, value_name = "FILE")]
+    pub input: String,
+
+    /// Smoothed mesh output file (exo | inp | mesh | vtk)
+    #[arg(long, short, value_name = "FILE")]
+    pub output: String,
+
+    /// Number of smoothing iterations
+    #[arg(default_value_t = 20, long, short = 'n', value_name = "NUM")]
+    pub iterations: usize,
+
+    /// Smoothing method (Laplace | Taubin) [default: Taubin]
+    #[arg(long, short, value_name = "NAME")]
+    pub method: Option<String>,
+
+    /// Pass-band frequency (for Taubin only)
+    #[arg(default_value_t = 0.1, long, short = 'k', value_name = "FREQ")]
+    pub pass_band: f64,
+
+    /// Scaling parameter for all smoothing methods
+    #[arg(default_value_t = 0.6307, long, short, value_name = "SCALE")]
+    pub scale: f64,
+
+    /// Quality metrics output file (csv | npy)
+    #[arg(long, value_name = "FILE")]
+    pub metrics: Option<String>,
+
+    /// Pass to quiet the terminal output
+    #[arg(action, long, short)]
+    pub quiet: bool,
+}
+
+#[derive(clap::Args)]
+pub struct SmoothTetArgs {
[…]
