+++ src/main.rs
@@ -1,6 +1,6 @@
 use automesh::{
-    FiniteElementMethods, HEX, HexahedralFiniteElements, Octree, Remove, Scale, Smoothing, TET,
-    TRI, TetrahedralFiniteElements, Translate, TriangularFiniteElements,
+    FiniteElementMethods, HEX, HexahedralFiniteElements, Octree, Remove, Scale, TET, TRI,
+    TetrahedralFiniteElements, Translate, TriangularFiniteElements,
 };
 use clap::{Parser, Subcommand};
 use conspire::math::TensorVec;
@@ -14,12 +14,10 @@ use cli::{
     diff::diff,
     extract::extract,
     input::{read_finite_elements, read_segmentation},
+    mesh::{MeshSubcommand, mesh, mesh_print_info},
     output::{write_finite_elements, write_metrics},
-    remesh::{MeshRemeshCommands, REMESH_DEFAULT_ITERS, remesh},
-    smooth::{
-        MeshSmoothCommands, SmoothSubcommand, TAUBIN_DEFAULT_BAND, TAUBIN_DEFAULT_ITERS,
-        TAUBIN_DEFAULT_SCALE, apply_smoothing_method, smooth,
-    },
+    remesh::{REMESH_DEFAULT_ITERS, remesh},
+    smooth::{SmoothSubcommand, smooth},
 };
 
 macro_rules! about {
@@ -294,271 +292,6 @@ enum Commands {
     },
 }
 
-#[derive(Subcommand)]
-enum MeshSubcommand {
-    /// Creates an all-hexahedral mesh from a segmentation
-    Hex(MeshHexArgs),
-    /// Creates an all-tetrahedral mesh from a segmentation
-    Tet(MeshTetArgs),
-    /// Creates all-triangular isosurface(s) from a segmentation
-    Tri(MeshTriArgs),
-}
-
-#[derive(clap::Args)]
-struct MeshHexArgs {
-    #[command(subcommand)]
-    smoothing: Option<MeshSmoothCommands>,
-
-    /// Segmentation input file (npy | spn)
-    #[arg(long, short, value_name = "FILE")]
-    input: String,
-
-    /// Mesh output file (exo | inp | mesh | vtk)
-    #[arg(long, short, value_name = "FILE")]
-    output: String,
-
-    /// Defeature clusters with less than NUM voxels
-    #[arg(long, short, value_name = "NUM")]
-    defeature: Option<usize>,
-
-    /// Number of voxels in the x-direction (spn)
-    #[arg(long, short = 'x', value_name = "NEL")]
-    nelx: Option<usize>,
-
-    /// Number of voxels in the y-direction (spn)
-    #[arg(long, short = 'y', value_name = "NEL")]
-    nely: Option<usize>,
-
-    /// Number of voxels in the z-direction (spn)
-    #[arg(long, short = 'z', value_name = "NEL")]
-    nelz: Option<usize>,
-
-    /// Voxel IDs to remove from the mesh
-    #[arg(long, num_args = 1.., short, value_delimiter = ' ', value_name = "ID")]
-    remove: Option<Vec<usize>>,
-
-    /// Scaling (> 0.0) in the x-direction, applied before translation
-    #[arg(default_value_t = 1.0, long, value_name = "SCALE")]
-    xscale: f64,
-
-    /// Scaling (> 0.0) in the y-direction, applied before translation
-    #[arg(default_value_t = 1.0, long, value_name = "SCALE")]
-    yscale: f64,
-
-    /// Scaling (> 0.0) in the z-direction, applied before translation
-    #[arg(default_value_t = 1.0, long, value_name = "SCALE")]
-    zscale: f64,
-
-    /// Translation in the x-direction
-    #[arg(
-        long,
[…]
