### 2025-09-02 653f2a7
python3.8 EOL and python3.14 coming online
