+++ .github/workflows/Book.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Python
         uses: actions/setup-python@v5
         with:
-          python-version: 3.13
+          python-version: 3.12
       - name: CLI
         run: cargo install --path .
       - name: Rust
