+++ .github/workflows/Python.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [macos-latest, ubuntu-latest, windows-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
+        python-version: ['3.10', '3.11', '3.12', '3.13', '3.14']
         toolchain: [beta, stable]
     runs-on: ${{ matrix.os }}
     steps:
@@ -54,7 +54,7 @@ jobs:
       - name: Python
         uses: actions/setup-python@v5
         with:
-          python-version: 3.13
+          python-version: 3.12
       - name: Dependencies
         run: pip install build maturin twine
       - name: Source
@@ -72,7 +72,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [macos-latest, ubuntu-latest, windows-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
+        python-version: ['3.10', '3.11', '3.12', '3.13', '3.14']
     steps:
     - name: Checkout
       uses: actions/checkout@v5
@@ -108,7 +108,7 @@ jobs:
       uses: actions/setup-python@v5
       with:
         check-latest: true
-        python-version: 3.13
+        python-version: 3.12
     - name: Dependencies
       run: pip install twine
     - name: Wheels
