+++ .github/workflows/Rust.yml
@@ -24,7 +24,7 @@ jobs:
         uses: actions/setup-python@v5
         with:
           check-latest: true
-          python-version: 3.13
+          python-version: 3.12
       - name: Dependencies
         run: pip install numpy pillow
       - name: Block
