+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
 ]
 dynamic = ["version"]
 name = "automesh"
-requires-python = ">=3.8,<3.14"
+requires-python = ">=3.10,<3.15"
 
 [project.optional-dependencies]
 dev = [
