+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
 ]
 dynamic = ["version"]
 name = "automesh"
-requires-python = ">=3.10,<3.15"
+requires-python = ">=3.10,<3.14"
 
 [project.optional-dependencies]
 dev = [
