### 2025-09-02 2eeae8d
python3.8 EOL and python3.14 coming online
