### 2025-09-03 4f51f56
fix remeshing issue and add printout
