+++ src/cli/remesh/mod.rs
@@ -33,33 +33,7 @@ pub fn remesh(
 ) -> Result<(), ErrorWrapper> {
     let mut finite_elements =
         read_finite_elements::<_, TriangularFiniteElements>(&input, quiet, true)?;
-    let time = Instant::now();
-    if !quiet {
-        println!("   \x1b[1;96mRemeshing\x1b[0m isotropically with {iterations} iterations")
-    }
-    finite_elements.node_element_connectivity()?;
-    finite_elements.node_node_connectivity()?;
-    finite_elements.remesh(
-        iterations,
-        &Smoothing::Taubin(
-            TAUBIN_DEFAULT_ITERS,
-            TAUBIN_DEFAULT_BAND,
-            TAUBIN_DEFAULT_SCALE,
-        ),
-    );
-    if !quiet {
-        let mut blocks = finite_elements.get_element_blocks().clone();
-        let elements = blocks.len();
-        blocks.sort();
-        blocks.dedup();
-        println!(
-            "        \x1b[1;92mDone\x1b[0m {:?} \x1b[2m[{} blocks, {} elements, {} nodes]\x1b[0m",
-            time.elapsed(),
-            blocks.len(),
-            elements,
-            finite_elements.get_nodal_coordinates().len()
-        );
-    }
+    apply_remeshing(&mut finite_elements, iterations, quiet, false)?;
     write_finite_elements(output, finite_elements, quiet)
 }
 
@@ -77,7 +51,7 @@ where
     if !quiet {
         println!("   \x1b[1;96mRemeshing\x1b[0m isotropically with {iterations} iterations")
     }
-    if smoothed {
+    if !smoothed {
         finite_elements.node_element_connectivity()?;
         finite_elements.node_node_connectivity()?;
     }
