### 2025-09-03 a447630
octree cell query speedup
