+++ .github/workflows/Rust.yml
@@ -21,7 +21,7 @@ jobs:
         with:
           toolchain: nightly
       - name: Python
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           check-latest: true
           python-version: 3.13
