### 2025-09-04 16cbdb8
option to mesh right away; cow is 80% there maybe
