### 2025-09-08 421deab
Bump maturin from 1.9.3 to 1.9.4
Bumps [maturin](https://github.com/pyo3/maturin) from 1.9.3 to 1.9.4.
- [Release notes](https://github.com/pyo3/maturin/releases)
- [Changelog](https://github.com/PyO3/maturin/blob/main/Changelog.md)
- [Commits](https://github.com/pyo3/maturin/compare/v1.9.3...v1.9.4)

---
updated-dependencies:
- dependency-name: maturin
  dependency-version: 1.9.4
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
