+++ pyproject.toml
@@ -47,7 +47,7 @@ testpaths = [
 [build-system]
 build-backend = "maturin"
 requires = [
-  "maturin==1.9.3"
+  "maturin==1.9.4"
 ]
 
 [tool.maturin]
