### 2025-09-08 fd1475f
Update chrono requirement from =0.4.41 to =0.4.42
Updates the requirements on [chrono](https://github.com/chronotope/chrono) to permit the latest version.
- [Release notes](https://github.com/chronotope/chrono/releases)
- [Changelog](https://github.com/chronotope/chrono/blob/main/CHANGELOG.md)
- [Commits](https://github.com/chronotope/chrono/compare/v0.4.41...v0.4.42)

---
updated-dependencies:
- dependency-name: chrono
  dependency-version: 0.4.42
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
