+++ Cargo.toml
@@ -16,7 +16,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 clap = { version = "=4.5.47", features = ["derive"] }
-chrono = "=0.4.41"
+chrono = "=0.4.42"
 conspire = { version = "=0.5.8", features = ["fem"] }
 ndarray = { version = "=0.16.1", features = ["rayon"] }
 ndarray-npy = "=0.9.1"
