### 2025-09-08 77b6853
notes
