### 2025-09-09 d46b0a4
closer but stuck again
