+++ src/fem/tri/mod.rs
@@ -22,7 +22,7 @@ use std::{
 };
 
 const FOUR_THIRDS: Scalar = 4.0 / 3.0;
-// const FOUR_FIFTHS: Scalar = 4.0 / 5.0;
+const FOUR_FIFTHS: Scalar = 4.0 / 5.0;
 const J_EQUILATERAL: Scalar = 0.8660254037844387;
 const REGULAR_DEGREE: i8 = 6;
 
@@ -273,11 +273,11 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
                     .norm()
             });
         average_length = lengths.iter().sum::<Scalar>() / (lengths.len() as Scalar);
-        split_edges(fem, &mut edges, &mut lengths, average_length);
-        // collapse_edges(fem, &mut edges, &lengths, average_length);
-        flip_edges(fem, &mut edges);
-        fem.nodal_influencers();
-        fem.smooth(smoothing_method).unwrap();
+        // split_edges(fem, &mut edges, &mut lengths, average_length);
+        collapse_edges(fem, &mut edges, &mut lengths, average_length);
+        // flip_edges(fem, &mut edges);
+        // fem.nodal_influencers();
+        // fem.smooth(smoothing_method).unwrap();
     });
 }
 
@@ -398,7 +398,204 @@ fn split_edges(
     lengths.append(&mut new_lengths);
 }
 
-// fn collapse_edges(
+fn collapse_edges(
+    fem: &mut TriangularFiniteElements,
+    edges: &mut Edges,
+    lengths: &mut Lengths,
+    average_length: Scalar,
+) {
+    let mut element_index_1 = 0;
+    let mut element_index_2 = 0;
+    // let element_blocks = &mut fem.element_blocks;
+    // let element_node_connectivity = &mut fem.element_node_connectivity;
+    // let node_element_connectivity = &mut fem.node_element_connectivity;
+    // let node_node_connectivity = &mut fem.node_node_connectivity;
+    // let nodal_coordinates = &mut fem.nodal_coordinates;
+    // let mut removed_edges = vec![];
+    // let mut removed_elements = vec![];
+    // let mut merged_nodes = vec![None; nodal_coordinates.len()];
+
+    let mut foo = 0;
+
+    let mut node_a;
+    let mut node_b;
+    let mut edge_index = 0;
+    while edge_index < edges.len() {
+        [node_a, node_b] = edges[edge_index];
+        if lengths[edge_index] < FOUR_FIFTHS * average_length {
+            if edge_index < 1826 {
+                // 1826: [299, 737] also some slivers nearby from beforehand
+                println!("{}: [{}, {}]", edge_index, node_a, node_b);
+                [element_index_1, element_index_2, _, _] = edge_info(
+                    node_a,
+                    node_b,
+                    &fem.element_node_connectivity,
+                    &fem.node_element_connectivity,
+                );
+                fem.nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] = (&fem.nodal_coordinates
+                    [node_a - NODE_NUMBERING_OFFSET]
+                    + fem.nodal_coordinates[node_b - NODE_NUMBERING_OFFSET].clone())
+                    * 0.5;
+                fem.nodal_coordinates.remove(node_b - NODE_NUMBERING_OFFSET);
+                if element_index_1 < element_index_2 {
+                    fem.element_blocks.remove(element_index_1);
+                    fem.element_blocks.remove(element_index_2 - 1);
+                    fem.element_node_connectivity.remove(element_index_1);
+                    fem.element_node_connectivity.remove(element_index_2 - 1);
+                } else if element_index_2 < element_index_1 {
+                    fem.element_blocks.remove(element_index_2);
[…]
