+++ src/fem/tri/mod.rs
@@ -30,6 +30,7 @@ const REGULAR_DEGREE: i8 = 6;
 pub const TRI: usize = 3;
 
 const NUM_NODES_FACE: usize = 1;
+const NUM_NODES_TRI: usize = 3;
 
 type Lengths = conspire::math::Vector;
 
@@ -423,8 +424,11 @@ fn collapse_edges(
     while edge_index < edges.len() {
         [node_a, node_b] = edges[edge_index];
         if lengths[edge_index] < FOUR_FIFTHS * average_length {
-            if edge_index < 1826 {
-                // 1826: [299, 737] also some slivers nearby from beforehand
+            if edge_index < 67 {
+                // 1204: [299, 495] creates 3 tris within a tri
+                // 1205: [299, 811] collapses a "tet" into two overlapping tris with all 3 same nodes
+                // 1826: [299, 737] collapses short edge of overlapping tris, which cannot find the unshared c/d nodes
+                // can you put in a case to handle where tris all have the same nodes? (delete them maybe?)
                 println!("{}: [{}, {}]", edge_index, node_a, node_b);
                 [element_index_1, element_index_2, _, _] = edge_info(
                     node_a,
@@ -459,7 +463,7 @@ fn collapse_edges(
                             }
                         })
                     });
-                fem.node_element_connectivity().unwrap();
+                fem.node_element_connectivity().unwrap(); // when do this later have to -1 or -2 the elements indices
                 edges.remove(edge_index);
                 lengths.remove(edge_index);
                 edges.iter_mut().for_each(|edge| {
@@ -505,6 +509,122 @@ fn collapse_edges(
                             - &fem.get_nodal_coordinates()[node_b - NODE_NUMBERING_OFFSET])
                             .norm()
                     });
+                //
+                // if one of the tris connected to node_a is connected to two others and contained by another larger tri, keep the larger tri
+                // - node_a is shared by two tris and the larger tri
+                // - another node in the tri is similar
+                // - a third node is shared by all three smaller tris but not the larger tri
+                // - the edge going from node_a to the center should have been a duplicate edge?
+                //   - this could be key to fast identification but unsure whether always true
+                //
+                // if there are 3 elements that share 1 node amongst all 3 and share 2 nodes among each 3 pairs
+                let elements = &fem.get_node_element_connectivity()[node_a - NODE_NUMBERING_OFFSET];
+                println!("{:?}", elements);
+                let mut nodes: Vec<usize> = elements
+                    .iter()
+                    .flat_map(|&element| {
+                        fem.get_element_node_connectivity()[element - ELEMENT_NUMBERING_OFFSET]
+                    })
+                    .collect();
+                nodes.sort();
+                nodes.dedup();
+                println!("{:?}", nodes);
+                let mut more_elements: Vec<usize> = nodes
+                    .iter()
+                    .flat_map(|&node| {
+                        fem.get_node_element_connectivity()[node - NODE_NUMBERING_OFFSET].clone()
+                    })
+                    .collect();
+                //
+                // is pool of more_elements smaller if use node-to-node connectivity for the first step?
+                //
+                more_elements.sort();
+                more_elements.dedup();
+                println!("{:?}", more_elements);
+                let (center_nodes, triangles): (Vec<usize>, Vec<Vec<usize>>) = nodes
+                    .iter()
+                    .map(|node| {
+                        (node, more_elements
+                            .iter()
+                            .filter(|&&element| {
+                                fem.get_element_node_connectivity()
+                                    [element - ELEMENT_NUMBERING_OFFSET]
+                                    .contains(node)
+                            })
+                            .cloned()
+                            .collect::<Vec<usize>>())
+                    })
[…]
