### 2025-09-11 c43d6f0
welp at least I got through one iteration
