+++ src/fem/tri/mod.rs
@@ -277,8 +277,8 @@ fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_metho
         // split_edges(fem, &mut edges, &mut lengths, average_length);
         collapse_edges(fem, &mut edges, &mut lengths, average_length);
         // flip_edges(fem, &mut edges);
-        // fem.nodal_influencers();
-        // fem.smooth(smoothing_method).unwrap();
+        fem.nodal_influencers();
+        fem.smooth(smoothing_method).unwrap();
     });
 }
 
@@ -416,31 +416,31 @@ fn collapse_edges(
     // let mut removed_elements = vec![];
     // let mut merged_nodes = vec![None; nodal_coordinates.len()];
 
-    let mut foo = 0;
-
     let mut node_a;
     let mut node_b;
     let mut edge_index = 0;
+    let mut foo = 0;
     while edge_index < edges.len() {
+edges.iter().for_each(|[a, b]|
+    if a==b {
+        panic!()
+    }
+);
         [node_a, node_b] = edges[edge_index];
         if lengths[edge_index] < FOUR_FIFTHS * average_length {
-            if edge_index < 67 {
-                // 1204: [299, 495] creates 3 tris within a tri
-                // 1205: [299, 811] collapses a "tet" into two overlapping tris with all 3 same nodes
-                // 1826: [299, 737] collapses short edge of overlapping tris, which cannot find the unshared c/d nodes
-                // can you put in a case to handle where tris all have the same nodes? (delete them maybe?)
-                println!("{}: [{}, {}]", edge_index, node_a, node_b);
+if foo <= 8900000 {
+println!("\nedge {foo} nodes: {:?}", edges[edge_index]);
                 [element_index_1, element_index_2, _, _] = edge_info(
                     node_a,
                     node_b,
                     &fem.element_node_connectivity,
                     &fem.node_element_connectivity,
                 );
+                let foobar = fem.nodal_coordinates.remove(node_b - NODE_NUMBERING_OFFSET);
                 fem.nodal_coordinates[node_a - NODE_NUMBERING_OFFSET] = (&fem.nodal_coordinates
                     [node_a - NODE_NUMBERING_OFFSET]
-                    + fem.nodal_coordinates[node_b - NODE_NUMBERING_OFFSET].clone())
+                    + foobar)
                     * 0.5;
-                fem.nodal_coordinates.remove(node_b - NODE_NUMBERING_OFFSET);
                 if element_index_1 < element_index_2 {
                     fem.element_blocks.remove(element_index_1);
                     fem.element_blocks.remove(element_index_2 - 1);
@@ -476,31 +476,20 @@ fn collapse_edges(
                     });
                     edge.sort()
                 });
-                //
-                // NEED TO REMOVE REDUNDANT EDGES (and lengths) SOMEHOW!
-                // collapsing one edge makes two sets of two on top of one another
-                // merging nodes a and b in edges then makes two duplicate edges
-                // then if one of those gets collapsed, you get another edge with [node_a, node_a] to hit later
-                //
-                let mut dups = vec![];
-                edges
-                    .iter()
-                    .enumerate()
-                    .for_each(|(index, [node_a, node_b])| {
-                        if node_a == node_b {
-                            dups.push(index)
-                        }
-                    });
-                println!("{:?}", dups);
-                let mut foo_index = 0;
-                dups.into_iter().for_each(|index| {
-                    edges.remove(index + foo_index);
-                    lengths.remove(index + foo_index);
-                    foo_index += 1;
+                let mut seen: std::collections::HashSet<[usize; 2]> = std::collections::HashSet::new();
+                let mut index = 0;
+                edges.retain(|edge| {
+                    let keep = seen.insert(edge.clone());
+                    if keep {
[…]
