### 2025-09-15 bac43d2
exterior nodes not working
