### 2025-09-15 ca6acea
exterior nodes not working
