### 2025-09-16 d3fe573
preliminary impl with exo read fixes; may need line search or at least more info on which element inverted
