### 2025-09-16 a371f14
running and finishing but creating poor quality elements and worsening overall distribution a bit possibly from trying to fight the bad elements or something
