### 2025-09-17 533296a
link for .inp files in smoothing, clean up defeature
