### 2025-09-17 37fe91f
energetic smoothing
