### 2025-09-17 b953498
energetic smoothing
