+++ src/fem/hex/test.rs
@@ -149,7 +149,7 @@ fn single() {
         [0.0, 1.0, 1.0],
         [1.0, 1.0, 1.0],
     ]);
-    let node_element_connectivity_gold = vec![vec![1]; 8];
+    let node_element_connectivity_gold = vec![vec![0]; 8];
     let node_node_connectivity_gold = vec![
         vec![2, 3, 5],
         vec![1, 4, 6],
@@ -208,18 +208,18 @@ fn double_x() {
         [2.0, 1.0, 1.0],
     ]);
     let node_element_connectivity_gold = vec![
+        vec![0],
+        vec![0, 1],
         vec![1],
-        vec![1, 2],
-        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1],
-        vec![1, 2],
-        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1],
-        vec![1, 2],
-        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1],
-        vec![1, 2],
-        vec![2],
     ];
     let node_node_connectivity_gold = vec![
         vec![2, 4, 7],
@@ -317,18 +317,18 @@ fn double_y() {
         [1.0, 2.0, 1.0],
     ]);
     let node_element_connectivity_gold = vec![
+        vec![0],
+        vec![0],
+        vec![0, 1],
+        vec![0, 1],
         vec![1],
         vec![1],
-        vec![1, 2],
-        vec![1, 2],
-        vec![2],
-        vec![2],
+        vec![0],
+        vec![0],
+        vec![0, 1],
+        vec![0, 1],
         vec![1],
         vec![1],
-        vec![1, 2],
-        vec![1, 2],
-        vec![2],
-        vec![2],
     ];
     let node_node_connectivity_gold = vec![
         vec![2, 3, 7],
@@ -390,22 +390,22 @@ fn triple() {
         [3.0, 1.0, 1.0],
     ]);
     let node_element_connectivity_gold = vec![
-        vec![1],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
-        vec![2, 3],
-        vec![3],
-        vec![1],
+        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
-        vec![2, 3],
-        vec![3],
-        vec![1],
+        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
-        vec![2, 3],
-        vec![3],
-        vec![1],
+        vec![2],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
-        vec![2, 3],
-        vec![3],
+        vec![2],
     ];
     let node_node_connectivity_gold = vec![
         vec![2, 5, 9],
@@ -476,26 +476,26 @@ fn quadruple() {
         [4.0, 1.0, 1.0],
     ]);
     let node_element_connectivity_gold = vec![
-        vec![1],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
         vec![2, 3],
-        vec![3, 4],
-        vec![4],
-        vec![1],
+        vec![3],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
         vec![2, 3],
-        vec![3, 4],
-        vec![4],
-        vec![1],
+        vec![3],
+        vec![0],
+        vec![0, 1],
         vec![1, 2],
         vec![2, 3],
-        vec![3, 4],
-        vec![4],
-        vec![1],
+        vec![3],
[…]
