+++ src/fem/hex/test.rs
@@ -1383,9 +1383,8 @@ fn bracket() {
         vec![2, 3, 6, 7],
         vec![3, 7],
         vec![4, 8],
-        vec![4, 8, 10],
-        vec![5, 9],
-        vec![6, 9],
+        vec![4, 5, 8, 9],
+        vec![5, 6, 9],
         vec![6, 7],
         vec![7],
         vec![8, 10],
@@ -1405,9 +1404,8 @@ fn bracket() {
         vec![2, 3, 6, 7],
         vec![3, 7],
         vec![4, 8],
-        vec![4, 8, 10],
-        vec![5, 9],
-        vec![6, 9],
+        vec![4, 5, 8, 9],
+        vec![5, 6, 9],
         vec![6, 7],
         vec![7],
         vec![8, 10],
@@ -2489,6 +2487,200 @@ fn sparse() {
         [4.0, 5.0, 5.0],
         [5.0, 5.0, 5.0],
     ]);
+    let node_element_connectivity_gold = vec![
+        vec![0],
+        vec![0],
+        vec![0, 1],
+        vec![0, 1],
+        vec![2],
+        vec![2, 3],
+        vec![3],
+        vec![4],
+        vec![1, 4, 5],
+        vec![1, 5],
+        vec![2, 6],
+        vec![2, 3, 6],
+        vec![3],
+        vec![4],
+        vec![4, 5, 7],
+        vec![5, 7],
+        vec![6, 8],
+        vec![6, 8, 9],
+        vec![9],
+        vec![10],
+        vec![7, 10, 11],
+        vec![7, 11],
+        vec![8],
+        vec![8, 9],
+        vec![9],
+        vec![10],
+        vec![10, 11],
+        vec![11],
+        vec![0],
+        vec![0],
+        vec![12],
+        vec![12, 13],
+        vec![13],
+        vec![14],
+        vec![0, 1, 14, 15],
+        vec![0, 1, 15],
+        vec![2, 12],
+        vec![2, 3, 12, 13],
+        vec![3, 13],
+        vec![4, 14, 16],
+        vec![1, 4, 5, 14, 15, 16],
+        vec![1, 5, 15],
+        vec![2, 6, 17],
+        vec![2, 3, 6, 17, 18],
+        vec![3, 18],
+        vec![4, 16, 19],
+        vec![4, 5, 7, 16, 19],
+        vec![5, 7],
+        vec![6, 8, 17, 20],
+        vec![6, 8, 9, 17, 18, 20],
+        vec![9, 18],
+        vec![10, 19],
+        vec![7, 10, 11, 19],
+        vec![7, 11, 21],
+        vec![8, 20, 21],
+        vec![8, 9, 20],
+        vec![9],
+        vec![10],
+        vec![10, 11],
+        vec![11, 21],
+        vec![21],
+        vec![22],
+        vec![22, 23],
+        vec![12, 23, 24],
+        vec![12, 13, 24],
+        vec![13],
+        vec![14],
+        vec![14, 15, 22],
+        vec![15, 22, 23],
+        vec![12, 23, 24, 25],
+        vec![12, 13, 24, 25],
+        vec![13],
+        vec![14, 16],
+        vec![14, 15, 16],
+        vec![15, 26],
+        vec![17, 25, 26, 27],
+        vec![17, 18, 25, 27],
+        vec![18],
+        vec![16, 19],
+        vec![16, 19],
+        vec![26, 28],
+        vec![17, 20, 26, 27, 28, 29],
+        vec![17, 18, 20, 27, 29],
+        vec![18],
+        vec![19],
+        vec![19, 30],
+        vec![21, 28, 30],
+        vec![20, 21, 28, 29, 31],
+        vec![20, 29, 31, 32],
+        vec![32],
[…]
