+++ src/fem/mod.rs
@@ -38,8 +38,7 @@ use vtkio::{
     },
 };
 
-const ELEMENT_NUMBERING_OFFSET: usize = 0;
-const ELEMENT_NUMBERING_OFFSET_OLD: usize = 1;
+const ELEMENT_NUMBERING_OFFSET: usize = 1;
 pub const NODE_NUMBERING_OFFSET: usize = 1;
 
 /// A vector of finite element block IDs.
@@ -302,7 +301,7 @@ where
                 .for_each(|(node, connected_elements)| {
                     connected_blocks = connected_elements
                         .iter()
-                        .map(|element| element_blocks[element - ELEMENT_NUMBERING_OFFSET])
+                        .map(|&element| element_blocks[element])
                         .collect();
                     connected_blocks.sort();
                     connected_blocks.dedup();
@@ -356,8 +355,7 @@ where
             .enumerate()
             .for_each(|(element, connectivity)| {
                 connectivity.iter().for_each(|node| {
-                    node_element_connectivity[node - NODE_NUMBERING_OFFSET]
-                        .push(element + ELEMENT_NUMBERING_OFFSET)
+                    node_element_connectivity[node - NODE_NUMBERING_OFFSET].push(element)
                 })
             });
         self.node_element_connectivity = node_element_connectivity;
@@ -381,10 +379,8 @@ where
                 .iter_mut()
                 .zip(node_element_connectivity.iter().enumerate())
                 .try_for_each(|(connectivity, (node, node_connectivity))| {
-                    node_connectivity.iter().try_for_each(|element| {
-                        element_connectivity.clone_from(
-                            &element_node_connectivity[element - ELEMENT_NUMBERING_OFFSET],
-                        );
+                    node_connectivity.iter().try_for_each(|&element| {
+                        element_connectivity.clone_from(&element_node_connectivity[element]);
                         if let Some(neighbors) = element_connectivity
                             .iter()
                             .position(|&n| n == node + NODE_NUMBERING_OFFSET)
@@ -1149,7 +1145,7 @@ fn write_element_node_connectivity_to_inp<const N: usize>(
                     file.write_all(
                         format!(
                             "{:>width$}",
-                            element + ELEMENT_NUMBERING_OFFSET_OLD,
+                            element + ELEMENT_NUMBERING_OFFSET,
                             width = element_number_width
                         )
                         .as_bytes(),
