+++ src/tree/mod.rs
@@ -6,8 +6,8 @@ mod hex;
 use super::{
     Coordinate, Coordinates, NSD, Vector,
     fem::{
-        Blocks, FiniteElementMethods, HEX, HexahedralFiniteElements, NODE_NUMBERING_OFFSET,
-        TriangularFiniteElements, hex::HexConnectivity,
+        Blocks, FiniteElementMethods, HEX, HexahedralFiniteElements, TriangularFiniteElements,
+        hex::HexConnectivity,
     },
     voxel::{Nel, Remove, Scale, Translate, VoxelData, Voxels},
 };
@@ -1436,7 +1436,7 @@ impl Octree {
             )
             .for_each(|(cell, (block, connectivity))| {
                 *block = cell.get_block();
-                *connectivity = from_fn(|n| n + index + NODE_NUMBERING_OFFSET);
+                *connectivity = from_fn(|n| n + index);
                 x_min = *cell.get_min_x() as f64 * scale.x() + translate.x();
                 y_min = *cell.get_min_y() as f64 * scale.y() + translate.y();
                 z_min = *cell.get_min_z() as f64 * scale.z() + translate.z();
@@ -1806,12 +1806,9 @@ impl From<Octree> for TriangularFiniteElements {
             })
         });
         non_manifold_edges.iter().for_each(|edge| {
-            let non_manifold_faces: Vec<usize> = node_face_connectivity
-                [edge[0] - NODE_NUMBERING_OFFSET]
+            let non_manifold_faces: Vec<usize> = node_face_connectivity[edge[0]]
                 .iter()
-                .filter(|face_a| {
-                    node_face_connectivity[edge[1] - NODE_NUMBERING_OFFSET].contains(face_a)
-                })
+                .filter(|face_a| node_face_connectivity[edge[1]].contains(face_a))
                 .copied()
                 .collect();
             let mut non_manifold_cells_vec: Vec<usize> = non_manifold_faces
@@ -1886,14 +1883,14 @@ impl From<Octree> for TriangularFiniteElements {
                         .contains(node)
                 })
                 .unwrap();
-            let faces: [usize; 3] = node_face_connectivity[node - NODE_NUMBERING_OFFSET]
+            let faces: [usize; 3] = node_face_connectivity[*node]
                 .iter()
                 .filter(|face| cell_faces[non_manifold_cells[cell_index]].contains(face))
                 .copied()
                 .collect::<Vec<usize>>()
                 .try_into()
                 .expect("Should be 3 faces.");
-            nodal_coordinates.push(nodal_coordinates[node - NODE_NUMBERING_OFFSET].clone());
+            nodal_coordinates.push(nodal_coordinates[*node].clone());
             let node_new = nodal_coordinates.len();
             let mut position = 0;
             faces.iter().for_each(|&face| {
@@ -1941,7 +1938,7 @@ impl From<Octree> for TriangularFiniteElements {
                             })
                         }) {
                             nodal_coordinates.push(nodal_coordinates[node_index].clone());
-                            let node = node_index + NODE_NUMBERING_OFFSET;
+                            let node = node_index;
                             let node_new = nodal_coordinates.len();
                             let mut position = 0;
                             faces[0].iter().for_each(|&face| {
@@ -1993,7 +1990,7 @@ fn invert_connectivity(faces_connectivity: &[[usize; 4]], num_nodes: usize) -> V
         .for_each(|(face, connectivity)| {
             connectivity
                 .iter()
-                .for_each(|node| node_face_connectivity[node - NODE_NUMBERING_OFFSET].push(face))
+                .for_each(|&node| node_face_connectivity[node].push(face))
         });
[…]
