+++ src/fem/hex/test.rs
@@ -138,7 +138,7 @@ fn test_finite_elements(
 #[test]
 fn single() {
     let element_blocks = vec![11];
-    let element_node_connectivity = vec![[1, 2, 4, 3, 5, 6, 8, 7]];
+    let element_node_connectivity = vec![[0, 1, 3, 2, 4, 5, 7, 6]];
     let nodal_coordinates = Coordinates::new(&[
         [0.0, 0.0, 0.0],
         [1.0, 0.0, 0.0],
@@ -151,16 +151,16 @@ fn single() {
     ]);
     let node_element_connectivity_gold = vec![vec![0]; 8];
     let node_node_connectivity_gold = vec![
-        vec![2, 3, 5],
-        vec![1, 4, 6],
+        vec![1, 2, 4],
+        vec![0, 3, 5],
+        vec![0, 3, 6],
+        vec![1, 2, 7],
+        vec![0, 5, 6],
         vec![1, 4, 7],
-        vec![2, 3, 8],
-        vec![1, 6, 7],
-        vec![2, 5, 8],
-        vec![3, 5, 8],
-        vec![4, 6, 7],
+        vec![2, 4, 7],
+        vec![3, 5, 6],
     ];
-    let exterior_nodes_gold = (1..=8).collect();
+    let exterior_nodes_gold = (0..8).collect();
     let interface_nodes_gold = vec![];
     let interior_nodes_gold = vec![];
     let laplacian_gold = Coordinates::new(&[
@@ -192,7 +192,7 @@ fn single() {
 #[test]
 fn double_x() {
     let element_blocks = vec![11; 2];
-    let element_node_connectivity = vec![[1, 2, 5, 4, 7, 8, 11, 10], [2, 3, 6, 5, 8, 9, 12, 11]];
+    let element_node_connectivity = vec![[0, 1, 4, 3, 6, 7, 10, 9], [1, 2, 5, 4, 7, 8, 11, 10]];
     let nodal_coordinates = Coordinates::new(&[
         [0.0, 0.0, 0.0],
         [1.0, 0.0, 0.0],
@@ -222,20 +222,20 @@ fn double_x() {
         vec![1],
     ];
     let node_node_connectivity_gold = vec![
-        vec![2, 4, 7],
-        vec![1, 3, 5, 8],
-        vec![2, 6, 9],
-        vec![1, 5, 10],
-        vec![2, 4, 6, 11],
-        vec![3, 5, 12],
-        vec![1, 8, 10],
-        vec![2, 7, 9, 11],
-        vec![3, 8, 12],
-        vec![4, 7, 11],
-        vec![5, 8, 10, 12],
-        vec![6, 9, 11],
+        vec![1, 3, 6],
+        vec![0, 2, 4, 7],
+        vec![1, 5, 8],
+        vec![0, 4, 9],
+        vec![1, 3, 5, 10],
+        vec![2, 4, 11],
+        vec![0, 7, 9],
+        vec![1, 6, 8, 10],
+        vec![2, 7, 11],
+        vec![3, 6, 10],
+        vec![4, 7, 9, 11],
+        vec![5, 8, 10],
     ];
-    let exterior_nodes_gold = (1..=12).collect();
+    let exterior_nodes_gold = (0..12).collect();
     let interface_nodes_gold = vec![];
     let interior_nodes_gold = vec![];
     let laplacian_gold = Coordinates::new(&[
@@ -301,7 +301,7 @@ fn double_x() {
 #[test]
 fn double_y() {
     let element_blocks = vec![11; 2];
-    let element_node_connectivity = vec![[1, 2, 4, 3, 7, 8, 10, 9], [3, 4, 6, 5, 9, 10, 12, 11]];
[…]

[…]
