+++ Cargo.toml
@@ -15,7 +15,7 @@ version = "0.3.4"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-clap = { version = "=4.5.47", features = ["derive"] }
+clap = { version = "=4.5.48", features = ["derive"] }
 chrono = "=0.4.42"
 conspire = { version = "=0.5.11", features = ["fem"] }
 ndarray = { version = "=0.16.1", features = ["rayon"] }
