### 2025-09-23 d5c9af4
very spooky renumbering of the nodes
