+++ src/fem/hex/test.rs
@@ -619,10 +619,10 @@ fn quadruple_2_voids() {
 fn quadruple_2_blocks() {
     let element_blocks = vec![11, 21, 21, 11];
     let element_node_connectivity = vec![
+        [0, 1, 6, 5, 10, 11, 16, 15],
         [1, 2, 7, 6, 11, 12, 17, 16],
         [2, 3, 8, 7, 12, 13, 18, 17],
         [3, 4, 9, 8, 13, 14, 19, 18],
-        [4, 5, 10, 9, 14, 15, 20, 19],
     ];
     let nodal_coordinates = Coordinates::new(&[
         [0.0, 0.0, 0.0],
@@ -669,29 +669,29 @@ fn quadruple_2_blocks() {
         vec![3],
     ];
     let node_node_connectivity_gold = vec![
-        vec![2, 6, 11],
+        vec![1, 5, 10],
+        vec![0, 2, 6, 11],
         vec![1, 3, 7, 12],
         vec![2, 4, 8, 13],
-        vec![3, 5, 9, 14],
-        vec![4, 10, 15],
-        vec![1, 7, 16],
+        vec![3, 9, 14],
+        vec![0, 6, 15],
+        vec![1, 5, 7, 16],
         vec![2, 6, 8, 17],
         vec![3, 7, 9, 18],
-        vec![4, 8, 10, 19],
-        vec![5, 9, 20],
-        vec![1, 12, 16],
+        vec![4, 8, 19],
+        vec![0, 11, 15],
+        vec![1, 10, 12, 16],
         vec![2, 11, 13, 17],
         vec![3, 12, 14, 18],
-        vec![4, 13, 15, 19],
-        vec![5, 14, 20],
-        vec![6, 11, 17],
+        vec![4, 13, 19],
+        vec![5, 10, 16],
+        vec![6, 11, 15, 17],
         vec![7, 12, 16, 18],
         vec![8, 13, 17, 19],
-        vec![9, 14, 18, 20],
-        vec![10, 15, 19],
+        vec![9, 14, 18],
     ];
-    let exterior_nodes_gold = (1..=20).collect();
-    let interface_nodes_gold = vec![2, 4, 7, 9, 12, 14, 17, 19];
+    let exterior_nodes_gold = (0..20).collect();
+    let interface_nodes_gold = vec![1, 3, 6, 8, 11, 13, 16, 18];
     let interior_nodes_gold = vec![];
     let nodal_influencers_gold = node_node_connectivity_gold.clone();
     test_finite_elements(
@@ -713,9 +713,9 @@ fn quadruple_2_blocks() {
 fn quadruple_2_blocks_void() {
     let element_blocks = vec![11, 21, 11];
     let element_node_connectivity = vec![
+        [0, 1, 6, 5, 10, 11, 16, 15],
         [1, 2, 7, 6, 11, 12, 17, 16],
-        [2, 3, 8, 7, 12, 13, 18, 17],
-        [4, 5, 10, 9, 14, 15, 20, 19],
+        [3, 4, 9, 8, 13, 14, 19, 18],
     ];
     let nodal_coordinates = Coordinates::new(&[
         [0.0, 0.0, 0.0],
@@ -762,29 +762,29 @@ fn quadruple_2_blocks_void() {
         vec![2],
     ];
     let node_node_connectivity_gold = vec![
-        vec![2, 6, 11],
-        vec![1, 3, 7, 12],
-        vec![2, 8, 13],
-        vec![5, 9, 14],
-        vec![4, 10, 15],
-        vec![1, 7, 16],
-        vec![2, 6, 8, 17],
[…]
