### 2025-09-23 9a84b7b
very spooky renumbering of the nodes
