### 2025-09-30 #8768 @hovey
Dear kind reviewers and editor @vijaysm @jeremylt @finsberg 👍

My coauthor @mrbuche and I greatly appreciate all of the time you have invested so far into this review.   Thank you!

Today, the last day of September, I wanted to reach out to see if there any steps I can assist with for the [paper](https://github.com/autotwin/automesh/blob/joss_2025/pubs/joss_2025/paper.md) we submitted the end of April?

* I see the reviews checklist has *installation* and *documentation* items remaining.
* I am happy to do a web meeting (e.g., Teams, Zoom, etc.) to help facilitate.
* **Documentation**
  * Our [mdbook documentation](https://autotwin.github.io/automesh/) documents our feature set in detail, and guides the user through steps to reproduce [Figure 1](https://autotwin.github.io/automesh/analysis/defeature/index.html) and [Figure 2](https://autotwin.github.io/automesh/smoothing/taubin.html) in our JOSS paper submission.
* **Installation**
  * Also in our documentation, there is a dedicated section on [installation](https://autotwin.github.io/automesh/installation.html) that covers macOS, Linux, and Windows platforms.
