+++ src/tree/mod.rs
@@ -16,6 +16,7 @@ use ndarray::{Axis, parallel::prelude::*, s};
 use std::{
     array::from_fn,
     collections::HashMap,
+    iter::repeat,
     ops::{Deref, DerefMut},
 };
 
@@ -1329,18 +1330,26 @@ impl Octree {
     }
     pub fn from_finite_elements<const M: usize, const N: usize, T>(
         finite_elements: T,
+        length: usize,
         levels: usize,
     ) -> Self
     where
         T: FiniteElementMethods<M, N>,
     {
-        let mut blocks = finite_elements.get_element_blocks().clone();
-        let mut samples = finite_elements.centroids();
-        let extra_samples = true;
-        if extra_samples {
-            blocks.append(&mut blocks.iter().flat_map(|&block| [block; N]).collect());
-            samples.append(&mut finite_elements.integration_points());
-        }
+        let mut blocks: Blocks = finite_elements.get_element_blocks().iter().flat_map(|&block| repeat(block).take(length.pow(3))).collect();
+        let mut samples = finite_elements.interior_points(length);
+println!("blocks: {}, samples: {}", blocks.len(), samples.len());
+        let mut exterior_face_samples = finite_elements.exterior_faces_interior_points(length);
+println!("faces: {}, samples: {}", finite_elements.exterior_faces().len(), exterior_face_samples.len());
+        blocks.extend(vec![PADDING; exterior_face_samples.len()]);
+        samples.append(&mut exterior_face_samples); // can you just append right away at the start?
+        // let mut blocks = finite_elements.get_element_blocks().clone();
+        // let mut samples = finite_elements.centroids();
+        // let extra_samples = true;
+        // if extra_samples {
+        //     blocks.append(&mut blocks.iter().flat_map(|&block| [block; N]).collect());
+        //     samples.append(&mut finite_elements.integration_points());
+        // }
         let (minimum, mut maximum) = samples.iter().fold(
             (
                 Coordinate::new([f64::INFINITY; NSD]),
@@ -1365,17 +1374,17 @@ impl Octree {
             *sample -= &minimum;
             *sample *= &scale;
         });
-        let mut exterior_face_centroids = finite_elements.exterior_faces_centroids();
-        if extra_samples {
-            exterior_face_centroids
-                .append(&mut finite_elements.exterior_faces_integration_points());
-        }
-        exterior_face_centroids.iter_mut().for_each(|sample| {
-            *sample -= &minimum;
-            *sample *= &scale;
-        });
-        blocks.extend(vec![PADDING; exterior_face_centroids.len()]);
-        samples.append(&mut exterior_face_centroids);
+        // let mut exterior_face_centroids = finite_elements.exterior_faces_centroids();
+        // if extra_samples {
+        //     exterior_face_centroids
+        //         .append(&mut finite_elements.exterior_faces_integration_points());
+        // }
+        // exterior_face_centroids.iter_mut().for_each(|sample| {
+        //     *sample -= &minimum;
+        //     *sample *= &scale;
+        // });
+        // blocks.extend(vec![PADDING; exterior_face_centroids.len()]);
+        // samples.append(&mut exterior_face_centroids); // can you just append right away at the start?
         let mut tree = Octree {
             nel: Nel::from([nel as usize; NSD]),
             octree: vec![],
