+++ src/fem/tet/mod.rs
@@ -32,11 +32,17 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for TetrahedralFiniteElements {
     fn exterior_faces_integration_points(&self) -> Coordinates {
         self.exterior_faces_centroids()
     }
+    fn exterior_faces_interior_points(&self, _grid_length: usize) -> Coordinates {
+        todo!()
+    }
     fn faces(&self) -> Connectivity<NUM_NODES_FACE> {
         todo!()
     }
     fn integration_points(&self) -> Coordinates {
         self.centroids()
+    }
+    fn interior_points(&self, _grid_length: usize) -> Coordinates {
+        todo!()
     }
     fn maximum_edge_ratios(&self) -> Metrics {
         todo!()
