+++ src/fem/tri/mod.rs
@@ -79,11 +79,17 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for TriangularFiniteElements {
     fn exterior_faces_integration_points(&self) -> Coordinates {
         self.exterior_faces_centroids()
     }
+    fn exterior_faces_interior_points(&self, _grid_length: usize) -> Coordinates {
+        todo!()
+    }
     fn faces(&self) -> Connectivity<NUM_NODES_FACE> {
         unimplemented!()
     }
     fn integration_points(&self) -> Coordinates {
         self.centroids()
+    }
+    fn interior_points(&self, _grid_length: usize) -> Coordinates {
+        todo!()
     }
     fn maximum_edge_ratios(&self) -> Metrics {
         // Knupp 2006

[…]
