### 2025-10-01 bb41e10
highlight the bin path with single quotes for clarity
