### 2025-10-01 bbb54b6
fixed a bunch of stuff
