### 2025-10-01 fa4afc8
fixed a bunch of stuff
