+++ src/voxel/mod.rs
@@ -372,13 +372,13 @@ impl Voxels {
     /// Constructs and returns a segmentation from a finite element mesh.
     pub fn from_finite_elements<const M: usize, const N: usize, T>(
         finite_elements: T,
-        length: usize,
-        levels: usize,
+        grid: usize,
+        size: f64,
     ) -> Self
     where
         T: FiniteElementMethods<M, N>,
     {
-        Octree::from_finite_elements(finite_elements, length, levels).into()
+        Octree::from_finite_elements(finite_elements, grid, size).into()
     }
     /// Constructs and returns a new voxels type from an NPY file.
     pub fn from_npy(
