### 2025-10-03 4d1d5b3
do build info during package release
