### 2025-10-03 faf6137
do build info during package release
