+++ .github/workflows/Rust.yml
@@ -51,6 +51,8 @@ jobs:
         uses: actions/checkout@v5
       - name: Cleanup
         run: rm -fr benches/ book/ sandbox/ tests/
+      - name: Info
+        run: sed -i "s/std::env::var(\"GIT_COMMIT_HASH\").unwrap_or(env!(\"CARGO_PKG_VERSION\").to_string()),/\"$(git rev-parse --short HEAD)\".to_string(),/" src/main.rs
       - name: Package
         run: cargo package
       - name: Login
@@ -58,10 +60,7 @@ jobs:
         run: cargo login ${{ secrets.CRATES_IO_TOKEN }}
       - name: Publish
         if: github.event_name == 'release'
-        run: |
-          commit_hash=$(git rev-parse --short HEAD)
-          sed -i "s/std::env::var(\"GIT_COMMIT_HASH\").unwrap_or(env!(\"CARGO_PKG_VERSION\").to_string()),/\"$commit_hash\".to_string(),/" src/main.rs
-          cargo publish
+        run: cargo publish
   Test:
     if: github.event_name != 'release'
     runs-on: ${{ matrix.os }}
