### 2025-10-03 631c812
do build info during package release
