+++ .github/workflows/Rust.yml
@@ -54,7 +54,7 @@ jobs:
       - name: Info
         run: sed -i "s/std::env::var(\"GIT_COMMIT_HASH\").unwrap_or(env!(\"CARGO_PKG_VERSION\").to_string()),/\"$(git rev-parse --short HEAD)\".to_string(),/" src/main.rs
       - name: Package
-        run: cargo package
+        run: cargo package --allow-dirty
       - name: Login
         if: github.event_name == 'release'
         run: cargo login ${{ secrets.CRATES_IO_TOKEN }}
