+++ .github/workflows/Rust.yml
@@ -12,6 +12,7 @@ on:
 permissions: {}
 jobs:
   Bench:
+    if: github.event_name != 'release'
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -32,6 +33,7 @@ jobs:
       - name: Bench
         run: rustup run nightly cargo bench
   Coverage:
+    if: github.event_name != 'release'
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -50,7 +52,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v5
       - name: Cleanup
-        run: rm -fr benches/ book/ sandbox/ tests/
+        run: rm -fr benches/ book/ pubs/ sandbox/ tests/
       - name: Info
         run: sed -i "s/std::env::var(\"GIT_COMMIT_HASH\").unwrap_or(env!(\"CARGO_PKG_VERSION\").to_string()),/\"$(git rev-parse --short HEAD)\".to_string(),/" src/main.rs
       - name: Package
