+++ src/fem/mod.rs
@@ -39,7 +39,7 @@ use vtkio::{
 };
 
 const ELEMENT_NUMBERING_OFFSET: usize = 1;
-pub const NODE_NUMBERING_OFFSET: usize = 1;
+const NODE_NUMBERING_OFFSET: usize = 1;
 
 /// A vector of finite element block IDs.
 pub type Blocks = Vec<u8>;
@@ -169,7 +169,7 @@ where
             .map(|nodes| {
                 nodes
                     .iter()
-                    .map(|node| coordinates[node - NODE_NUMBERING_OFFSET].clone())
+                    .map(|&node| coordinates[node].clone())
                     .sum::<Coordinate>()
                     / number_of_nodes
             })
@@ -189,7 +189,7 @@ where
             .iter()
             .map(|face| {
                 face.iter()
-                    .map(|node| coordinates[node - NODE_NUMBERING_OFFSET].clone())
+                    .map(|&node| coordinates[node].clone())
                     .sum::<Coordinate>()
                     / number_of_nodes
             })
@@ -246,7 +246,7 @@ where
                 } else {
                     connectivity
                         .iter()
-                        .map(|node_j| nodal_coordinates[node_j - NODE_NUMBERING_OFFSET].clone())
+                        .map(|&node_j| nodal_coordinates[node_j].clone())
                         .sum::<Coordinate>()
                         / (connectivity.len() as f64)
                         - &nodal_coordinates[node_index_i]
@@ -263,16 +263,16 @@ where
             let mut boundary_nodes = self.get_boundary_nodes().clone();
             boundary_nodes
                 .retain(|boundary_node| prescribed_nodes.binary_search(boundary_node).is_err());
-            boundary_nodes.iter().for_each(|boundary_node| {
-                nodal_influencers[boundary_node - NODE_NUMBERING_OFFSET].retain(|node| {
+            boundary_nodes.iter().for_each(|&boundary_node| {
+                nodal_influencers[boundary_node].retain(|node| {
                     boundary_nodes.binary_search(node).is_ok()
                         || prescribed_nodes.binary_search(node).is_ok()
                 })
             });
         }
-        prescribed_nodes.iter().for_each(|prescribed_node| {
-            nodal_influencers[prescribed_node - NODE_NUMBERING_OFFSET].clear()
-        });
+        prescribed_nodes
+            .iter()
+            .for_each(|&prescribed_node| nodal_influencers[prescribed_node].clear());
         self.nodal_influencers = nodal_influencers;
         #[cfg(feature = "profile")]
         println!(
@@ -308,18 +308,18 @@ where
                     number_of_connected_blocks = connected_blocks.len();
                     number_of_connected_elements = connected_elements.len();
                     if number_of_connected_blocks > 1 {
-                        interface_nodes.push(node + NODE_NUMBERING_OFFSET);
+                        interface_nodes.push(node);
                         //
                         // THIS IS WHERE IT IS ASSUMED THAT THE MESH IS PERFECTLY STRUCTURED
                         // ONLY AFFECTS HIERARCHICAL SMOOTHING
                         //
                         if number_of_connected_elements < HEX {
-                            exterior_nodes.push(node + NODE_NUMBERING_OFFSET);
+                            exterior_nodes.push(node);
                         }
                     } else if number_of_connected_elements < HEX {
-                        exterior_nodes.push(node + NODE_NUMBERING_OFFSET);
+                        exterior_nodes.push(node);
                     } else {
-                        interior_nodes.push(node + NODE_NUMBERING_OFFSET);
+                        interior_nodes.push(node);
                     }
                 });
             exterior_nodes.sort();
@@ -354,9 +354,9 @@ where
             .iter()
             .enumerate()
[…]
