+++ src/fem/tri/mod.rs
@@ -5,9 +5,8 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-    super::tree::Edges, Connectivity, Coordinate, Coordinates, FiniteElementMethods,
-    FiniteElementSpecifics, FiniteElements, Metrics, NODE_NUMBERING_OFFSET, Smoothing,
-    Tessellation, VecConnectivity, Vector,
+    super::tree::Edges, Connectivity, Coordinate, FiniteElementMethods, FiniteElementSpecifics,
+    FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity, Vector,
 };
 use conspire::{
     math::{Tensor, TensorArray, TensorVec},
@@ -48,13 +47,7 @@ impl From<Tessellation> for TriangularFiniteElements {
         let element_node_connectivity = data
             .faces
             .iter()
-            .map(|face| {
-                [
-                    face.vertices[0] + NODE_NUMBERING_OFFSET,
-                    face.vertices[1] + NODE_NUMBERING_OFFSET,
-                    face.vertices[2] + NODE_NUMBERING_OFFSET,
-                ]
-            })
+            .map(|face| [face.vertices[0], face.vertices[1], face.vertices[2]])
             .collect();
         TriangularFiniteElements::from_data(
             element_blocks,
@@ -96,14 +89,11 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for TriangularFiniteElements {
         self.get_element_node_connectivity()
             .iter()
             .map(|connectivity| {
-                l0 = (&nodal_coordinates[connectivity[2] - NODE_NUMBERING_OFFSET]
-                    - &nodal_coordinates[connectivity[1] - NODE_NUMBERING_OFFSET])
+                l0 = (&nodal_coordinates[connectivity[2]] - &nodal_coordinates[connectivity[1]])
                     .norm();
-                l1 = (&nodal_coordinates[connectivity[0] - NODE_NUMBERING_OFFSET]
-                    - &nodal_coordinates[connectivity[2] - NODE_NUMBERING_OFFSET])
+                l1 = (&nodal_coordinates[connectivity[0]] - &nodal_coordinates[connectivity[2]])
                     .norm();
-                l2 = (&nodal_coordinates[connectivity[1] - NODE_NUMBERING_OFFSET]
-                    - &nodal_coordinates[connectivity[0] - NODE_NUMBERING_OFFSET])
+                l2 = (&nodal_coordinates[connectivity[1]] - &nodal_coordinates[connectivity[0]])
                     .norm();
                 [l0, l1, l2].into_iter().reduce(f64::max).unwrap()
                     / [l0, l1, l2].into_iter().reduce(f64::min).unwrap()
@@ -213,10 +203,8 @@ impl TriangularFiniteElements {
         self.get_element_node_connectivity()
             .iter()
             .map(|connectivity| {
-                l0 = &nodal_coordinates[connectivity[2] - NODE_NUMBERING_OFFSET]
-                    - &nodal_coordinates[connectivity[1] - NODE_NUMBERING_OFFSET];
-                l1 = &nodal_coordinates[connectivity[0] - NODE_NUMBERING_OFFSET]
-                    - &nodal_coordinates[connectivity[2] - NODE_NUMBERING_OFFSET];
+                l0 = &nodal_coordinates[connectivity[2]] - &nodal_coordinates[connectivity[1]];
+                l1 = &nodal_coordinates[connectivity[0]] - &nodal_coordinates[connectivity[2]];
                 0.5 * (l0.cross(&l1)).norm()
             })
             .collect()
@@ -230,12 +218,9 @@ impl TriangularFiniteElements {
[…]
