+++ src/voxel/mod.rs
@@ -11,7 +11,7 @@ use super::{
     Coordinate, Coordinates, NSD, Octree, Vector,
     fem::{
         Blocks, Connectivity, FiniteElementMethods, HEX, HexahedralFiniteElements,
-        NODE_NUMBERING_OFFSET, TetrahedralFiniteElements, TriangularFiniteElements,
+        TetrahedralFiniteElements, TriangularFiniteElements,
     },
 };
 use conspire::math::TensorArray;
@@ -21,8 +21,6 @@ use std::{
     fs::File,
     io::{BufRead, BufReader, BufWriter, Error, Write},
 };
-
-const NODE_NUMBERING_OFFSET_PLUS_ONE: usize = NODE_NUMBERING_OFFSET + 1;
 
 type InitialNodalCoordinates = Vec<Option<Coordinate>>;
 type VoxelDataFlattened = Blocks;
@@ -624,20 +622,14 @@ fn initial_element_node_connectivity(
         .par_iter()
         .map(|&[i, j, k]| {
             [
-                i + j * nelxplus1 + k * nelxplus1_mul_nelyplus1 + NODE_NUMBERING_OFFSET,
-                i + j * nelxplus1 + k * nelxplus1_mul_nelyplus1 + NODE_NUMBERING_OFFSET_PLUS_ONE,
-                i + (j + 1) * nelxplus1
-                    + k * nelxplus1_mul_nelyplus1
-                    + NODE_NUMBERING_OFFSET_PLUS_ONE,
-                i + (j + 1) * nelxplus1 + k * nelxplus1_mul_nelyplus1 + NODE_NUMBERING_OFFSET,
-                i + j * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1 + NODE_NUMBERING_OFFSET,
-                i + j * nelxplus1
-                    + (k + 1) * nelxplus1_mul_nelyplus1
-                    + NODE_NUMBERING_OFFSET_PLUS_ONE,
-                i + (j + 1) * nelxplus1
-                    + (k + 1) * nelxplus1_mul_nelyplus1
-                    + NODE_NUMBERING_OFFSET_PLUS_ONE,
-                i + (j + 1) * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1 + NODE_NUMBERING_OFFSET,
+                i + j * nelxplus1 + k * nelxplus1_mul_nelyplus1,
+                i + j * nelxplus1 + k * nelxplus1_mul_nelyplus1 + 1,
+                i + (j + 1) * nelxplus1 + k * nelxplus1_mul_nelyplus1 + 1,
+                i + (j + 1) * nelxplus1 + k * nelxplus1_mul_nelyplus1,
+                i + j * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1,
+                i + j * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1 + 1,
+                i + (j + 1) * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1 + 1,
+                i + (j + 1) * nelxplus1 + (k + 1) * nelxplus1_mul_nelyplus1,
             ]
         })
         .collect();
@@ -674,13 +666,12 @@ fn initial_nodal_coordinates(
         .zip(element_node_connectivity.iter())
         .for_each(|(&[x, y, z], connectivity)| {
             offsets.iter().enumerate().for_each(|(node, [cx, cy, cz])| {
-                if nodal_coordinates[connectivity[node] - NODE_NUMBERING_OFFSET].is_none() {
-                    nodal_coordinates[connectivity[node] - NODE_NUMBERING_OFFSET] =
-                        Some(Coordinate::new([
[…]
