+++ src/fem/hex/mod.rs
@@ -5,8 +5,13 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
+<<<<<<< HEAD
     Connectivity, Coordinates, FiniteElementMethods, FiniteElementSpecifics, FiniteElements,
     Metrics, NODE_NUMBERING_OFFSET, Smoothing, Tessellation, Vector,
+=======
+    Connectivity, FiniteElementMethods, FiniteElementSpecifics, FiniteElements, Metrics, Smoothing,
+    Tessellation, Vector,
+>>>>>>> 50e2111 (very spooky renumbering of the nodes)
 };
 use conspire::math::{Tensor, TensorArray, TensorVec};
 use ndarray::{Array2, s};
@@ -229,32 +234,29 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for HexahedralFiniteElements {
                     node_6,
                     node_7,
                 ]| {
-                    l1 = (&nodal_coordinates[node_1 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_0 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_2 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_3 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_5 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_4 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_6 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_7 - NODE_NUMBERING_OFFSET])
+                    l1 = (&nodal_coordinates[node_1] - &nodal_coordinates[node_0]
+                        + &nodal_coordinates[node_2]
+                        - &nodal_coordinates[node_3]
+                        + &nodal_coordinates[node_5]
+                        - &nodal_coordinates[node_4]
+                        + &nodal_coordinates[node_6]
+                        - &nodal_coordinates[node_7])
                         .norm();
-                    l2 = (&nodal_coordinates[node_3 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_0 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_2 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_1 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_7 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_4 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_6 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_5 - NODE_NUMBERING_OFFSET])
+                    l2 = (&nodal_coordinates[node_3] - &nodal_coordinates[node_0]
+                        + &nodal_coordinates[node_2]
+                        - &nodal_coordinates[node_1]
+                        + &nodal_coordinates[node_7]
+                        - &nodal_coordinates[node_4]
+                        + &nodal_coordinates[node_6]
+                        - &nodal_coordinates[node_5])
                         .norm();
-                    l3 = (&nodal_coordinates[node_4 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_0 - NODE_NUMBERING_OFFSET]
-                        + &nodal_coordinates[node_5 - NODE_NUMBERING_OFFSET]
-                        - &nodal_coordinates[node_1 - NODE_NUMBERING_OFFSET]
[…]
