+++ src/fem/hex/mod.rs
@@ -5,13 +5,11 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-<<<<<<< HEAD
     Connectivity, Coordinates, FiniteElementMethods, FiniteElementSpecifics, FiniteElements,
     Metrics, NODE_NUMBERING_OFFSET, Smoothing, Tessellation, Vector,
-=======
     Connectivity, FiniteElementMethods, FiniteElementSpecifics, FiniteElements, Metrics, Smoothing,
     Tessellation, Vector,
->>>>>>> 50e2111 (very spooky renumbering of the nodes)
+    Metrics, Smoothing, Tessellation, Vector,
 };
 use conspire::math::{Tensor, TensorArray, TensorVec};
 use ndarray::{Array2, s};
@@ -117,8 +115,8 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for HexahedralFiniteElements {
                             nodes
                                 .iter()
                                 .zip(shape_functions.iter())
-                                .map(|(node, shape_function)| {
-                                    &nodal_coordinates[node - NODE_NUMBERING_OFFSET]
+                                .map(|(&node, shape_function)| {
+                                    &nodal_coordinates[node]
                                         * shape_function
                                 })
                                 .sum(),
@@ -203,8 +201,8 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for HexahedralFiniteElements {
                                     nodes
                                         .iter()
                                         .zip(shape_functions.iter())
-                                        .map(|(node, shape_function)| {
-                                            &nodal_coordinates[node - NODE_NUMBERING_OFFSET]
+                                        .map(|(&node, shape_function)| {
+                                            &nodal_coordinates[node]
                                                 * shape_function
                                         })
                                         .sum(),
