+++ src/fem/hex/mod.rs
@@ -113,8 +113,7 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for HexahedralFiniteElements {
                                 .iter()
                                 .zip(shape_functions.iter())
                                 .map(|(&node, shape_function)| {
-                                    &nodal_coordinates[node]
-                                        * shape_function
+                                    &nodal_coordinates[node] * shape_function
                                 })
                                 .sum(),
                         );
@@ -199,8 +198,7 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for HexahedralFiniteElements {
                                         .iter()
                                         .zip(shape_functions.iter())
                                         .map(|(&node, shape_function)| {
-                                            &nodal_coordinates[node]
-                                                * shape_function
+                                            &nodal_coordinates[node] * shape_function
                                         })
                                         .sum(),
                                 );
