+++ src/fem/tri/mod.rs
@@ -5,8 +5,9 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-    super::tree::Edges, Connectivity, Coordinate, Coordinates, FiniteElementMethods, FiniteElementSpecifics,
-    FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity, Vector,
+    super::tree::Edges, Connectivity, Coordinate, Coordinates, FiniteElementMethods,
+    FiniteElementSpecifics, FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity,
+    Vector,
 };
 use conspire::{
     math::{Tensor, TensorArray, TensorVec},
