### 2025-10-06 469c01c
Bump maturin from 1.9.4 to 1.9.5
Bumps [maturin](https://github.com/pyo3/maturin) from 1.9.4 to 1.9.5.
- [Release notes](https://github.com/pyo3/maturin/releases)
- [Changelog](https://github.com/PyO3/maturin/blob/main/Changelog.md)
- [Commits](https://github.com/pyo3/maturin/compare/v1.9.4...v1.9.5)

---
updated-dependencies:
- dependency-name: maturin
  dependency-version: 1.9.5
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
