+++ book/installation.md
@@ -24,32 +24,7 @@ For macOS and Linux, use a terminal.  For Windows, use a Command Prompt (CMD) or
 
 ### Rust Prerequisites
 
-Install Rust and Cargo for your operating system:
-
-#### macOS and Linux
-
-1. Open a terminal.  Install Rust using [`rustup`](https://www.rust-lang.org/tools/install):
-
-```sh
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
-```
-
-2. Follow the on-screen instructions to complete the installation.
-3. Add Cargo's bin directory to your `PATH`:
-
-```sh
-source $HOME/.cargo/env
-```
-
-#### Windows
-
-1. Download `rustup-init.exe` from run the [Rust installer](https://www.rust-lang.org/tools/install).
-2. Follow the installation instructions in the command prompt.
-    * Use the default settings (`standard installation`) to add `cargo`, `rustc`, `rustup` to your `PATH`.
-    * The Cargo home directory is, e.g., `C:\Users\<User>\.cargo`, which can be modified with the `CARGO_HOME` environment variable.
-    * You may need to restart your command prompt or system.
-    * Ensure that Cargo's bin directory is in your `PATH`.
-3. Additional Windows installation details are available in [The rustup book](https://rust-lang.github.io/rustup/installation/windows.html).
+It is recommended to install Rust using [Rustup](https://rust-lang.org/learn/get-started/), which is an installer and version management tool.
 
 ### Python Prerequisites
 
