+++ src/fem/tri/mod.rs
@@ -7,6 +7,9 @@ use std::time::Instant;
 use super::{
     super::tree::Edges, Connectivity, Coordinate, FiniteElementMethods, FiniteElementSpecifics,
     FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity, Vector,
+    super::tree::Edges, Connectivity, Coordinate, Coordinates, FiniteElementMethods,
+    FiniteElementSpecifics, FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity,
+    Vector,
 };
 use conspire::{
     math::{Tensor, TensorArray, TensorVec},
