+++ Cargo.toml
@@ -15,7 +15,7 @@ version = "0.3.5"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-clap = { version = "=4.5.48", features = ["derive"] }
+clap = { version = "=4.5.49", features = ["derive"] }
 chrono = "=0.4.42"
 conspire = { version = "=0.5.11", features = ["fem"] }
 ndarray = { version = "=0.16.1", features = ["rayon"] }
