+++ Cargo.toml
@@ -18,7 +18,7 @@ crate-type = ["cdylib", "rlib"]
 clap = { version = "=4.5.48", features = ["derive"] }
 chrono = "=0.4.42"
 conspire = { version = "=0.5.11", features = ["fem"] }
-ndarray = { version = "=0.16.1", features = ["rayon"] }
+ndarray = { version = "=0.17.0", features = ["rayon"] }
 ndarray-npy = "=0.9.1"
 netcdf = { version = "=0.11.0", features = ["ndarray", "static"] }
 pyo3 = { version = "=0.26.0", features = ["extension-module"], optional = true }
