### 2025-10-20 #8768 @jeremylt
Sorry for the slow review - had a family situation come up.

Yeah, the issues are fixed and I just want to do one more read through of the paper, but otherwise this is looking good for me to finish this week.
