### 2025-10-20 f868e95
exclude in manifest means no cleanup needed
