### 2025-10-21 73e0c18
add segment example
