+++ book/SUMMARY.md
@@ -38,6 +38,7 @@
     - [Conforming Mesh Cubit](analysis/sphere_with_shells/conforming.md)
     - [Smoothed Mesh `automesh`](analysis/sphere_with_shells/smoothed.md)
     - [Comparisons](analysis/sphere_with_shells/comparisons.md)
+  - [Conforming v Segmented Meshes](analysis/conforming_v_segmented/README.md)
 
 # Appendices
 
