### 2025-10-22 2fd035a
baseline prior to refactor
