+++ book/analysis/conforming_v_segmented/nested_cubes.py
@@ -1,0 +1,218 @@
+"""Visualize nested segmented cubes with spheres at their centers."""
+
+from typing import Final, List, NamedTuple
+
+# from matplotlib.colors import LightSource
+import matplotlib.pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+from mpl_toolkits.mplot3d.art3d import Poly3DCollection
+import numpy as np
+import numpy.typing as npt
+
+
+class ViewParams(NamedTuple):
+    """Parameters for 3D view setup."""
+
+    elev: float = 63  # elevation angle
+    azim: float = -110  # azimuth angle
+    roll: float = 0  # roll angle
+
+
+DPI: Final[int] = 300  # resolution, dots per inch
+OUTER_SIZE = 10.0  # mm
+SHOW: Final[bool] = True  # Post-processing visuals, show on screen
+SAVE: Final[bool] = True  # Save the .png file
+SPHERE_RADIUS = 0.3
+VIEW: Final[ViewParams] = ViewParams()
+# el, az, roll = 63, -110, 0  # used for most visuals
+# lightsource = LightSource(azdeg=325, altdeg=45)  # azimuth, elevation
+
+
+def draw_cube(
+    ax: Axes3D,
+    origin: List[float],
+    size: float,
+    alpha: float = 0.1,
+    edge_color: str = "b",
+    face_color: str = "cyan",
+) -> npt.NDArray[np.float64]:
+    """
+    Draw a transparent cube with edges using a simplified vertex definition.
+
+    Args:
+        origin: The [x, y, z] coordinate of the cube's minimum corner.
+    """
+
+    x, y, z = origin
+
+    # Define the vertices of the cube
+    vertices = np.array(
+        [
+            [x, y, z],
+            [x + size, y, z],
+            [x + size, y + size, z],
+            [x, y + size, z],  # bottom
+            [x, y, z + size],
+            [x + size, y, z + size],
+            [x + size, y + size, z + size],
+            [x, y + size, z + size],  # top
+        ]
+    )
+
+    # Define the 6 faces of the cube
+    faces = [
+        [vertices[0], vertices[1], vertices[5], vertices[4]],  # front
+        [vertices[2], vertices[3], vertices[7], vertices[6]],  # back
+        [vertices[0], vertices[3], vertices[7], vertices[4]],  # left
+        [vertices[1], vertices[2], vertices[6], vertices[5]],  # right
+        [vertices[0], vertices[1], vertices[2], vertices[3]],  # bottom
+        [vertices[4], vertices[5], vertices[6], vertices[7]],  # top
+    ]
+
+    # Create the 3D polygon collection
+    cube = Poly3DCollection(
+        faces, alpha=alpha, facecolor=face_color, edgecolor=edge_color, linewidth=1.5
+    )
+    ax.add_collection3d(cube)
+
+    return vertices
+
+
+def draw_sphere(
+    ax: Axes3D,
+    center: List[float],
+    radius: float,
+    color: str = "red",
+    alpha: float = 0.8,
+) -> None:
+    """Draw a sphere at the given center."""
+    u = np.linspace(0, 2 * np.pi, 20)
+    v = np.linspace(0, np.pi, 20)
+    x = center[0] + radius * np.outer(np.cos(u), np.sin(v))
+    y = center[1] + radius * np.outer(np.sin(u), np.sin(v))
[…]
