### 2025-10-22 0d32bb0
illustrate the grid and size
