+++ book/analysis/conforming_v_segmented/nested_cubes.py
@@ -1,8 +1,7 @@
 """Visualize nested segmented cubes with spheres at their centers."""
 
-from typing import Final, List, NamedTuple
+from typing import Final, NamedTuple
 
-# from matplotlib.colors import LightSource
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
@@ -10,209 +9,187 @@ import numpy as np
 import numpy.typing as npt
 
 
+# --- Configuration using NamedTuple (Idiomatic and Robust) ---
 class ViewParams(NamedTuple):
-    """Parameters for 3D view setup."""
+    """Container for 3D view angles."""
 
-    elev: float = 63  # elevation angle
-    azim: float = -110  # azimuth angle
-    roll: float = 0  # roll angle
+    elev: float = 63  # elevation
+    azim: float = -110  # azimuth
+    roll: float = 0
 
 
 DPI: Final[int] = 300  # resolution, dots per inch
-OUTER_SIZE = 10.0  # mm
+OUTER_SIZE: Final[float] = 10.0  # mm, Common size for the bounding box
 SHOW: Final[bool] = True  # Post-processing visuals, show on screen
 SAVE: Final[bool] = True  # Save the .png file
-SPHERE_RADIUS = 0.3
 VIEW: Final[ViewParams] = ViewParams()
-# el, az, roll = 63, -110, 0  # used for most visuals
-# lightsource = LightSource(azdeg=325, altdeg=45)  # azimuth, elevation
+
+# -------------------- Helper Functions --------------------
 
 
 def draw_cube(
     ax: Axes3D,
-    origin: List[float],
+    origin: npt.ArrayLike,
     size: float,
     alpha: float = 0.1,
     edge_color: str = "b",
     face_color: str = "cyan",
-) -> npt.NDArray[np.float64]:
+) -> None:
     """
     Draw a transparent cube with edges using a simplified vertex definition.
 
     Args:
         origin: The [x, y, z] coordinate of the cube's minimum corner.
     """
-
     x, y, z = origin
 
-    # Define the vertices of the cube
-    vertices = np.array(
-        [
-            [x, y, z],
-            [x + size, y, z],
-            [x + size, y + size, z],
-            [x, y + size, z],  # bottom
-            [x, y, z + size],
-            [x + size, y, z + size],
-            [x + size, y + size, z + size],
-            [x, y + size, z + size],  # top
-        ]
-    )
+    # Generate all 8 vertices efficiently using broadcasting
+    _coords = [np.array([c, c + size]) for c in origin]
+    X, Y, Z = np.meshgrid(*_coords, indexing="ij")
 
-    # Define the 6 faces of the cube
-    faces = [
-        [vertices[0], vertices[1], vertices[5], vertices[4]],  # front
-        [vertices[2], vertices[3], vertices[7], vertices[6]],  # back
-        [vertices[0], vertices[3], vertices[7], vertices[4]],  # left
-        [vertices[1], vertices[2], vertices[6], vertices[5]],  # right
-        [vertices[0], vertices[1], vertices[2], vertices[3]],  # bottom
-        [vertices[4], vertices[5], vertices[6], vertices[7]],  # top
+    # Vertices (1D array of 8 x 3 points)
+    vertices = np.stack([X.ravel(), Y.ravel(), Z.ravel()], axis=1)
+
+    # Define the indices of the 4 vertices for each of the 6 faces
[…]
