### 2025-10-23 997ed5d
weighted average and scale to [0,1] using pi
