### 2025-10-24 d57e2b4
adaptive no longer hidden
