### 2025-10-24 ce21bce
will never adaptively mesh segmentations, will always adaptively mesh tessellations (whether input or during pipeline of surface reconstruction), so do not need to have this option
