### 2025-10-24 9596e2a
wiring things up
