+++ src/fem/hex/mod.rs
@@ -5,8 +5,8 @@ pub mod test;
 use std::time::Instant;
 
 use super::{
-    Connectivity, Coordinates, FiniteElementMethods, FiniteElementSpecifics, FiniteElements,
-    Metrics, Smoothing, Tessellation, Vector,
+    super::Octree, Connectivity, Coordinates, FiniteElementMethods, FiniteElementSpecifics,
+    FiniteElements, Metrics, Smoothing, Tessellation, Vector,
 };
 use conspire::math::{Tensor, TensorArray, TensorVec};
 use ndarray::{Array2, s};
@@ -455,7 +455,7 @@ impl HexahedralFiniteElements {
 }
 
 impl From<Tessellation> for HexahedralFiniteElements {
-    fn from(_tessellation: Tessellation) -> Self {
-        unimplemented!()
+    fn from(tessellation: Tessellation) -> Self {
+        Octree::from(tessellation).into()
     }
 }
