+++ src/cli/mesh/mod.rs
@@ -1,17 +1,17 @@
 use super::{
     ErrorWrapper,
-    input::read_segmentation,
+    input::{invalid_input, read_segmentation, read_tessellation},
     output::{validate_output, write_finite_elements, write_metrics},
     remesh::{MeshRemeshCommands, apply_remeshing},
     smooth::{MeshSmoothCommands, apply_smoothing_method},
 };
 use automesh::{
     FiniteElementMethods, HEX, HexahedralFiniteElements, Octree, Remove, Scale, TET, TRI,
-    TetrahedralFiniteElements, Translate, TriangularFiniteElements,
+    Tessellation, TetrahedralFiniteElements, Translate, TriangularFiniteElements,
 };
 use clap::Subcommand;
 use conspire::math::TensorVec;
-use std::time::Instant;
+use std::{path::Path, time::Instant};
 
 #[derive(Subcommand)]
 pub enum MeshSubcommand {
@@ -110,7 +110,7 @@ pub enum MeshBasis {
 }
 
 #[allow(clippy::too_many_arguments)]
-pub fn mesh<const N: usize>(
+pub fn mesh<const M: usize, const N: usize, T>(
     smoothing: Option<MeshSmoothCommands>,
     input: String,
     output: String,
@@ -127,17 +127,58 @@ pub fn mesh<const N: usize>(
     ztranslate: f64,
     metrics: Option<String>,
     quiet: bool,
-) -> Result<(), ErrorWrapper> {
-    let mut time = Instant::now();
-    let scale_temporary = Scale::from([xscale, yscale, zscale]);
-    let translate_temporary = Translate::from([xtranslate, ytranslate, ztranslate]);
-    let remove = Remove::from(remove);
+) -> Result<(), ErrorWrapper>
+where
+    T: FiniteElementMethods<M, N> + From<Tessellation>,
+    Tessellation: From<T>,
+{
     let scale = Scale::from([xscale, yscale, zscale]);
     let translate = Translate::from([xtranslate, ytranslate, ztranslate]);
+    let input_extension = Path::new(&input).extension().and_then(|ext| ext.to_str());
+    match input_extension {
+        Some("npy") | Some("spn") => mesh_segmentation::<N>(
+            smoothing, input, output, defeature, nelx, nely, nelz, remove, scale, translate,
+            metrics, quiet,
+        ),
+        Some("stl") => {
+            mesh_tessellation::<M, N, T>(smoothing, input, output, scale, translate, metrics, quiet)
+        }
+        _ => Err(invalid_input(&input, input_extension)),
+    }
+}
+
+#[allow(clippy::too_many_arguments)]
+pub fn mesh_segmentation<const N: usize>(
+    smoothing: Option<MeshSmoothCommands>,
+    input: String,
+    output: String,
+    defeature: Option<usize>,
+    nelx: Option<usize>,
+    nely: Option<usize>,
+    nelz: Option<usize>,
+    remove: Option<Vec<usize>>,
+    scale: Scale,
+    translate: Translate,
+    metrics: Option<String>,
+    quiet: bool,
+) -> Result<(), ErrorWrapper> {
+    let mut time = Instant::now();
+    let remove: Remove = Remove::from(remove);
     let mut input_type = read_segmentation(
-        input, nelx, nely, nelz, remove, scale, translate, quiet, true,
+        input,
+        nelx,
+        nely,
+        nelz,
+        remove,
+        scale.clone(),
+        translate.clone(),
+        quiet,
+        true,
     )?;
     validate_output("mesh", &output)?;
+    //
+    // Might be able to generalize below if pass in T and stuff like `mesh_tessellation` does.
+    //
     match N {
         HEX => {
[…]
