+++ src/main.rs
@@ -375,6 +375,7 @@ fn main() -> Result<(), ErrorWrapper> {
                     args.nely,
                     args.nelz,
                     args.remove,
+                    args.size,
                     args.xscale,
                     args.yscale,
                     args.zscale,
@@ -396,6 +397,7 @@ fn main() -> Result<(), ErrorWrapper> {
                     args.nely,
                     args.nelz,
                     args.remove,
+                    args.size,
                     args.xscale,
                     args.yscale,
                     args.zscale,
@@ -417,6 +419,7 @@ fn main() -> Result<(), ErrorWrapper> {
                     args.nely,
                     args.nelz,
                     args.remove,
+                    args.size,
                     args.xscale,
                     args.yscale,
                     args.zscale,

[…]
