+++ .github/workflows/Python.yml
@@ -103,7 +103,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - name: Download
-      uses: actions/download-artifact@v5
+      uses: actions/download-artifact@v6
     - name: Python
       uses: actions/setup-python@v6
       with:
