+++ .github/workflows/Python.yml
@@ -62,7 +62,7 @@ jobs:
       - name: Check
         run: twine check dist/*
       - name: Dist
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: dist
           path: dist
@@ -94,7 +94,7 @@ jobs:
     - name: Check
       run: twine check target/wheels/*
     - name: Upload
-      uses: actions/upload-artifact@v4
+      uses: actions/upload-artifact@v5
       with:
         name: ${{ matrix.os }}-${{ matrix.python-version }}-dist
         path: target/wheels
