+++ src/tree/mod.rs
@@ -3,12 +3,12 @@ use std::time::Instant;
 
 mod hex;
 mod tessellation;
+mod tri;
 
 use super::{
     Coordinate, Coordinates, NSD, Vector,
     fem::{
-        Blocks, FiniteElementMethods, HEX, HexahedralFiniteElements, Size,
-        TriangularFiniteElements, hex::HexConnectivity,
+        Blocks, FiniteElementMethods, HEX, HexahedralFiniteElements, Size, hex::HexConnectivity,
     },
     tessellation::Tessellation,
     voxel::{Nel, Remove, Scale, Translate, VoxelData, Voxels},
@@ -531,85 +531,6 @@ impl Cell {
                 min_z: val_z,
             },
         ]
-    }
-}
-
-impl From<Voxels> for Octree {
-    fn from(voxels: Voxels) -> Octree {
-        #[cfg(feature = "profile")]
-        let time = Instant::now();
-        let (data_voxels, remove, scale, translate) = voxels.data();
-        let mut nel_i = 0;
-        let nel_padded = data_voxels
-            .shape()
-            .iter()
-            .map(|nel_0| {
-                nel_i = *nel_0;
-                while (nel_i & (nel_i - 1)) != 0 {
-                    nel_i += 1
-                }
-                nel_i
-            })
-            .max()
-            .unwrap();
-        let nel = Nel::from([nel_padded; NSD]);
-        let mut data = VoxelData::from(nel);
-        data.iter_mut().for_each(|entry| *entry = PADDING);
-        if data_voxels.iter().any(|entry| entry == &PADDING) {
-            panic!("Segmentation cannot use 255 as an ID with octree padding.")
-        }
-        data.axis_iter_mut(Axis(2))
-            .zip(data_voxels.axis_iter(Axis(2)))
-            .for_each(|(mut data_i, data_voxels_i)| {
-                data_i
-                    .axis_iter_mut(Axis(1))
-                    .zip(data_voxels_i.axis_iter(Axis(1)))
-                    .for_each(|(mut data_ij, data_voxels_ij)| {
-                        data_ij
-                            .iter_mut()
-                            .zip(data_voxels_ij.iter())
-                            .for_each(|(data_ijk, data_voxels_ijk)| *data_ijk = *data_voxels_ijk)
-                    })
-            });
-        let nel_min = nel.iter().min().unwrap();
-        let lngth = *nel_min as u16;
-        let mut tree = Octree {
-            nel,
-            octree: vec![],
-            remove,
-            scale,
-            translate,
-        };
-        (0..(nel.x() / nel_min)).for_each(|i| {
-            (0..(nel.y() / nel_min)).for_each(|j| {
-                (0..(nel.z() / nel_min)).for_each(|k| {
-                    tree.push(Cell {
-                        block: None,
-                        cells: None,
-                        faces: [None; NUM_FACES],
-                        lngth,
-                        min_x: lngth * i as u16,
-                        min_y: lngth * j as u16,
-                        min_z: lngth * k as u16,
-                    })
-                })
-            })
-        });
-        let mut index = 0;
-        while index < tree.len() {
-            if let Some(block) = tree[index].homogeneous(&data) {
-                tree[index].block = Some(block)
-            } else {
-                tree.subdivide(index)
-            }
-            index += 1;
-        }
-        #[cfg(feature = "profile")]
-        println!(
-            "           \x1b[1;93m⤷ Octree initialization\x1b[0m {:?} ",
-            time.elapsed()
-        );
-        tree
     }
 }
 
@@ -1624,464 +1545,81 @@ impl Octree {
     }
 }
 
-impl From<Octree> for TriangularFiniteElements {
[…]
