### 2025-10-27 89fee9f
one/more of the hex templates is messed up
