+++ src/fem/mod.rs
@@ -11,7 +11,7 @@ pub use tri::{TRI, TriangularFiniteElements};
 #[cfg(feature = "profile")]
 use std::time::Instant;
 
-use super::{Coordinate, Coordinates, NSD, Octree, Remove, Tessellation, Vector};
+use super::{Coordinate, Coordinates, NSD, Octree, Tessellation, Vector};
 use chrono::Utc;
 use conspire::{
     constitutive::solid::hyperelastic::NeoHookean,
@@ -238,27 +238,40 @@ where
         ))
     }
     fn from_tessellation(tessellation: Tessellation, size: Size) -> Self {
-        let mut tree = Octree::from_tessellation(tessellation, size);
+        let tree = Octree::from_tessellation(tessellation, size);
+        match N {
+            HEX => {
+                let hexes = HexahedralFiniteElements::from(tree);
+                let (element_blocks, hex_connectivity, nodal_coordinates) = hexes.data();
+                let mut element_node_connectivity = vec![[0; N]; hex_connectivity.len()];
+                element_node_connectivity
+                    .iter_mut()
+                    .zip(hex_connectivity)
+                    .for_each(|(a, b)| a.iter_mut().zip(b).for_each(|(c, d)| *c = d));
+                Self::from_data(element_blocks, element_node_connectivity, nodal_coordinates)
+            }
+            _ => panic!(),
+        }
         //
         // Below is temporary for octree visualization.
         //
-        tree.prune();
-        let remove = match tree.remove() {
-            Remove::Some(blocks) => Some(blocks.clone()),
-            Remove::None => None,
-        };
-        let scale = tree.scale().clone();
-        let translate = tree.translate().clone();
-        let hexes = tree
-            .octree_into_finite_elements(remove, scale, translate)
-            .unwrap();
-        let (element_blocks, hex_connectivity, nodal_coordinates) = hexes.data();
-        let mut element_node_connectivity = vec![[0; N]; hex_connectivity.len()];
-        element_node_connectivity
-            .iter_mut()
-            .zip(hex_connectivity)
-            .for_each(|(a, b)| a.iter_mut().zip(b).for_each(|(c, d)| *c = d));
-        Self::from_data(element_blocks, element_node_connectivity, nodal_coordinates)
+        // tree.prune();
+        // let remove = match tree.remove() {
+        //     super::Remove::Some(blocks) => Some(blocks.clone()),
+        //     super::Remove::None => None,
+        // };
+        // let scale = tree.scale().clone();
+        // let translate = tree.translate().clone();
+        // let hexes = tree
+        //     .octree_into_finite_elements(remove, scale, translate)
+        //     .unwrap();
+        // let (element_blocks, hex_connectivity, nodal_coordinates) = hexes.data();
+        // let mut element_node_connectivity = vec![[0; N]; hex_connectivity.len()];
+        // element_node_connectivity
+        //     .iter_mut()
+        //     .zip(hex_connectivity)
+        //     .for_each(|(a, b)| a.iter_mut().zip(b).for_each(|(c, d)| *c = d));
+        // Self::from_data(element_blocks, element_node_connectivity, nodal_coordinates)
     }
     fn laplacian(&self, node_node_connectivity: &VecConnectivity) -> Coordinates {
         let nodal_coordinates = self.get_nodal_coordinates();
