### 2025-10-28 f6955ec
translate to center instead, handle buffer between model edge and octree edge cell centers using that and going up in nel if length is 2^k
