+++ src/fem/mod.rs
@@ -19,7 +19,7 @@ use conspire::{
         ElementBlock, FiniteElementBlock, FirstOrderMinimize, LinearHexahedron, LinearTetrahedron,
     },
     math::{
-        Scalar, Tensor, TensorArray, TensorVec,
+        Scalar, Tensor, TensorArray, TensorRank1List, TensorVec,
         optimize::{EqualityConstraint, GradientDescent, LineSearch},
     },
 };
@@ -40,6 +40,9 @@ use vtkio::{
 
 const ELEMENT_NUMBERING_OFFSET: usize = 1;
 const NODE_NUMBERING_OFFSET: usize = 1;
+
+/// A bounding box described by two opposing corners.
+pub type BoundingBox = TensorRank1List<3, 1, 2>;
 
 /// A vector of finite element block IDs.
 pub type Blocks = Vec<u8>;
@@ -83,6 +86,8 @@ pub trait FiniteElementMethods<const M: usize, const N: usize>
 where
     Self: FiniteElementSpecifics<M> + Sized,
 {
+    /// Calculates and returns the bounding box.
+    fn bounding_box(&self) -> BoundingBox;
     /// Calculates and returns the coordinates of the centroids.
     fn centroids(&self) -> Coordinates;
     /// Returns and moves the data associated with the finite elements.
@@ -111,6 +116,8 @@ where
     fn node_element_connectivity(&mut self) -> Result<(), &str>;
     /// Calculates and sets the node-to-node connectivity.
     fn node_node_connectivity(&mut self) -> Result<(), &str>;
+    /// Remove elements with nodes that are outside a given bounding box.
+    fn remove_outside_bounding_box(&mut self, bounding_box: BoundingBox);
     /// Smooths the nodal coordinates according to the provided smoothing method.
     fn smooth(&mut self, method: &Smoothing) -> Result<(), String>;
     /// Writes the finite elements data to a new Exodus file.
@@ -163,6 +170,25 @@ impl<const M: usize, const N: usize> FiniteElementMethods<M, N> for FiniteElemen
 where
     Self: FiniteElementSpecifics<M> + Sized,
 {
+    fn bounding_box(&self) -> BoundingBox {
+        let (minimum, maximum) = self.get_nodal_coordinates().iter().fold(
+            (
+                Coordinate::new([f64::INFINITY; NSD]),
+                Coordinate::new([f64::NEG_INFINITY; NSD]),
+            ),
+            |(mut minimum, mut maximum), coordinate| {
+                minimum
+                    .iter_mut()
+                    .zip(maximum.iter_mut().zip(coordinate.iter()))
+                    .for_each(|(min, (max, &coord))| {
+                        *min = min.min(coord);
+                        *max = max.max(coord);
+                    });
+                (minimum, maximum)
+            },
+        );
+        BoundingBox::new([minimum.into(), maximum.into()])
+    }
     fn centroids(&self) -> Coordinates {
         let coordinates = self.get_nodal_coordinates();
         let number_of_nodes = N as f64;
@@ -238,10 +264,13 @@ where
         ))
     }
     fn from_tessellation(tessellation: Tessellation, size: Size) -> Self {
-        let tree = Octree::from_tessellation(tessellation, size);
+        let triangular_finite_elements = TriangularFiniteElements::from(tessellation);
+        let bounding_box = triangular_finite_elements.bounding_box();
+        let tree = Octree::from_triangular_finite_elements(triangular_finite_elements, size);
         match N {
             HEX => {
-                let hexes = HexahedralFiniteElements::from(tree);
+                let mut hexes = HexahedralFiniteElements::from(tree);
+                hexes.remove_outside_bounding_box(bounding_box);
[…]
