### 2025-10-28 13db629
working and is now fast!
