+++ src/fem/mod.rs
@@ -50,6 +50,7 @@ pub type Blocks = Vec<u8>;
 /// An element-to-node connectivity.
 pub type Connectivity<const N: usize> = Vec<[usize; N]>;
 
+pub type Data<const N: usize> = (Blocks, Connectivity<N>, Coordinates);
 pub type Metrics = Array1<Scalar>;
 pub type Nodes = Vec<usize>;
 pub type ReorderedConnectivity = Vec<Vec<u32>>;
@@ -91,7 +92,7 @@ where
     /// Calculates and returns the coordinates of the centroids.
     fn centroids(&self) -> Coordinates;
     /// Returns and moves the data associated with the finite elements.
-    fn data(self) -> (Blocks, Connectivity<N>, Coordinates);
+    fn data(self) -> Data<N>;
     /// Returns the centroid for each exterior face.
     fn exterior_faces_centroids(&self) -> Coordinates;
     /// Constructs and returns a new finite elements type from data.
@@ -117,7 +118,7 @@ where
     /// Calculates and sets the node-to-node connectivity.
     fn node_node_connectivity(&mut self) -> Result<(), &str>;
     /// Remove elements with nodes that are outside a given bounding box.
-    fn remove_outside_bounding_box(&mut self, bounding_box: BoundingBox);
+    fn remove_nodes(self, removed_nodes: Nodes) -> Data<N>;
     /// Smooths the nodal coordinates according to the provided smoothing method.
     fn smooth(&mut self, method: &Smoothing) -> Result<(), String>;
     /// Writes the finite elements data to a new Exodus file.
@@ -203,7 +204,7 @@ where
             })
             .collect()
     }
-    fn data(self) -> (Blocks, Connectivity<N>, Coordinates) {
+    fn data(self) -> Data<N> {
         (
             self.element_blocks,
             self.element_node_connectivity,
@@ -264,14 +265,29 @@ where
         ))
     }
     fn from_tessellation(tessellation: Tessellation, size: Size) -> Self {
-        let triangular_finite_elements = TriangularFiniteElements::from(tessellation);
-        let bounding_box = triangular_finite_elements.bounding_box();
-        let tree = Octree::from_triangular_finite_elements(triangular_finite_elements, size);
+        let triangular_finite_elements = TriangularFiniteElements::from(tessellation); // profile?
+        let [[min_x, min_y, min_z], [max_x, max_y, max_z]] =
+            triangular_finite_elements.bounding_box().as_array(); // profile?
+        let tree = Octree::from_triangular_finite_elements(triangular_finite_elements, size); // profile certain parts? and combine balancing/pairing for less printout?
         match N {
             HEX => {
-                let mut hexes = HexahedralFiniteElements::from(tree);
-                hexes.remove_outside_bounding_box(bounding_box);
-                let (element_blocks, hex_connectivity, nodal_coordinates) = hexes.data();
+                let hexes = HexahedralFiniteElements::from(tree);
+                let removed_nodes = hexes
+                    .get_nodal_coordinates()
+                    .iter()
+                    .enumerate()
+                    .filter(|(_, coordinates)| {
+                        coordinates[0] < min_x
+                            || coordinates[0] > max_x
+                            || coordinates[1] < min_y
+                            || coordinates[1] > max_y
+                            || coordinates[2] < min_z
+                            || coordinates[2] > max_z
+                    })
+                    .map(|(node, _)| node)
+                    .collect();
+                let (element_blocks, hex_connectivity, nodal_coordinates) =
+                    hexes.remove_nodes(removed_nodes); // profile? with/without above?
                 let mut element_node_connectivity = vec![[0; N]; hex_connectivity.len()];
                 element_node_connectivity
                     .iter_mut()
@@ -477,24 +493,23 @@ where
[…]
