### 2025-10-28 c2e35e1
working and is now fast!
