+++ src/tree/tessellation/mod.rs
@@ -6,6 +6,9 @@ use crate::{
 };
 use conspire::math::{Scalar, Tensor, TensorArray};
 
+#[cfg(feature = "profile")]
+use std::time::Instant;
+
 impl From<Octree> for Tessellation {
     fn from(tree: Octree) -> Self {
         TriangularFiniteElements::from(tree).into()
@@ -23,6 +26,8 @@ pub fn octree_from_triangular_finite_elements(
     }
     if let Some(size) = size {
         let mut tree = octree_from_bounding_cube(&mut surface_coordinates, size);
+        #[cfg(feature = "profile")]
+        let time = Instant::now();
         let mut index = 0;
         while index < tree.len() {
             if tree[index].is_voxel() || !tree[index].any_coordinates_inside(&surface_coordinates) {
@@ -32,6 +37,11 @@ pub fn octree_from_triangular_finite_elements(
             }
             index += 1;
         }
+        #[cfg(feature = "profile")]
+        println!(
+            "             \x1b[1;93mSubdivision from size\x1b[0m {:?}",
+            time.elapsed()
+        );
         tree.balance_and_pair(true);
         tree
     } else {
@@ -40,6 +50,8 @@ pub fn octree_from_triangular_finite_elements(
 }
 
 pub fn octree_from_bounding_cube(samples: &mut Coordinates, minimum_cell_size: Scalar) -> Octree {
+    #[cfg(feature = "profile")]
+    let time = Instant::now();
     let (minimum, maximum) = samples.iter().fold(
         (
             Coordinate::new([f64::INFINITY; NSD]),
@@ -83,5 +95,10 @@ pub fn octree_from_bounding_cube(samples: &mut Coordinates, minimum_cell_size: S
         min_y: 0,
         min_z: 0,
     });
+    #[cfg(feature = "profile")]
+    println!(
+        "             \x1b[1;93mOctree initialization\x1b[0m {:?}",
+        time.elapsed()
+    );
     tree
 }
