### 2025-10-29 a210f27
faster when consuming data
