### 2025-10-30 ac7b93c
does not need to consume Octree
