### 2025-10-30 e524b6a
does not need to consume Octree
