### 2025-10-30 78349fe
named tuples and From<> instead of separate methods
