+++ src/fem/mod.rs
@@ -96,18 +96,10 @@ where
     fn data(self) -> Data<N>;
     /// Returns the centroid for each exterior face.
     fn exterior_faces_centroids(&self) -> Coordinates;
-    /// Constructs and returns a new finite elements type from data.
-    fn from_data(
-        element_blocks: Blocks,
-        element_node_connectivity: Connectivity<N>,
-        nodal_coordinates: Coordinates,
-    ) -> Self;
     /// Constructs and returns a new finite elements type from an Exodus input file.
     fn from_exo(file_path: &str) -> Result<Self, ErrorNetCDF>;
     /// Constructs and returns a new finite elements type from an Abaqus input file.
     fn from_inp(file_path: &str) -> Result<Self, ErrorIO>;
-    /// Creates a conformal mesh within a tessellation.
-    fn from_tessellation(tessellation: Tessellation, size: Size) -> Self;
     /// Calculates and returns the discrete Laplacian for the given node-to-node connectivity.
     fn laplacian(&self, node_node_connectivity: &VecConnectivity) -> Coordinates;
     /// Calculates and sets the nodal influencers.
@@ -235,79 +227,23 @@ where
             })
             .collect()
     }
-    fn from_data(
-        element_blocks: Blocks,
-        element_node_connectivity: Connectivity<N>,
-        nodal_coordinates: Coordinates,
-    ) -> Self {
-        Self {
-            boundary_nodes: vec![],
-            element_blocks,
-            element_node_connectivity,
-            exterior_nodes: vec![],
-            interface_nodes: vec![],
-            interior_nodes: vec![],
-            nodal_coordinates,
-            nodal_influencers: vec![],
-            node_element_connectivity: vec![],
-            node_node_connectivity: vec![],
-            prescribed_nodes: vec![],
-            prescribed_nodes_homogeneous: vec![],
-            prescribed_nodes_inhomogeneous: vec![],
-            prescribed_nodes_inhomogeneous_coordinates: Coordinates::zero(0),
-        }
-    }
     fn from_exo(file_path: &str) -> Result<Self, ErrorNetCDF> {
         let (element_blocks, element_node_connectivity, nodal_coordinates) =
             finite_element_data_from_exo(file_path)?;
-        Ok(Self::from_data(
+        Ok(Self::from((
             element_blocks,
             element_node_connectivity,
             nodal_coordinates,
-        ))
+        )))
     }
     fn from_inp(file_path: &str) -> Result<Self, ErrorIO> {
         let (element_blocks, element_node_connectivity, nodal_coordinates) =
             finite_element_data_from_inp(file_path)?;
-        Ok(Self::from_data(
+        Ok(Self::from((
             element_blocks,
             element_node_connectivity,
             nodal_coordinates,
-        ))
-    }
-    fn from_tessellation(tessellation: Tessellation, size: Size) -> Self {
-        let triangular_finite_elements = TriangularFiniteElements::from(tessellation);
-        let bounding_box = triangular_finite_elements.bounding_box();
-        let tree = Octree::from_triangular_finite_elements(triangular_finite_elements, size);
-        let hexes = match N {
-            HEX => HexahedralFiniteElements::from(&tree),
-            _ => panic!(),
-        };
-        let (element_blocks, connectivity, nodal_coordinates) =
-            hexes.remove_nodes_outside(bounding_box);
-        let element_node_connectivity =
-            unsafe { transmute::<Connectivity<8>, Connectivity<N>>(connectivity) };
-        Self::from_data(element_blocks, element_node_connectivity, nodal_coordinates)
-        //
-        // Below is temporary code used for octree visualization.
-        //
-        // tree.prune();
-        // let remove = match tree.remove() {
[…]
