### 2025-10-30 57d7404
named tuples and From<> instead of separate methods
