### 2025-10-30 e6b9254
named tuples and From<> instead of separate methods
