### 2025-10-30 ce5103e
named tuples and From<> instead of separate methods
