### 2025-10-30 9aa5966
round down to home base to pick the cell a node is within
