### 2025-10-30 0da22b0
can use u16 to avoid memory issues since octree already bound by that; should try to use octree cells to speed this up though
