+++ src/fem/tri/mod.rs
@@ -6,8 +6,8 @@ use std::time::Instant;
 
 use super::{
     super::tree::Edges, Connectivity, Coordinate, Coordinates, FiniteElementMethods,
-    FiniteElementSpecifics, FiniteElements, Metrics, Smoothing, Tessellation, VecConnectivity,
-    Vector,
+    FiniteElementSpecifics, FiniteElements, Metrics, Size, Smoothing, Tessellation,
+    VecConnectivity, Vector,
 };
 use conspire::{
     math::{Tensor, TensorArray, TensorVec, Vector as VectorConspire},
@@ -122,8 +122,8 @@ impl FiniteElementSpecifics<NUM_NODES_FACE> for TriangularFiniteElements {
             .map(|angle| angle.sin() / J_EQUILATERAL)
             .collect()
     }
-    fn remesh(&mut self, iterations: usize, smoothing_method: &Smoothing) {
-        remesh(self, iterations, smoothing_method)
+    fn remesh(&mut self, iterations: usize, smoothing_method: &Smoothing, size: Size) {
+        remesh(self, iterations, smoothing_method, size)
     }
     fn write_metrics(&self, file_path: &str) -> Result<(), ErrorIO> {
         let areas = self.areas();
@@ -294,9 +294,61 @@ impl TriangularFiniteElements {
             })
             .collect()
     }
+    /// Iteratively refine until all edges are smaller than a size.
+    pub fn refine(&mut self, size: Scalar) {
+        #[cfg(feature = "profile")]
+        let time = Instant::now();
+        let mut edges: Edges = self
+            .get_element_node_connectivity()
+            .iter()
+            .flat_map(|&[node_0, node_1, node_2]| {
+                [[node_0, node_1], [node_1, node_2], [node_2, node_0]].into_iter()
+            })
+            .collect();
+        edges.iter_mut().for_each(|edge| edge.sort());
+        edges.sort();
+        edges.dedup();
+        let nodal_coordinates = self.get_nodal_coordinates();
+        let mut lengths = Lengths::zero(edges.len());
+        edges
+            .iter()
+            .zip(lengths.iter_mut())
+            .for_each(|(&[node_a, node_b], length)| {
+                *length = (&nodal_coordinates[node_a] - &nodal_coordinates[node_b]).norm()
+            });
+        self.boundary_nodes = vec![];
+        self.exterior_nodes = vec![];
+        self.interface_nodes = vec![];
+        self.interior_nodes = vec![];
+        loop {
+            if lengths.iter().any(|length| length > &size) {
+                split_edges(self, &mut edges, &mut lengths, size / FOUR_THIRDS)
+                //
+                // Would be nice to bake this into remeshing with two options based on input args:
+                // (1) just go for some number of iterations like typically do
+                // (2) iterate until all edge below size
+                //     - would still do edge collapse (get all edges close to size, just no larger than size)
+                //
+            } else {
+                break;
+            }
+        }
+        #[cfg(feature = "profile")]
+        println!(
+            "             \x1b[1;93mSplitting large edges\x1b[0m {:?}",
+            time.elapsed()
+        );
+    }
 }
 
-fn remesh(fem: &mut TriangularFiniteElements, iterations: usize, smoothing_method: &Smoothing) {
+fn remesh(
+    fem: &mut TriangularFiniteElements,
+    iterations: usize,
+    smoothing_method: &Smoothing,
+    size: Size,
+) {
+    #[cfg(feature = "profile")]
+    let time = Instant::now();
     let mut edges: Edges = fem
         .get_element_node_connectivity()
         .iter()
[…]
