+++ src/fem/mod.rs
@@ -11,7 +11,10 @@ pub use tri::{TRI, TriangularFiniteElements};
 #[cfg(feature = "profile")]
 use std::time::Instant;
 
-use crate::{Coordinate, Coordinates, NSD, Tessellation, Vector, tree::{HexesAndCoords, OctreeAndSamples}};
+use crate::{
+    Coordinate, Coordinates, NSD, Tessellation, Vector,
+    tree::{HexesAndCoords, OctreeAndSamples},
+};
 use chrono::Utc;
 use conspire::{
     constitutive::solid::hyperelastic::NeoHookean,
@@ -63,6 +66,7 @@ pub enum Smoothing {
     Energetic,
     Laplacian(usize, Scalar),
     Taubin(usize, Scalar, Scalar),
+    None,
 }
 
 /// The finite elements type.
@@ -362,7 +366,7 @@ where
         self.node_element_connectivity = node_element_connectivity;
         #[cfg(feature = "profile")]
         println!(
-            "             \x1b[1;93mNode-to-element connectivity\x1b[0m {:?} ",
+            "             \x1b[1;93mNode-element connectivity\x1b[0m {:?} ",
             time.elapsed()
         );
         Ok(())
@@ -403,7 +407,7 @@ where
             self.node_node_connectivity = node_node_connectivity;
             #[cfg(feature = "profile")]
             println!(
-                "             \x1b[1;93mNode-to-node connectivity\x1b[0m {:?} ",
+                "             \x1b[1;93mNode-node connectivity\x1b[0m {:?} ",
                 time.elapsed()
             );
             Ok(())
@@ -412,6 +416,8 @@ where
         }
     }
     fn remove_nodes(self, removed_nodes: Nodes) -> Data<N> {
+        #[cfg(feature = "profile")]
+        let time = Instant::now();
         let mut is_removed_node = vec![false; self.nodal_coordinates.len()];
         removed_nodes
             .iter()
@@ -473,11 +479,16 @@ where
                 },
             )
             .collect();
+        #[cfg(feature = "profile")]
+        println!(
+            "             \x1b[1;93mRemoving marked nodes\x1b[0m {:?}",
+            time.elapsed()
+        );
         (element_blocks, element_node_connectivity, nodal_coordinates)
     }
     fn remove_nodes_outside(self, bounding_box: BoundingBox) -> Data<N> {
-        #[cfg(feature = "profile")]
-        let time = Instant::now();
+        // #[cfg(feature = "profile")]
+        // let time = Instant::now();
         let [[min_x, min_y, min_z], [max_x, max_y, max_z]] = bounding_box.as_array();
         let removed_nodes = self
             .get_nodal_coordinates()
@@ -493,13 +504,14 @@ where
             })
             .map(|(node, _)| node)
             .collect();
-        let data = self.remove_nodes(removed_nodes);
-        #[cfg(feature = "profile")]
-        println!(
-            "             \x1b[1;93mRemoved nodes outside\x1b[0m {:?}",
-            time.elapsed()
-        );
-        data
+        self.remove_nodes(removed_nodes)
+        // let data = self.remove_nodes(removed_nodes);
+        // #[cfg(feature = "profile")]
+        // println!(
+        //     "             \x1b[1;93mRemoved nodes outside\x1b[0m {:?}",
+        //     time.elapsed()
+        // );
+        // data
     }
     fn smooth(&mut self, method: &Smoothing) -> Result<(), String> {
         if !self.get_node_node_connectivity().is_empty() {
@@ -589,6 +601,7 @@ where
                         }
                     }
                 }
+                Smoothing::None => return Ok(()),
             }
[…]
