+++ src/fem/tri/mod.rs
@@ -297,7 +297,9 @@ fn split_edges(
     let mut node_e = 0;
     let mut spot_a = 0;
     let mut spot_b = 0;
-    let mut new_edge = [0; 2];
+    let mut edge_eb = [0; 2];
+    let mut edge_ec = [0; 2];
+    let mut edge_ed = [0; 2];
     let mut new_edges = vec![];
     let mut new_lengths = Lengths::zero(0);
     let element_blocks = &mut fem.element_blocks;
@@ -388,12 +390,20 @@ fn split_edges(
             node_node_connectivity[node_d - NODE_NUMBERING_OFFSET].sort();
             node_node_connectivity.push(vec![*node_a, *node_b, node_c, node_d]);
             node_node_connectivity[node_e - NODE_NUMBERING_OFFSET].sort();
-            new_edge = [node_e, *node_b];
-            new_edge.sort();
-            new_edges.push(new_edge);
+            edge_eb = [node_e, *node_b];
+            edge_eb.sort();
+            new_edges.push(edge_eb);
             *node_b = node_e;
+            edge_ec = [node_e, node_c];
+            edge_ec.sort();
+            new_edges.push(edge_ec);
+            edge_ed = [node_e, node_d];
+            edge_ed.sort();
+            new_edges.push(edge_ed);
             *length *= 0.5;
             new_lengths.push(*length);
+            new_lengths.push((&nodal_coordinates[node_e] - &nodal_coordinates[node_c]).norm());
+            new_lengths.push((&nodal_coordinates[node_e] - &nodal_coordinates[node_d]).norm());
         });
     edges.append(&mut new_edges);
     lengths.append(&mut new_lengths);
