### 2025-11-03 8563229
~25% faster to round down earlier and use when finding points in cells
