### 2025-11-03 5eafcdd
samples map is same as outside
