+++ src/tree/mod.rs
@@ -6,7 +6,7 @@ mod tessellation;
 mod tri;
 
 pub use hex::HexesAndCoords;
-pub use tessellation::OctreeAndSamples;
+pub use tessellation::Foo;
 
 use super::{
     Coordinate, Coordinates, NSD, Vector,
@@ -156,14 +156,14 @@ pub struct Cell {
 }
 
 impl Cell {
-    pub fn any_samples_inside(&self, samples_map: &[Vec<Vec<bool>>]) -> bool {
+    pub fn any_samples_inside(&self, samples: &[Vec<Vec<bool>>]) -> bool {
         let min_x = self.min_x as usize;
         let min_y = self.min_y as usize;
         let min_z = self.min_z as usize;
         let max_x = (self.min_x + self.lngth) as usize;
         let max_y = (self.min_y + self.lngth) as usize;
         let max_z = (self.min_z + self.lngth) as usize;
-        (min_x..max_x).any(|i| (min_y..max_y).any(|j| (min_z..max_z).any(|k| samples_map[i][j][k])))
+        (min_x..max_x).any(|i| (min_y..max_y).any(|j| (min_z..max_z).any(|k| samples[i][j][k])))
     }
     pub const fn get_block(&self) -> u8 {
         if let Some(block) = self.block {
