### 2025-11-04 0b875ce
Add readme field to pyproject.toml
Added 'readme' field to project metadata.
