+++ pyproject.toml
@@ -9,9 +9,9 @@ classifiers = [
   "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
   "Development Status :: 5 - Production/Stable",
   "Intended Audience :: Science/Research",
-  "Topic :: Scientific/Engineering",
   "Programming Language :: Python",
   "Programming Language :: Rust",
+  "Topic :: Scientific/Engineering",
 ]
 description = "automesh"
 dependencies = [
@@ -20,6 +20,7 @@ dependencies = [
 dynamic = ["version"]
 name = "automesh"
 requires-python = ">=3.10,<3.14"
+readme = "README.md"
 
 [project.optional-dependencies]
 dev = [
